### 2017-01-02 #151 @whedon
[REVIEW]: Kactus2: A graphical EDA tool built on the IP-XACT standard
**Submitting author:** @TermosPullo (<a href="http://orcid.org/0000-0003-2254-3332">Janne Virtanen</a>)
**Repository:** <a href="https://github.com/kactus2/kactus2dev" target ="_blank">https://github.com/kactus2/kactus2dev</a>
**Version:** 3.4.0

- [x] **Authorship:** Has the submitting author (TermosPullo) made major contributions to the software?
