### 2017-01-02 #151 @labarba
The pre-review issue mentions that we (the JOSS editors) are unclear whether this submission satisfies the submission guidelines.

In particular we require that (quoting from http://joss.theoj.org/about#author_guidelines):

> - Your software should have a research application
>   -Should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

Here is the author's reply to that question:
https://github.com/openjournals/joss-reviews/issues/117#issuecomment-260583610

@jflorence : Please help us with a comment providing your assessment of this question here.
