### 2017-01-03 a478d28
[PERFECTIVE] Updated help for component editor (refs #20).
