+++ editors/ComponentEditor/instantiations/ComponentInstantiationEditor.cpp
@@ -33,7 +33,7 @@ component_(component),
 componentInstantiation_(componentInstantiation),
 nameGroupEditor_(componentInstantiation, this, tr("Component instance name and description")),
 languageEditor_(this), 
-languageStrict_(tr("Strict"), this),
+languageStrict_(tr("strict"), this),
 libraryEditor_(this),
 packageEditor_(this),
 modulelNameEditor_(this),
