+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.36.0
+#define VERSION_FULL           3.3.39.0
 
 #define VERSION_BASEYEAR       0
-#define VERSION_DATE           "2016-12-19"
-#define VERSION_TIME           "12:12:14"
+#define VERSION_DATE           "2017-01-03"
+#define VERSION_TIME           "10:12:43"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        36
+#define VERSION_BUILDNO        39
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,36,0
-#define VERSION_PRODUCT        3,3,36,0
-#define VERSION_FILESTR        "3,3,36,0"
-#define VERSION_PRODUCTSTR     "3,3,36,0"
+#define VERSION_FILE           3,3,39,0
+#define VERSION_PRODUCT        3,3,39,0
+#define VERSION_FILESTR        "3,3,39,0"
+#define VERSION_PRODUCTSTR     "3,3,39,0"
 
 #endif
