+++ Administrative/releaseFiles/release_notes.txt
@@ -1,6 +1,6 @@
 	===============================================================================
                             RELEASE NOTES : Kactus2
-           Copyright (c) 2012-2016 Tampere University of Technology
+           Copyright (c) 2012-2017 Tampere University of Technology
                             http://funbase.cs.tut.fi
 ===============================================================================
               + Improvement or feature added
