### 2017-01-03 5831c14
[CLEANUP] Minor changes to parameters of the verilog generator.
