### 2017-01-04 aa92667
[PERFECTIVE] Restructured the memory designer diagram class into different memory item handlers (refs #80).
