+++ designEditors/MemoryDesigner/MemoryDesignConstructor.h
@@ -1,0 +1,587 @@
+//-----------------------------------------------------------------------------
+// File: MemoryDesignConstructor.h
+//-----------------------------------------------------------------------------
+// Project: Kactus2
+// Author: Mikko Teuho
+// Date: 02.01.2017
+//
+// Description:
+// Constructs the items for the memory design diagram.
+//-----------------------------------------------------------------------------
+
+#ifndef MEMORYDESIGNCONSTRUCTOR_H
+#define MEMORYDESIGNCONSTRUCTOR_H
+
+#include <common/graphicsItems/GraphicsColumnLayout.h>
+
+#include <designEditors/MemoryDesigner/ConnectivityGraphFactory.h>
+
+#include <QSharedPointer>
+
+class MemoryColumnHandler;
+class MemoryGraphicsItemHandler;
+class MemoryConnectionHandler;
+
+//-----------------------------------------------------------------------------
+//! Constructs the items for the memory design diagram.
+//-----------------------------------------------------------------------------
+class MemoryDesignConstructor
+{
+
+public:
+
+    /*!
+     *  The constructor.
+     *
+     *      @param [in] layout  Layout for the memory graphics items.
+     */    
+    MemoryDesignConstructor(QSharedPointer<GraphicsColumnLayout> layout);
+
+    /*!
+     *  The destructor.
+     */
+    ~MemoryDesignConstructor();
+
+    /*!
+     *  Change the condensing of memory graphics items.
+     *
+     *      @param [in] condenseItems   Value for condensing memory graphics items.
+     */
+    void setCondenseMemoryItems(bool condenseMemoryItems);
+
+    /*!
+     *  Check whether the memory graphics items are condensed or not.
+     *
+     *      @return True, if the memory graphics items are condensed, otherwise false.
+     */
+    bool memoryItemsAreCondensed() const;
+
+    /*!
+     *  Set the filtering for chained address space memory connections.
+     *
+     *      @param [in] filterChains    Value for the chained address space memory connection filtering.
+     */
+    void setFilterAddressSpaceChains(bool filterChains);
+
+    /*!
+     *  Check whether the chained address space memory connections are filtered or not.
+     *
+     *      @return True, if the chained address space memory connections are filtered, false otherwise.
+     */
+    bool addressSpaceChainsAreFiltered() const;
+
+    /*!
+     *  Condense all the field graphics items to their original width.
+     */
+    void condenseFieldItems();
+
+    /*!
+     *  Extend all the field graphics items to fully display their name label.
+     */
+    void extendFieldItems();
+
+    /*!
+     *  Set a new value for filtering the address space segments.
+     *
+     *      @param [in] filterSegments  Value for filtering address space segments.
+     */
+    void setFilterAddressSpaceSegments(bool filterSegments);
+
+    /*!
+     *  Check if the address space segments are filtered.
+     *
+     *      @return True, if the address space segments are filtered, false otherwise.
+     */
+    bool addressSpaceSegmentsAreFiltered() const;
+
+    /*!
+     *  Set a new value for filtering the address blocks.
+     *
+     *      @param [in] filterBlocks    Value for filtering address blocks.
+     */
+    void setFilterAddressBlocks(bool filterBlocks);
+
+    /*!
+     *  Check if the address blocks are filtered.
+     *
+     *      @return True, if the address blocks are filtered, false otherwise.
+     */
+    bool addressBlocksAreFiltered() const;
+
+    /*!
+     *  Set a new value for filtering the address block registers.
+     *
+     *      @param [in] filterSegments  Value for filtering address block registers.
+     */
+    void setFilterAddressBlockRegisters(bool filterRegisters);
+
+    /*!
[…]
