+++ designEditors/MemoryDesigner/MemoryGraphicsItemHandler.cpp
@@ -1,0 +1,157 @@
+//-----------------------------------------------------------------------------
+// File: MemoryGraphicsItemHandler.cpp
+//-----------------------------------------------------------------------------
+// Project: Kactus2
+// Author: Mikko Teuho
+// Date: 03.01.2017
+//
+// Description:
+// Constructs the memory graphics items for the memory design diagram.
+//-----------------------------------------------------------------------------
+
+#include "MemoryGraphicsItemHandler.h"
+
+#include <designEditors/MemoryDesigner/ConnectivityGraph.h>
+#include <designEditors/MemoryDesigner/ConnectivityComponent.h>
+#include <designEditors/MemoryDesigner/MemoryItem.h>
+
+#include <designEditors/MemoryDesigner/MemoryDesignerConstants.h>
+#include <designEditors/MemoryDesigner/AddressSpaceGraphicsItem.h>
+#include <designEditors/MemoryDesigner/MemoryMapGraphicsItem.h>
+#include <designEditors/MemoryDesigner/MemoryColumn.h>
+
+//-----------------------------------------------------------------------------
+// Function: MemoryGraphicsItemHandler::MemoryGraphicsItemHandler()
+//-----------------------------------------------------------------------------
+MemoryGraphicsItemHandler::MemoryGraphicsItemHandler():
+filterAddressSpaceSegments_(true),
+filterAddressBlocks_(true),
+filterRegisters_(true),
+filterFields_(true)
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: MemoryGraphicsItemHandler::~MemoryGraphicsItemHandler()
+//-----------------------------------------------------------------------------
+MemoryGraphicsItemHandler::~MemoryGraphicsItemHandler()
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: MemoryGraphicsItemHandler::setFilterAddressSpaceSegments()
+//-----------------------------------------------------------------------------
+void MemoryGraphicsItemHandler::setFilterAddressSpaceSegments(bool filterSegments)
+{
+    filterAddressSpaceSegments_ = filterSegments;
+}
+
+//-----------------------------------------------------------------------------
+// Function: MemoryGraphicsItemHandler::addressSpaceSegmentsAreFiltered()
+//-----------------------------------------------------------------------------
+bool MemoryGraphicsItemHandler::addressSpaceSegmentsAreFiltered() const
+{
+    return filterAddressSpaceSegments_;
+}
+
+//-----------------------------------------------------------------------------
+// Function: MemoryGraphicsItemHandler::setFilterAddressBlocks()
+//-----------------------------------------------------------------------------
+void MemoryGraphicsItemHandler::setFilterAddressBlocks(bool filterBlocks)
+{
+    filterAddressBlocks_ = filterBlocks;
+}
+
+//-----------------------------------------------------------------------------
+// Function: MemoryGraphicsItemHandler::addressBlocksAreFiltered()
+//-----------------------------------------------------------------------------
+bool MemoryGraphicsItemHandler::addressBlocksAreFiltered() const
+{
+    return filterAddressBlocks_;
+}
+
+//-----------------------------------------------------------------------------
+// Function: MemoryGraphicsItemHandler::setFilterAddressBlockRegisters()
+//-----------------------------------------------------------------------------
+void MemoryGraphicsItemHandler::setFilterAddressBlockRegisters(bool filterRegisters)
+{
+    filterRegisters_ = filterRegisters;
+}
+
+//-----------------------------------------------------------------------------
+// Function: MemoryGraphicsItemHandler::addressBlockRegistersAreFiltered()
[…]
