### 2017-01-04 6b699e7
[ADDITIVE] Proceeded with refactoring of Verilog generator plugin.
