+++ Plugins/PluginSystem/GeneratorPlugin/GeneratorConfigurationDialog.cpp
@@ -126,6 +126,7 @@ void GeneratorConfigurationDialog::accept()
 //-----------------------------------------------------------------------------
 void GeneratorConfigurationDialog::onSelectedFileChanged()
 {
+
     qDebug() << "x vv" << endl;
 }
 
