### 2017-01-04 04740a5
[CORRECTIVE] Made changes to code so that it compiles and passes the tests again.
