### 2017-01-04 91bab6b
[CORRECTIVE] Now HDL parser leaves the top level module name as the same as in the associated component instantiation.
