### 2017-01-04 9d1b286
[CORRECTIVE] Now the width of generated Verilog port assignment is omitted, if the parsed port assignment has width of 1. Previously this was determined based on the bounds of the associated wire.
