### 2017-01-05 408bca2
[CORRECTIVE] Fixed register ranges in memory designer (refs #80). [CORRECTIVE] Fixed field construction in memory designer (refs #80).
