### 2017-01-11 #151 @TermosPullo
Hello

It appears that the management decided to close our original repository at SourceForge after we migrated to GitHub: https://github.com/kactus2/kactus2dev.

On the other hand, it appears that you are still not quite sure if the submission is eligible for the journal. Can we help you with this somehow?
