### 2017-01-11 #151 @labarba
@TermosPullo : I was going to tell @jflorence that he can open issues on the code repo to report any problems with installation, etc., when I had a quick look at the repo and see that it has issues disabled. Is there a particular reason to disallow issue reporting?
