### 2017-01-12 #151 @TermosPullo
Our project management, including issue tracker is concentrated at our redmine service: https://kactus2.cs.tut.fi/

Please notice that it requires registeration and logging in to file issues.
