### 2017-01-12 6d393a9
[PERFECTIVE] Improved suggested items for file set groups and file types. [PERFECTIVE] Cleaner layout for component editor item editors. [CLEANUP] Added missing comments and cleaned functions.
