+++ editors/ComponentEditor/fileSet/groupmanagerdelegate.h
@@ -10,42 +10,46 @@
 
 #include <common/delegates/ComboDelegate/combodelegate.h>
 
-/*! \brief Provides editor to select a group for a file set.
+class BasicLineContentMatcher;
+
+/*! Provides editor to select a group for a file set.
  *
  */
-class GroupManagerDelegate : public ComboDelegate {
+class GroupManagerDelegate : public QStyledItemDelegate {
 	Q_OBJECT
 
 public:
 
-	/*! \brief The constructor
+	/*! The constructor
 	 *
-	 * \param parent Pointer to the owner of the delegate.
-	 *
+	 *      @param [in] parent The owner of the delegate.
 	*/
 	GroupManagerDelegate(QObject *parent);
 	
-	//! \brief The destructor
+	//! The destructor
 	virtual ~GroupManagerDelegate();
 
-	/*! \brief Create a new editor for the given item
+	/*! Create a new editor for the given item
 	 *
-	 * \param parent Owner for the editor.
-	 * \param option Contains options for the editor.
-	 * \param index Model index identifying the item.
+	 *      @param [in] parent  Owner for the editor.
+	 *      @param [in] option  Contains options for the editor.
+	 *      @param [in] index   Model index identifying the item.
 	 *
-	 * \return Pointer to the editor to be used to edit the item.
+	 *      @return The editor to be used to edit the item.
 	*/
-	virtual QWidget* createEditor(QWidget* parent, 
-		const QStyleOptionViewItem& option, 
-		const QModelIndex& index) const;
+	virtual QWidget* createEditor(QWidget* parent, QStyleOptionViewItem const& option, 
+        QModelIndex const& index) const;
+
+    virtual void setEditorData(QWidget* editor, QModelIndex const& index) const;
+
+    virtual void setModelData(QWidget* editor, QAbstractItemModel* model, QModelIndex const& index) const;
 
 private:
-	//! \brief No copying
+	//! No copying
 	GroupManagerDelegate(const GroupManagerDelegate& other);
-
-	//! \brief No assignment
 	GroupManagerDelegate& operator=(const GroupManagerDelegate& other);
+
+    BasicLineContentMatcher* matcher_;
 };
 
 #endif // GROUPMANAGERDELEGATE_H
