### 2017-01-12 6a70230
[ADDITIVE] Added a preliminary version of redesigned HDL parsing.
It should be noticed that this version lacks support for ad-hocs and flat components and is not unit tested.
