### 2017-01-13 19b4393
[ADDITIVE] Refactored HDL parser supports now multilevel hierarchies
