+++ Plugins/common/HDLParser/MetaDesign.cpp
@@ -13,8 +13,6 @@
 
 #include <library/LibraryManager/libraryinterface.h>
 
-#include <editors/ComponentEditor/common/ComponentParameterFinder.h>
-#include <editors/ComponentEditor/common/MultipleParameterFinder.h>
 #include <editors/ComponentEditor/common/ListParameterFinder.h>
 
 #include <IPXACTmodels/Design/Design.h>
@@ -29,6 +27,8 @@
 #include <IPXACTmodels/Component/BusInterface.h>
 #include <IPXACTmodels/Component/PortMap.h>
 #include <IPXACTmodels/Component/Model.h>
+
+#include <QQueue>
 
 //-----------------------------------------------------------------------------
 // Function: MetaDesign::MetaDesign()
@@ -55,28 +55,90 @@ MetaDesign::~MetaDesign()
 }
 
 //-----------------------------------------------------------------------------
-// Function: MetaDesign::parseDesign()
+// Function: MetaDesign::parseHierarchy()
 //-----------------------------------------------------------------------------
-void MetaDesign::parseDesign(QSharedPointer<GenerationComponent> topComponent,
-    QSharedPointer<View> topComponentView, QSharedPointer<GenerationInstance> topInstance)
+void MetaDesign::parseHierarchy(LibraryInterface* library, QSharedPointer<Component> topComponent,
+    QSharedPointer<Design> design, QSharedPointer<DesignConfiguration> designConf,
+    QSharedPointer<View> topComponentView)
 {
-    parseInstances();
-    parseInsterconnections();
+    // Creating null parameters for function call.
+    QSharedPointer<ListParameterFinder> topFinder;
+    QSharedPointer<QList<QSharedPointer<ConfigurableElementValue> > > cevs;
+
+    // Instantiate the top component with the selected design.
+    // Obviously it cannot have CEVs or parameters of any other component.
+    QSharedPointer<MetaInstance> topMostInstance(new MetaInstance
+        (library, topComponent, topComponentView, topFinder, cevs));
+
+    // Create the design associated with the top component.
+    QSharedPointer<MetaDesign> topMostDesign(new MetaDesign(library, design, designConf, topMostInstance));
+
+    // Add it to the queue of designs that are to be parsed.
+    QQueue<QSharedPointer<MetaDesign> > designs;
+    designs.enqueue(topMostDesign);
+
+    // How many sub designs are encountered while parsing.
+    int countOfSubDesigns = 0;
+    // How many sub design are allowed in a hierarchy in maximum.
+    const int MAXIMUM_SUBDESIGNS = 1000;
+
+    // Each module name, except the topmost instance, is associated with the count of the same name.
+    QMap<QString,int> names;
+
+    // Parse designs until no more are encountered.
+    while (!designs.isEmpty())
+    {
+         // In each iteration the next from the queue and parse it.
+         QSharedPointer<MetaDesign> currentDesign = designs.dequeue();
+         currentDesign->parseDesign();
+
+         // Now take the encountered sub designs into consideration.
+         countOfSubDesigns += currentDesign->subDesigns_.count();
+
+         // Must not have too many.
+         if (countOfSubDesigns >= MAXIMUM_SUBDESIGNS)
+         {
+             //emit reportError(tr("Hit the limit: %1 DESIGNS IN ONE HIERARCHY!!!").arg(MAXIMUM_SUBDESIGNS));
+             return;
+         }
+
+         foreach (QSharedPointer<MetaDesign> subDesign, currentDesign->subDesigns_)
+         {
+             // Put the sub design to the queue of designs that are to be parsed.
+             designs.enqueue(subDesign);
+
+             // Take the module name associated with the instantiated top component.
[…]
