### 2017-01-13 a6eed89
[CORRECTIVE] Revised some of the HDL parsing test cases to work with the refactored structure. Moreover, dropped some conditions that were needless clutter and verified in other cases.
