+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -33,6 +33,7 @@
 
 #include <Plugins/common/HDLParser/HDLComponentParser.h>
 #include <Plugins/common/HDLParser/HDLDesignParser.h>
+#include <Plugins/common/HDLParser/MetaDesign.h>
 
 class tst_HDLParser : public QObject
 {
@@ -53,7 +54,6 @@ private slots:
 
     void testHierarchicalConnections();
     void testHierarchicalConnectionsWithExpressions();
-    void testSlicedHierarchicalConnection();
 
     void testMasterToSlaveInterconnection();
     void testEmptyBounds();
@@ -62,12 +62,11 @@ private slots:
     void testInterconnectionToVaryingSizeLogicalMaps();
     void testSlicedInterconnection();
     void testAbsDefDefault();
-    void testAbsDefWidth();
     void testMasterInterconnectionToMirroredMaster();
     void testMirroredSlaveInterconnectionToSlaves();  
-
     void testPortMapsWithoutBoundsInInterconnection();
-    void testAdhocConnectionBetweenComponentInstances();    
+
+    /*void testAdhocConnectionBetweenComponentInstances();    
     void testAdhocConnectionToVaryingSizePorts();    
     void testAdhocConnectionWithPartSelect();
     void testAdhocTieOffInComponentInstance();
@@ -76,7 +75,7 @@ private slots:
     void testAdHocConnectionBetweenComponentInstancesWithExpressions();
     void testHierarchicalAdhocConnection();
     void testHierarchicalAdHocTieOffValues();
-    void testAdHocConnectionBetweenMultipleComponentInstances();
+    void testAdHocConnectionBetweenMultipleComponentInstances();*/
 
     void testInstanceParametersAreCulled();
 	void testTopComponentParametersAreUtilized();
@@ -87,9 +86,9 @@ private slots:
     void testParameterPropagationFromTopWire();
     void testMultiLevelHierachy();
 
-	void testParameterSorting();
+	/*void testParameterSorting();
     void testParameterSorting2();
-    void testParameterSorting3();
+    void testParameterSorting3();*/
 
 private:
 
@@ -131,14 +130,20 @@ private:
 
     void addHierAdhocConnection(QString const& topPort, QString const& targetInstance, QString const& targetPort);
    
-    //! The top level component for which the generator is run.
+    //! The top level component used in the tests.
     QSharedPointer<Component> topComponent_;
     
-    //! The design for which the generator is run.
+    //! The design used in the tests.
     QSharedPointer<Design> design_;
 
-	//! The design configuration for which the generator is run.
-	QSharedPointer<DesignConfiguration> designConf_;
+	//! The design configuration used in the tests.
+    QSharedPointer<DesignConfiguration> designConf_;
+
+    //! The abstraction definitions used in the tests.
+    QSharedPointer<ConfigurableVLNVReference> clkAbstractionVLNV_;
+    QSharedPointer<AbstractionDefinition> clkAbstractionDefinition_;
+    QSharedPointer<ConfigurableVLNVReference> dataAbstractionVLNV_;
+    QSharedPointer<AbstractionDefinition> dataAbstractionDefinition_;
 
     //! The view of the top component chosen as the active view.
 	QSharedPointer<View> topView_;
@@ -179,18 +184,45 @@ void tst_HDLParser::init()
     VLNV vlnv(VLNV::COMPONENT, "Test", "TestLibrary", "TestComponent", "1.0");
     topComponent_ = QSharedPointer<Component>(new Component(vlnv));
[…]
