### 2017-01-16 6bc1233
[ADDITIVE] Further adjustments to HDL parser test suite.
