+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -62,9 +62,6 @@ private slots:
     void testInterconnectionToVaryingSizeLogicalMaps();
     void testSlicedInterconnection();
     void testAbsDefDefault();
-    void testMasterInterconnectionToMirroredMaster();
-    void testMirroredSlaveInterconnectionToSlaves();  
-    void testPortMapsWithoutBoundsInInterconnection();
 
     /*void testAdhocConnectionBetweenComponentInstances();    
     void testAdhocConnectionToVaryingSizePorts();    
@@ -360,16 +357,16 @@ void tst_HDLParser::testHierarchicalConnections()
     QList<QSharedPointer<MetaDesign> > designs = MetaDesign::parseHierarchy
         (&library_, topComponent_, design_, designConf_, topView_);
 
-    QCOMPARE( designs.size(), 1 );
+    QCOMPARE(designs.size(), 1 );
     QSharedPointer<MetaDesign> design = designs.first();
 
-    QCOMPARE( design->instances_.size(), 1 );
+    QCOMPARE(design->instances_.size(), 1);
 
     QSharedPointer<MetaInstance> mInstance = design->instances_["instance1"];
-    QCOMPARE( mInstance->interfaces_.size(), 2 );
+    QCOMPARE(mInstance->interfaces_.size(), 2);
 
     QSharedPointer<MetaInterface> mInterface = mInstance->interfaces_["clk"];
-    QCOMPARE( mInterface->ports_.size(), 1 );
+    QCOMPARE(mInterface->ports_.size(), 1);
 
     QSharedPointer<MetaPort> mPort = mInterface->ports_["clk"];
     QSharedPointer<MetaPortAssignMent> mpa = mPort->assignments_.value("CLK");
@@ -1202,7 +1199,7 @@ void tst_HDLParser::testSlicedInterconnection()
     QSharedPointer<View> recvView = addReceiverComponentToLibrary(receiver, General::SLAVE);
     QSharedPointer<Component> receiverComponent = library_.getModel(receiver).dynamicCast<Component>();
     QSharedPointer<PortMap> enableMap = receiverComponent->getBusInterface("data_bus")->getPortMaps()->last();
-    enableMap->getLogicalPort()->range_ = QSharedPointer<Range>(new Range("0", "1"));
+    enableMap->getLogicalPort()->range_ = QSharedPointer<Range>(new Range("1", "0"));
 
 	addInstanceToDesign("receiver", receiver, recvView);
 
@@ -1229,48 +1226,43 @@ void tst_HDLParser::testSlicedInterconnection()
 
     addConnectionToDesign("sender", "data_bus", "receiver", "data_bus");
 
-    QSharedPointer<HDLComponentParser> componentParser =
-    QSharedPointer<HDLComponentParser>(new HDLComponentParser(&library_, topComponent_));
+    QList<QSharedPointer<MetaDesign> > designs = MetaDesign::parseHierarchy
+        (&library_, topComponent_, design_, designConf_, topView_);
 
-    componentParser->parseComponent(topView_);
-    QSharedPointer<HDLDesignParser> designParser =
-    QSharedPointer<HDLDesignParser>(new HDLDesignParser(&library_, design_, designConf_));
-	designParser->parseDesign(componentParser->getParsedComponent(), topView_);
+    QCOMPARE(designs.size(), 1);
+    QSharedPointer<MetaDesign> design = designs.first();
 
-    QList<QSharedPointer<GenerationDesign> > designs = designParser->getParsedDesigns();
+    QCOMPARE(design->instances_.size(), 2);
 
-    QCOMPARE( designs.size(), 1 );
[…]
