### 2017-01-16 6a8cb90
[ADDITIVE] Further adjusted the HDL parser test suite.
