+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -2218,30 +2218,21 @@ void tst_HDLParser::testInstanceParametersAreCulled()
 
     designConf_->getViewConfiguration("sender")->getViewConfigurableElements()->append(parameterOverride);
 
-    QSharedPointer<HDLComponentParser> componentParser =
+    QList<QSharedPointer<MetaDesign> > designs = MetaDesign::parseHierarchy
+        (&library_, topComponent_, design_, designConf_, topView_);
 
-    QSharedPointer<HDLComponentParser>(new HDLComponentParser(&library_, topComponent_));
+    QCOMPARE(designs.size(), 1);
+    QSharedPointer<MetaDesign> design = designs.first();
 
-    componentParser->parseComponent(topView_);
-    QSharedPointer<HDLDesignParser> designParser =
-    QSharedPointer<HDLDesignParser>(new HDLDesignParser(&library_, design_, designConf_));
-	designParser->parseDesign(componentParser->getParsedComponent(), topView_);
+    QCOMPARE(design->instances_.size(), 1);
 
-    QList<QSharedPointer<GenerationDesign> > designs = designParser->getParsedDesigns();
+    QSharedPointer<MetaInstance> mInstance = design->instances_["sender"];
+    QCOMPARE(mInstance->parameters_.size(), 2);
 
-    QCOMPARE( designs.size(), 1 );
-    QSharedPointer<GenerationDesign> design = designs.first();
-
-    QCOMPARE( design->instances_.size(), 1 );
-
-    QSharedPointer<GenerationInstance> gi0 = design->instances_["sender"];
-
-    QCOMPARE( gi0->parameters_.size(), 2 );
-
-    QCOMPARE( gi0->parameters_.at(0)->name(), QString("componentParameter") );
-    QCOMPARE( gi0->parameters_.at(0)->getValue(), QString("3") );
-    QCOMPARE( gi0->parameters_.at(1)->name(), QString("moduleParameter") );
-    QCOMPARE( gi0->parameters_.at(1)->getValue(), QString("2") );
+    QCOMPARE(mInstance->parameters_.at(0)->name(), QString("componentParameter"));
+    QCOMPARE(mInstance->parameters_.at(0)->getValue(), QString("3"));
+    QCOMPARE(mInstance->parameters_.at(1)->name(), QString("moduleParameter"));
+    QCOMPARE(mInstance->parameters_.at(1)->getValue(), QString("2"));
 }
 
 //-----------------------------------------------------------------------------
@@ -2283,28 +2274,19 @@ void tst_HDLParser::testTopComponentParametersAreUtilized()
 
     designConf_->getViewConfiguration("sender")->getViewConfigurableElements()->append(parameterOverride);
 
-    QSharedPointer<HDLComponentParser> componentParser =
+    QList<QSharedPointer<MetaDesign> > designs = MetaDesign::parseHierarchy
+        (&library_, topComponent_, design_, designConf_, topView_);
 
-    QSharedPointer<HDLComponentParser>(new HDLComponentParser(&library_, topComponent_));
+    QCOMPARE(designs.size(), 1);
+    QSharedPointer<MetaDesign> design = designs.first();
 
-    componentParser->parseComponent(topView_);
-    QSharedPointer<HDLDesignParser> designParser =
-    QSharedPointer<HDLDesignParser>(new HDLDesignParser(&library_, design_, designConf_));
-	designParser->parseDesign(componentParser->getParsedComponent(), topView_);
+    QCOMPARE(design->instances_.size(), 1);
 
-    QList<QSharedPointer<GenerationDesign> > designs = designParser->getParsedDesigns();
[…]
