### 2017-01-16 b9f5dff
[ADDITIVE] Now the refactored HLD parsing has the ad-hocs and associated unit tests.
