+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -63,16 +63,13 @@ private slots:
     void testSlicedInterconnection();
     void testAbsDefDefault();
 
-    /*void testAdhocConnectionBetweenComponentInstances();    
+    void testAdhocConnectionBetweenComponentInstances();    
     void testAdhocConnectionToVaryingSizePorts();    
     void testAdhocConnectionWithPartSelect();
     void testAdhocTieOffInComponentInstance();
     void testPortDefaultValueInComponentInstance();
-    void testMultipleAdhocConnectionsBetweenComponentInstances();
-    void testAdHocConnectionBetweenComponentInstancesWithExpressions();
     void testHierarchicalAdhocConnection();
-    void testHierarchicalAdHocTieOffValues();
-    void testAdHocConnectionBetweenMultipleComponentInstances();*/
+    void testAdHocConnectionBetweenMultipleComponentInstances();
 
     void testInstanceParametersAreCulled();
 	void testTopComponentParametersAreUtilized();
@@ -1345,9 +1342,9 @@ void tst_HDLParser::testAbsDefDefault()
 }
 
 //-----------------------------------------------------------------------------
-// Function: tst_HDLParser::testAdhocConnectionBetweenTwoComponentInstances()
+// Function: tst_HDLParser::testAdhocConnectionBetweenComponentInstances()
 //-----------------------------------------------------------------------------
-/*void tst_HDLParser::testAdhocConnectionBetweenComponentInstances()
+void tst_HDLParser::testAdhocConnectionBetweenComponentInstances()
 {
     VLNV senderVLNV(VLNV::COMPONENT, "Test", "TestLibrary", "TestSender", "1.0");
     QSharedPointer<View> senderView = addSenderComponentToLibrary(senderVLNV, General::MASTER);
@@ -1362,63 +1359,56 @@ void tst_HDLParser::testAbsDefDefault()
     addAdhocConnection("enableAdHoc", "sender", "enable_out", "receiver2", "enable_in");
     addAdhocConnection("dataAdHoc", "sender", "data_out", "receiver1", "data_in");
 
-    QSharedPointer<HDLComponentParser> componentParser =
+    QList<QSharedPointer<MetaDesign> > designs = MetaDesign::parseHierarchy
+        (&library_, topComponent_, design_, designConf_, topView_);
 
-    QSharedPointer<HDLComponentParser>(new HDLComponentParser(&library_, topComponent_));
+    QCOMPARE(designs.size(), 1);
+    QSharedPointer<MetaDesign> design = designs.first();
 
-    componentParser->parseComponent(topView_);
-    QSharedPointer<HDLDesignParser> designParser =
-    QSharedPointer<HDLDesignParser>(new HDLDesignParser(&library_, design_, designConf_));
-	designParser->parseDesign(componentParser->getParsedComponent(), topView_);
+    QCOMPARE(design->instances_.size(), 3);
 
-    QList<QSharedPointer<GenerationDesign> > designs = designParser->getParsedDesigns();
+    QSharedPointer<MetaInstance> mInstance = design->instances_["sender"];
 
-    QCOMPARE( designs.size(), 1 );
-    QSharedPointer<GenerationDesign> design = designs.first();
+    QSharedPointer<MetaPort> mPort = mInstance->ports_["data_out"];
[…]
