### 2017-01-17 f8bd9ee
[ADDITIVE] Refactored parsing of a flat component: It is now independent from parsing of a hierarchy.
