+++ Plugins/common/HDLParser/HDLParserCommon.h
@@ -30,7 +30,7 @@ class AbstractionDefinition;
 class Port;
 class Parameter;
 
-struct  MetaPort;
+struct MetaPort;
 
 struct MetaWire
 {
@@ -142,154 +142,6 @@ struct GenerationRemapState
     QList<QSharedPointer<QPair<QSharedPointer<Port>, QString> > > ports_;
     //! The name of the remap state.
     QString stateName_;
-};
-
-struct GenerationInterface
-{
-    //! The matching IP-XACT interface.
-    QSharedPointer<BusInterface> interface_;
-    //! The used abstraction type.
-    QSharedPointer<AbstractionType> absType_;
-    //! The abstraction definition of the abstraction type.
-    QSharedPointer<AbstractionDefinition> absDef_;
-    //! The used interface mode.
-    QString mode_;
-    //! The description of the interface.
-    QString description_;
-};
-
-struct GenerationPort
-{
-    //! The matching IP-XACT port.
-    QSharedPointer<Port> port_;
-    //! The bounds for the port.
-    QPair<QString,QString> vectorBounds_;
-    //! The array bounds, if the port is arranged as an array.
-    QPair<QString,QString> arrayBounds_;
-};
-
-struct GenerationComponent
-{
-    //! The matching IP-XACT component.
-    QSharedPointer<Component> component_;
-    //! The original, non-formatted parameters of the component.
-    QList<QSharedPointer<Parameter> > originalParameters_;
-    //! The formatted parameters of the component, ready for writing.
-    QList<QSharedPointer<Parameter> > formattedParameters_;
-    //! The parsed interfaces of the component, keyed with its name.
-    QMap<QString,QSharedPointer<GenerationInterface> > interfaces_;
-    //! The parsed ports of the component keyed with its physical name.
-    QMap<QString,QSharedPointer<GenerationPort> > ports_;
-
-    //! The filename for the document.
-    QString fileName_;
-    //! The module name for HDL.
-    QString moduleName_;
-
-    //! The parsed remap states.
-    QList<QSharedPointer<GenerationRemapState> > remapStates_;
-    //! The parsed remaps, including the default memory map.
-    QList<QSharedPointer<GenerationRemap> > remaps_;
-    //! The addressable unit bits to be used within the component.
-    QString aub_;
-    //! Total memory within the component counted in AUBs.
-    QString totalRange_;
-};
-
-struct GenerationWire
-{
-    //! The bounds of the wire.
-    QPair<QString,QString> bounds_;
-    //! The name of the wire, to tell it apart from other wires of the same interconnection.
-	QString name_;
-};
-
-struct GenerationPortAssignMent
-{
-    //! The port which is assigned.
-	QSharedPointer<GenerationPort> port_;
-    //! The bounds of the assignment.
-	QPair<QString,QString> bounds_;
-    //! The wire of interconnection where port is assigned. Is null if there is none.
-	QSharedPointer<GenerationWire> wire_;
-    //! The name of the physical port of the top component where the port is assigned. Is empty if there is none.
-	QString topPort_;
-    //! The assigned tie-off value. Is empty if there is none.
-    QString tieOff_;
-    //! True, if the port is assigned to ad-hoc interconnection, else false.
-    bool adhoc_;
-    //! The width used in the corresponding abstraction port wire.
-    QString abstractionWidth_;
-};
-
-struct GenerationInterconnection
-{
-    //! The interfaces assigned to the interconnection.
-    QList<QSharedPointer<GenerationInterface> > interfaces_;
-    //! The assignments of the ports that are mapped to the interconnection, keyed by logical port name.
-    QMultiMap<QString,QSharedPointer<GenerationPortAssignMent> > ports_;
[…]
