### 2017-01-18 cf7fc04
[ADDITIVE] Now the flat component parser is inherited from MetaInstance, so that same class interface may be used to conjure up the relevant values.
