+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -48,7 +48,7 @@ private slots:
     void init();
     void cleanup();
 
-    // Test cases:
+    // Hierarchy test cases:
     void testTopLevelComponent();
     void testTopLevelComponentExpressions();
 
@@ -80,7 +80,11 @@ private slots:
     void testParameterPropagationFromTopWire();
     void testMultiLevelHierachy();
 
-	void testParameterSorting();
+    // Flat test cases:
+    void testFlatComponent();
+    void testFlatComponentExpressions();
+
+    void testParameterSorting();
     void testParameterSorting2();
     void testParameterSorting3();
 
@@ -89,7 +93,7 @@ private:
     void addPort(QString const& portName, int portSize, DirectionTypes::Direction direction, 
         QSharedPointer<Component> component);
 
-    void addModuleParameter(QString const& name, QString const& value);
+    void addModuleParameter(QString const& name, QString const& value, QString const& valueID);
 
     void createHierarchicalConnection(QString const& topInterfaceRef, QString const& instanceInterfaceRef);
 
@@ -236,31 +240,36 @@ void tst_HDLParser::testTopLevelComponent()
     addPort("rst_n", 1, DirectionTypes::IN, topComponent_);
     addPort("dataOut", 8, DirectionTypes::OUT, topComponent_);
     addPort("dataIn", 8, DirectionTypes::IN, topComponent_);
-    addModuleParameter("dataWidth", "8");
-    addModuleParameter("freq", "100000");
+    addModuleParameter("dataWidth", "8", "firstParameter");
+    addModuleParameter("freq", "100000", "secondParameter");
 
-    QSharedPointer<HDLComponentParser> componentParser =
-    QSharedPointer<HDLComponentParser>(new HDLComponentParser(&library_, topComponent_));
-    componentParser->parseComponent(topView_);
+    QList<QSharedPointer<MetaDesign> > designs = MetaDesign::parseHierarchy
+        (&library_, topComponent_, design_, designConf_, topView_);
 
-    QCOMPARE(componentParser->parameters_.size(), 2);
-    QCOMPARE(componentParser->parameters_.at(0)->name(), QString("dataWidth"));
-    QCOMPARE(componentParser->parameters_.at(0)->getValue(), QString("8"));
-    QCOMPARE(componentParser->parameters_.at(1)->name(), QString("freq"));
-    QCOMPARE(componentParser->parameters_.at(1)->getValue(), QString("100000"));
+    QCOMPARE(designs.size(), 1);
+    QSharedPointer<MetaDesign> design = designs.first();
 
-    QCOMPARE(componentParser->ports_.size(), 4);
-    QCOMPARE(componentParser->ports_["clk"]->port_->getDirection(), DirectionTypes::IN);
+    QVERIFY(design->topInstance_);
+    QSharedPointer<MetaInstance> mInstance = design->topInstance_;
 
-    QCOMPARE(componentParser->ports_["dataIn"]->port_->getDirection(), DirectionTypes::IN);
-    QCOMPARE(componentParser->ports_["dataIn"]->vectorBounds_.first, QString("7"));
-    QCOMPARE(componentParser->ports_["dataIn"]->vectorBounds_.second, QString("0"));
+    QCOMPARE(mInstance->parameters_.size(), 2);
+    QCOMPARE(mInstance->parameters_.at(0)->name(), QString("dataWidth"));
+    QCOMPARE(mInstance->parameters_.at(0)->getValue(), QString("8"));
+    QCOMPARE(mInstance->parameters_.at(1)->name(), QString("freq"));
+    QCOMPARE(mInstance->parameters_.at(1)->getValue(), QString("100000"));
 
[…]
