### 2017-01-18 3dc61a7
[ADAPTIVE] Preliminary version of Verilog generator, that is adapted to the HDL parser refactoring.
