+++ Plugins/VerilogGenerator/ComponentInstanceVerilogWriter/ComponentInstanceVerilogWriter.h
@@ -25,7 +25,7 @@
 
 #include <QTextStream>
 #include <QSharedPointer>
-#include <Plugins/common/HDLParser/HDLParserCommon.h>
+#include <Plugins/common/HDLParser/MetaDesign.h>
 
 //-----------------------------------------------------------------------------
 //! Class for writing a component instance as a Verilog instance.
@@ -41,7 +41,7 @@ public:
      *      @param [in] sorter                  Sorter for the ports in the component.
      *      @param [in] useInterfaces           True, if interfaces are used where applicable. False means always ports.
 	 */
-	ComponentInstanceVerilogWriter(QSharedPointer<GenerationInstance> instance,
+	ComponentInstanceVerilogWriter(QSharedPointer<MetaInstance> instance,
         QSharedPointer<const PortSorter> sorter, bool useInterfaces);
 
 	//! The destructor.
@@ -139,7 +139,7 @@ private:
     *
     *      @return The Verilog description of the instance port connection assignment.
     */
-    QString assignmentForPort(QSharedPointer<GenerationPortAssignMent> gab) const;
+    QString assignmentForPort(QSharedPointer<MetaPortAssignMent> mpa) const;
 
     /*!
     *  Checks if a given port is not connected.
@@ -181,7 +181,7 @@ private:
     // Data.
     //-----------------------------------------------------------------------------
 
-	QSharedPointer<GenerationInstance> instance_;
+	QSharedPointer<MetaInstance> instance_;
 
 	//! Sorter for the ports of the component.
     QSharedPointer<const PortSorter> sorter_;
