### 2017-01-19 2d42be9
[CLEANUP] Moved Verilog implementation selection to a separate test suite.
