+++ tests/Plugins/VerilogGenerator/VerilogGenerator/tst_VerilogGenerator.cpp
@@ -3,7 +3,7 @@
 //-----------------------------------------------------------------------------
 // Project: Kactus2
 // Author: Esko Pekkarinen
-// Date: 28.7.2014
+// Date: 28.07.2014
 //
 // Description:
 // Unit test for class VerilogGenerator.
@@ -57,18 +57,6 @@ private slots:
 
     void testTopLevelParametersAreWritten();
     void testInstanceParametersAreWritten();*/
-
-	void testImplementationSelection();
-    void testImplementationSelectionWithEvilComments();
-	void testImplementationSelectionWithTag();
-	void testImplementationSelectionWithoutParameters();
-	void testImplementationSelectionWithoutPorts();
-	void testImplementationSelectionWithInstantiation();
-	void testImplementationSelectionWithPostModule();
-	void testImplementationSelectionWithTooManyModules();
-	void testImplementationSelectionWithNoModuleHeaderStart();
-	void testImplementationSelectionWithNoModuleHeaderEnd();
-	void testImplementationSelectionWithNoModuleEnd();
 
 	void testGenerationWithImplementation();
 	void testGenerationWithImplementationWithTag();
@@ -911,484 +899,6 @@ void tst_VerilogGenerator::testInstanceParametersAreWritten()
         "        .testParameter       (2))\n"
         "sender();");
 }*/
-
-//-----------------------------------------------------------------------------
-// Function: tst_VerilogGenerator::testImplementationSelection()
-//-----------------------------------------------------------------------------
-void tst_VerilogGenerator::testImplementationSelection()
-{
-	QFile existingFile("./generatorOutput.v");
-
-	existingFile.open(QIODevice::WriteOnly);
-	QTextStream outputStream(&existingFile);
-
-	QString content = QString(
-		"module TestComponent #(\n"
-		"    parameter                              dataWidth        = 8,\n"
-		"    parameter                              freq             = 100000\n"
-		") (\n"
-		"    // These ports are not in any interface\n"         
-		"    input                               clk,\n"
-		"    input          [7:0]                dataIn,\n"
-		"    input                               rst_n,\n"
-		"    output         [7:0]                dataOut\n"
-		");\n"
-		"foo\n"
-		"bar\n"
-		"endmodule\n"
-		); 
-
-	outputStream << content;
-
-	existingFile.close();
-
-	VerilogGenerator generator(&library_,false,false,"","");
-
-	QString implementation;
-	QString postModule;
-
-	generator.selectImplementation("./generatorOutput.v", implementation, postModule);
-
-	implementation = implementation.trimmed();
-
-	QCOMPARE(implementation,QString("foo\nbar"));
-}
-
-//-----------------------------------------------------------------------------
-// Function: tst_VerilogGenerator::testImplementationSelectionWithEvilComments()
-//-----------------------------------------------------------------------------
-void tst_VerilogGenerator::testImplementationSelectionWithEvilComments()
-{
-    QFile existingFile("./generatorOutput.v");
-
-    existingFile.open(QIODevice::WriteOnly);
-    QTextStream outputStream(&existingFile);
-
-    QString content = QString(
-        "module TestComponent #(\n"
-        "    parameter                              dataWidth        = 8, // module ending );\n"
-        "    parameter                              freq             = 100000 /* ending module header );\n"
-        ") (\n"
-        "    // These ports are not in any interface\n"         
-        "    input                               clk,\n"
-        "    input          [7:0]                dataIn,\n"
-        "    input                               rst_n,\n"
-        "    output         [7:0]                dataOut\n"
-        ");\n"
[…]
