+++ tests/Plugins/VerilogGenerator/ImplementationSelection/tst_ImplementationSelection.cpp
@@ -1,0 +1,574 @@
+//-----------------------------------------------------------------------------
+// File: tst_ImplementationSelection.cpp
+//-----------------------------------------------------------------------------
+// Project: Kactus2
+// Author: Janne Virtanen
+// Date: 19.01.2017
+//
+// Description:
+// Unit test for Verilog implementation selection.
+//-----------------------------------------------------------------------------
+
+#include <QtTest>
+
+#include <Plugins/VerilogGenerator/VerilogGenerator/VerilogGenerator.h>
+
+#include <tests/MockObjects/LibraryMock.h>
+
+#include <Plugins/VerilogImport/VerilogSyntax.h>
+
+class tst_ImplementationSelection : public QObject
+{
+    Q_OBJECT
+
+public:
+    tst_ImplementationSelection();
+
+private slots:
+    void initTestCase();
+    void cleanupTestCase();
+    void init();
+    void cleanup();
+
+    // Test cases:
+
+	void testImplementationSelection();
+    void testImplementationSelectionWithEvilComments();
+	void testImplementationSelectionWithTag();
+	void testImplementationSelectionWithoutParameters();
+	void testImplementationSelectionWithoutPorts();
+	void testImplementationSelectionWithInstantiation();
+	void testImplementationSelectionWithPostModule();
+	void testImplementationSelectionWithTooManyModules();
+	void testImplementationSelectionWithNoModuleHeaderStart();
+	void testImplementationSelectionWithNoModuleHeaderEnd();
+	void testImplementationSelectionWithNoModuleEnd();
+
+private:
+
+
+    //! The test mock for library interface.
+    LibraryMock library_;
+};
+
+//-----------------------------------------------------------------------------
+// Function: tst_ImplementationSelection::tst_ImplementationSelection()
+//-----------------------------------------------------------------------------
+tst_ImplementationSelection::tst_ImplementationSelection(): library_(this)
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: tst_ImplementationSelection::initTestCase()
+//-----------------------------------------------------------------------------
+void tst_ImplementationSelection::initTestCase()
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: tst_ImplementationSelection::cleanupTestCase()
+//-----------------------------------------------------------------------------
+void tst_ImplementationSelection::cleanupTestCase()
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: tst_ImplementationSelection::init()
+//-----------------------------------------------------------------------------
+void tst_ImplementationSelection::init()
+{
+    library_.clear();
+}
+
+//-----------------------------------------------------------------------------
+// Function: tst_ImplementationSelection::cleanup()
+//-----------------------------------------------------------------------------
+void tst_ImplementationSelection::cleanup()
+{
+    QFile::remove("./generatorOutput.v");
+}
+
+//-----------------------------------------------------------------------------
+// Function: tst_ImplementationSelection::testImplementationSelection()
+//-----------------------------------------------------------------------------
+void tst_ImplementationSelection::testImplementationSelection()
+{
+	QFile existingFile("./generatorOutput.v");
+
+	existingFile.open(QIODevice::WriteOnly);
+	QTextStream outputStream(&existingFile);
+
+	QString content = QString(
+		"module TestComponent #(\n"
+		"    parameter                              dataWidth        = 8,\n"
+		"    parameter                              freq             = 100000\n"
+		") (\n"
[…]
