### 2017-01-19 f9e1e9b
[PERFECTIVE] Refined fileset and file editors. [CLEANUP] Removed obsolete files.
