### 2017-01-19 bcd02eb
[ADAPTIVE] Now most of the Verilog generation test suite is adapted to the refactoring, except hierarchical assignment writer.
