+++ tests/Plugins/VerilogGenerator/VerilogGenerator/tst_VerilogGenerator.cpp
@@ -43,20 +43,21 @@ private slots:
     void testFileHeaderIsPrinted();
 
     void testHierarchicalConnections();
+    void testInstanceSlicedHierarchicalConnections();
 
     void testMasterToSlaveInterconnection();
-    /*void testMasterToMultipleSlavesInterconnections();
+    void testMasterToMultipleSlavesInterconnections();
 
     void testAdhocConnectionBetweenComponentInstances();    
     void testAdhocTieOffInComponentInstance();
     void testHierarchicalAdhocConnection();
-    void testHierarchicalAdHocTieOffValues();
+    //void testHierarchicalAdHocTieOffValues();
     
     void testDescriptionAndVLNVIsPrintedAboveInstance();
     void testDescriptionAndVLNVIsPrintedAboveInstance_data();
 
     void testTopLevelParametersAreWritten();
-    void testInstanceParametersAreWritten();*/
+    void testInstanceParametersAreWritten();
 
 	void testGenerationWithImplementation();
 	void testGenerationWithImplementationWithTag();
@@ -67,7 +68,7 @@ private:
     QSharedPointer<MetaInterconnection> addInterconnectToDesign(QSharedPointer<MetaInterface> first,
         QSharedPointer<MetaInterface> second);
 
-    QSharedPointer<MetaWire> addWireToDesign(QSharedPointer<MetaInterconnection> gic, QString name, QString leftBound, QString rightBound);
+    QSharedPointer<MetaWire> addWireToDesign(QString name, QString leftBound, QString rightBound, QSharedPointer<MetaInterconnection> mInterconnect = QSharedPointer<MetaInterconnection>::QSharedPointer());
 
     QSharedPointer<MetaPort> addPort(QString const& portName, int portSize, DirectionTypes::Direction direction, 
         QSharedPointer<MetaInstance> component, QSharedPointer<MetaInterface> mInterface =  QSharedPointer<MetaInterface>::QSharedPointer());
@@ -76,7 +77,7 @@ private:
 
     void runGenerator(bool useDesign);
 
-    void createPortAssignments(QSharedPointer<MetaPort> instancePort, QSharedPointer<MetaPort> hierPort, QSharedPointer<MetaWire> wire, QString const& leftBound, QString const& rightBound);
+    void createPortAssignment(QSharedPointer<MetaPort> mPort, QSharedPointer<MetaWire> wire, QString const& leftBound, QString const& rightBound);
 
     void createWirePortAssignment(QSharedPointer<MetaPort> port, QSharedPointer<MetaWire> wire, QSharedPointer<MetaInstance> instance, QString const& leftBound, QString const& rightBound);
 
@@ -84,20 +85,19 @@ private:
 
     QSharedPointer<MetaInstance> addInstanceToDesign(QString instanceName, QSharedPointer<Component> component);
 
+    void addInterfacesToInstance(QSharedPointer<MetaInstance> mInstance);
+
     void createInterface(QSharedPointer<MetaPort> port, QString const& interfaceName, QSharedPointer<MetaInstance> component);
     
     QSharedPointer<MetaInterface> addInterfaceToComponent(QString const& interfaceName, QSharedPointer<MetaInstance> component);
 
-    QSharedPointer<MetaInstance> addSenderComponentToLibrary(QString const& senderVLNV, General::InterfaceMode mode);
+    QSharedPointer<MetaInstance> addSender(QString const& instanceName);
 
-    QSharedPointer<MetaInstance> addReceiverComponentToLibrary(QString const& receiverVLNV, General::InterfaceMode mode);
[…]
