+++ Plugins/VerilogGenerator/ComponentInstanceVerilogWriter/ComponentInstanceVerilogWriter.cpp
@@ -134,70 +134,48 @@ QString ComponentInstanceVerilogWriter::portConnections() const
 		return "";
 	}
 
-	QString previousInterfaceName = "";
+    QString previousInterfaceName = "";
 
-    foreach(QSharedPointer<MetaPort> mPort, instance_->ports_)
+    // Pick the ports in sorted order.
+    QList<QSharedPointer<Port> > ports = sorter_->sortedPorts(instance_->component_);
+
+    foreach(QSharedPointer<Port> cPort, ports)
     {
-        // If not applicable, skip.
-        if (mPort->assignments_.size() < 1)
+        QSharedPointer<MetaPort> mPort = instance_->ports_[cPort->name()];
+
+        if (!mPort)
         {
+            // TODO: error
             continue;
         }
 
-        QSharedPointer<MetaPortAssignMent> mpa = mPort->assignments_.first();
+        QSharedPointer<QList<QSharedPointer<BusInterface> > > busInterfaces =
+            instance_->component_->getInterfacesUsedByPort(cPort->name());
         QString interfaceName;
+
+        if (busInterfaces->size() == 1)
+        {
+            interfaceName = busInterfaces->first()->name();
+        }
+        else if (!busInterfaces->isEmpty())
+        {
+            interfaceName = QLatin1String("several");
+        }
+        else
+        {
+            interfaceName = QLatin1String("none");
+        }
+
 
         QString interfaceSeparatorLine = createInterfaceSeparator(interfaceName, previousInterfaceName);
         previousInterfaceName = interfaceName;
 
         QString portAssignment = interfaceSeparatorLine + indentation().repeated(2) + ".<port>(<connection>)";
         portAssignment.replace("<port>", mPort->port_->name().leftJustified(20));
-        // TODO: use the list of port assignments rather than a single assignment,
-        portAssignment.replace("<connection>", assignmentForPort(mpa));
+        portAssignment.replace("<connection>", assignmentForPort(mPort));
 
         portAssignments.append(portAssignment);
     }
-
-	/*foreach(QString portName, sorter_->sortedPortNames(instance_->component_))
-    {
-        // Seek for a port assignment.
-        QSharedPointer<MetaPort> mPort = instance_->ports_.value(portName);
-
-        // If none found, or is not applicable, skip.
-        if (!mPort || mPort->assignments_.size() < 1)
-        {
-            continue;
-        }
-
-        QSharedPointer<MetaPortAssignMent> mpa = mPort->assignments_.first();
-        
-        QSharedPointer<QList<QSharedPointer<BusInterface> > > busInterfaces =
-        instance_->component_->getInterfacesUsedByPort(portName);
-		QString interfaceName;
-
-		if (busInterfaces->size() == 1)
-		{
-			interfaceName = busInterfaces->first()->name();
-		}
-		else if (!busInterfaces->isEmpty())
-		{
-			interfaceName = QLatin1String("several");
-		}
-		else
-		{
-			interfaceName = QLatin1String("none");
-		}
-
-		QString interfaceSeparatorLine = createInterfaceSeparator(interfaceName, previousInterfaceName);
-		previousInterfaceName = interfaceName;
-
-		QString portAssignment = interfaceSeparatorLine + indentation().repeated(2) + ".<port>(<connection>)";
-		portAssignment.replace("<port>", portName.leftJustified(20));
-        // TODO: use the list of port assignments rather than a single assignment,
-		portAssignment.replace("<connection>", assignmentForPort(mpa));
-
-		portAssignments.append(portAssignment);
-    }*/
 
     return portAssignments.join(",\n");
 }
@@ -240,51 +218,86 @@ QString ComponentInstanceVerilogWriter::createInterfaceSeparator(QString const& 
 //-----------------------------------------------------------------------------
[…]
