### 2017-01-20 7037ccf
[CORRECTIVE] Now the HDL parser picks both physical and logical bounds from the port mapping.
