### 2017-01-23 9febd35
[CORRECTIVE] Fixed some glitches in hierarchical port assignments.
