+++ Plugins/VerilogGenerator/ComponentInstanceVerilogWriter/ComponentInstanceVerilogWriter.cpp
@@ -11,7 +11,6 @@
 
 #include "ComponentInstanceVerilogWriter.h"
 
-#include <Plugins/VerilogGenerator/VerilogAssignmentWriter/VerilogAssignmentWriter.h>
 #include <Plugins/VerilogGenerator/CommentWriter/CommentWriter.h>
 #include <IPXACTmodels/Component/Port.h>
 #include <IPXACTmodels/Component/BusInterface.h>
@@ -173,7 +172,7 @@ QString ComponentInstanceVerilogWriter::portConnections() const
 
         QString portAssignment = interfaceSeparatorLine + indentation().repeated(2) + ".<port>(<connection>)";
         portAssignment.replace("<port>", mPort->port_->name().leftJustified(20));
-        portAssignment.replace("<connection>", VerilogAssignmentWriter::assignmentForPort(mPort, DirectionTypes::IN));
+        portAssignment.replace("<connection>", assignmentForInstancePort(mPort));
 
         portAssignments.append(portAssignment);
     }
@@ -230,4 +229,79 @@ bool ComponentInstanceVerilogWriter::assignAllBitsInConnection(General::PortBoun
 bool ComponentInstanceVerilogWriter::assignSingleBitInConnection(General::PortBounds const& signalBounds) const
 {
     return  signalBounds.left_ == signalBounds.right_;
+}
+
+//-----------------------------------------------------------------------------
+// Function: ComponentInstanceVerilogWriter::assignmentForPort()
+//-----------------------------------------------------------------------------
+QString ComponentInstanceVerilogWriter::assignmentForInstancePort(QSharedPointer<MetaPort> mPort)
+{
+    bool isInPort =
+        (mPort->port_->getDirection() == DirectionTypes::IN || mPort->port_->getDirection() == DirectionTypes::INOUT);
+
+    // Use the default value of port, if no assignments exist.
+    if (mPort->assignments_.size() < 1)
+    {
+        if (isInPort)
+        {
+            return mPort->defaultValue_;
+        }
+
+        return "";
+    }
+
+    QString assignmentString;
+
+    if (mPort->assignments_.size() > 1)
+    {
+        assignmentString = "{";
+    }
+
+    foreach (QSharedPointer<MetaPortAssignMent> mpa, mPort->assignments_)
+    {
+        QString subAssignment;
+
+        if (mpa->wire_)
+        {
+            subAssignment = "<wireName>[<left>:<right>]";
+
+            QPair<QString,QString> wireBounds = mpa->logicalBounds_;
+
+            // Use bounds only if they are not the same.
+            if (wireBounds.first == wireBounds.second)
+            {
+                subAssignment.remove("[<left>:<right>]");
+            }
+            else
+            {
+                subAssignment.replace("<left>", wireBounds.first);
+                subAssignment.replace("<right>", wireBounds.second);
+            }
+
+            // Write a wire as a connection.
+            subAssignment.replace("<wireName>", mpa->wire_->name_);
+        }
+        else if (isInPort)
+        {
+            // If a default value is assigned to a physical port, it shall be used.
+            subAssignment = mpa->defaultValue_;
+        }
+
+        if (!subAssignment.isEmpty())
+        {
+            assignmentString += subAssignment;
+
+            if (mpa != mPort->assignments_.last())
+            {
+                assignmentString += ", ";
+            }
+        }
+    }
+
+    if (mPort->assignments_.size() > 1)
+    {
+        assignmentString += "}";
+    }
+
+    return assignmentString;
 }
