### 2017-01-24 #151 @TermosPullo
@jflorence Your account is activated. If you cannot log in, please contact the project management with email: kactus2@cs.tut.fi
