### 2017-01-24 710229a
[CORRECTIVE] Fixed csv-import of abstract wire ports (refs #328). [PERFECTIVE] Enabled sorting of abstract wire ports.
