### 2017-01-24 ed11d4a
[CORRECTIVE] Fixed crashing when closing a design in some specific cases.
