### 2017-01-25 dd8aa3d
[ADDITIVE] Verilog generator creates a wire for each physical port, and then uses this for assigning wires. (refs #329, #327)
