### 2017-01-25 fabf66c
[CORRECTIVE] Now the generated Verilog wires and assignments come with comments. Moreover, fixed the test suites.
