+++ Plugins/VerilogGenerator/VerilogGenerator/VerilogGenerator.h
@@ -109,18 +109,6 @@ public:
      *      @remark If parsing is not called before generation, nothing is generated.
      */
     void prepareDesign(QString const& outputPath, QSharedPointer<MetaDesign> design);
-
-    /*!
-     *  Parses the module implementation out of verilog file given as output, if it already exists.
-     *
-	 *      @param [in] outputPath		The path to the output file.
-	 *      @param [in] implementation	The module implementation.
-	 *      @param [in] postModule		Anything that exists after the module declaration.
-	 *
-	 *      @return False, if the file exists, but could not be opened or parsed properly.
-     */
-	bool selectImplementation(QString const& outputPath, QString& implementation,
-		QString& postModule) const;
     
     /*!
      *  Writes the Verilog to files.
