+++ tests/Plugins/VerilogGenerator/ImplementationSelection/tst_ImplementationSelection.cpp
@@ -11,8 +11,6 @@
 
 #include <QtTest>
 
-#include <Plugins/VerilogGenerator/VerilogGenerator/VerilogGenerator.h>
-
 #include <tests/MockObjects/LibraryMock.h>
 
 #include <Plugins/VerilogImport/VerilogSyntax.h>
@@ -121,12 +119,13 @@ void tst_ImplementationSelection::testImplementationSelection()
 
 	existingFile.close();
 
-	VerilogGenerator generator(&library_,false,false,"","");
+	
 
 	QString implementation;
-	QString postModule;
+    QString postModule;
+    QString error;
 
-	generator.selectImplementation("./generatorOutput.v", implementation, postModule);
+	VerilogSyntax::selectImplementation("./generatorOutput.v", implementation, postModule, error);
 
 	implementation = implementation.trimmed();
 
@@ -163,12 +162,13 @@ void tst_ImplementationSelection::testImplementationSelectionWithEvilComments()
 
     existingFile.close();
 
-    VerilogGenerator generator(&library_,false,false,"","");
+    
 
     QString implementation;
     QString postModule;
+    QString error;
 
-    generator.selectImplementation("./generatorOutput.v", implementation, postModule);
+    VerilogSyntax::selectImplementation("./generatorOutput.v", implementation, postModule, error);
 
     implementation = implementation.trimmed();
 
@@ -206,12 +206,13 @@ void tst_ImplementationSelection::testImplementationSelectionWithTag()
 
 	existingFile.close();
 
-	VerilogGenerator generator(&library_,false,false,"","");
+	
 
 	QString implementation;
-	QString postModule;
+    QString postModule;
+    QString error;
 
-	generator.selectImplementation("./generatorOutput.v", implementation, postModule);
+	VerilogSyntax::selectImplementation("./generatorOutput.v", implementation, postModule, error);
 
 	implementation = implementation.trimmed();
 
@@ -246,12 +247,13 @@ void tst_ImplementationSelection::testImplementationSelectionWithoutParameters()
 
 	existingFile.close();
 
-	VerilogGenerator generator(&library_,false,false,"","");
+	
 
 	QString implementation;
-	QString postModule;
+    QString postModule;
+    QString error;
 
-	generator.selectImplementation("./generatorOutput.v", implementation, postModule);
+	VerilogSyntax::selectImplementation("./generatorOutput.v", implementation, postModule, error);
 
 	implementation = implementation.trimmed();
 
@@ -278,12 +280,13 @@ void tst_ImplementationSelection::testImplementationSelectionWithoutPorts()
 
 	existingFile.close();
 
-	VerilogGenerator generator(&library_,false,false,"","");
+	
 
 	QString implementation;
-	QString postModule;
+    QString postModule;
+    QString error;
 
-	generator.selectImplementation("./generatorOutput.v", implementation, postModule);
+	VerilogSyntax::selectImplementation("./generatorOutput.v", implementation, postModule, error);
 
 	implementation = implementation.trimmed();
 
@@ -331,12 +334,13 @@ void tst_ImplementationSelection::testImplementationSelectionWithInstantiation()
 
 	existingFile.close();
 
-	VerilogGenerator generator(&library_,false,false,"","");
+	
 
 	QString actualImplementation;
-	QString postModule;
+    QString postModule;
+    QString error;
 
-	generator.selectImplementation("./generatorOutput.v", actualImplementation, postModule);
+	VerilogSyntax::selectImplementation("./generatorOutput.v", actualImplementation, postModule, error);
 
 	actualImplementation = actualImplementation.trimmed();
 
[…]
