### 2017-01-26 04b89d5
[ADDITIVE] Initial implementation for parsing parameter values in hierarchy for memory designs.
