+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.107.0
+#define VERSION_FULL           3.3.112.0
 
 #define VERSION_BASEYEAR       0
-#define VERSION_DATE           "2017-01-24"
-#define VERSION_TIME           "10:11:49"
+#define VERSION_DATE           "2017-01-26"
+#define VERSION_TIME           "10:03:05"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        107
+#define VERSION_BUILDNO        112
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,107,0
-#define VERSION_PRODUCT        3,3,107,0
-#define VERSION_FILESTR        "3,3,107,0"
-#define VERSION_PRODUCTSTR     "3,3,107,0"
+#define VERSION_FILE           3,3,112,0
+#define VERSION_PRODUCT        3,3,112,0
+#define VERSION_FILESTR        "3,3,112,0"
+#define VERSION_PRODUCTSTR     "3,3,112,0"
 
 #endif
