### 2017-01-26 7f0892d
[CLEANUP] Improved readability and minor fixes.
