### 2017-01-26 d43ea6e
[CLEANUP] Verilog generation output is no longer directly directed to output files.
