+++ Plugins/VerilogGenerator/VerilogWriterFactory/VerilogWriterFactory.h
@@ -1,0 +1,149 @@
+//-----------------------------------------------------------------------------
+// File: VerilogWriterFactory.h
+//-----------------------------------------------------------------------------
+// Project: Kactus2
+// Author: Janne Virtanen
+// Date: 26.01.2017
+//
+// Description:
+// Creates writes for generating Verilog.
+//-----------------------------------------------------------------------------
+
+#ifndef VERILOGWRITERFACTORY_H
+#define VERILOGWRITERFACTORY_H
+
+#include "../veriloggeneratorplugin_global.h"
+
+#include "VerilogDocument.h"
+
+#include <QTextStream>
+#include <Plugins/common/HDLParser/MetaDesign.h>
+
+class ComponentVerilogWriter;
+class ComponentInstanceVerilogWriter;
+class VerilogAssignmentWriter;
+class LibraryInterface;
+class PortSorter;
+class VerilogHeaderWriter;
+class VerilogWireWriter;
+class CommentWriter;
+class Writer;
+class WriterGroup;
+
+//-----------------------------------------------------------------------------
+// Verilog file generator.
+//-----------------------------------------------------------------------------
+class VerilogWriterFactory : public QObject
+{
+	Q_OBJECT
+
+public:
+
+	/*!
+	 *  The constructor.
+	 *
+	 *      @param [in] library                 The component library.
+     *      @param [in] useInterfaces           True, if interfaces are utilized separately from physical ports, else false.
+     *      @param [in] generateMemory          If true, definitions for registers are generated as well, else it is false.
+	 *      @param [in] kactusVersion		    The version of the current Kactus build.
+	 *      @param [in] generatorVersion	    The current version of the generator.
+	 */
+    VerilogWriterFactory(LibraryInterface* library, bool useInterfaces, bool generateMemory,
+        QString const& kactusVersion, QString const& generatorVersion);
+
+    //! The destructor.
+    ~VerilogWriterFactory();
+
+    /*!
+     *  Prepares a given HDL component for generation, creating writers for the task.
+     *
+     *      @param [in] outputPath			The path to the output file.
+     *      @param [in] component           The component to which is prepared for generation.
+     *
+     *      @return False, if something went wrong.
+     */
+    QSharedPointer<VerilogDocument> prepareComponent(QString const& outputPath, QSharedPointer<MetaInstance> component);
+
+    /*!
+     *  Prepares the given set of HDL designs for generation, creating writers for the task.
+     *
+     *      @param [in] outputPath			The path to the output file.
+     *      @param [in] designs             The designs to which are prepared for generation.
+     *
+     *      @remark If parsing is not called before generation, nothing is generated.
+     */
+    QSharedPointer<VerilogDocument> prepareDesign(QString const& outputPath, QSharedPointer<MetaDesign> design);
+
+signals:
+	
+    /*!
+     *  Emitted when reportable error occurs.
+     */
+	void reportError(const QString& errorMessage) const;
+
+private:
+
+    // Disable copying.
+    VerilogWriterFactory(VerilogWriterFactory const& rhs);
+    VerilogWriterFactory& operator=(VerilogWriterFactory const& rhs);
+
+    /*!
+    *  Checks if the generator should write nothing.
+    *
+    *      @return True, if the generator has nothing to write, otherwise false.
+    */
+    bool nothingToWrite() const;
+
+    /*!
+     *  Initializes writers for the given top component.
+
+     *      @return The document, which has writers associated with the component writing.
+     */
+    QSharedPointer<VerilogDocument> initializeComponentWriters(QSharedPointer<MetaInstance> topComponent);
+    
+    /*!
+    *  Initializes writers for the given design.
+     *
+	 *      @param [in] design		        The design, which data will be written.
[…]
