+++ tests/Plugins/VerilogGenerator/VerilogWriterFactory/tst_VerilogWriterFactory.cpp
@@ -1,0 +1,1588 @@
+//-----------------------------------------------------------------------------
+// File: tst_VerilogWriterFactory.cpp
+//-----------------------------------------------------------------------------
+// Project: Kactus2
+// Author: Janne Virtanen
+// Date: 26.01.2017
+//
+// Description:
+// Unit test for class VerilogWriterFactory.
+//-----------------------------------------------------------------------------
+
+#include <QtTest>
+
+#include <Plugins/VerilogGenerator/VerilogWriterFactory/VerilogWriterFactory.h>
+
+#include <tests/MockObjects/LibraryMock.h>
+
+#include <Plugins/common/HDLParser/MetaDesign.h>
+#include <Plugins/VerilogImport/VerilogSyntax.h>
+
+#include <IPXACTmodels/common/Parameter.h>
+#include <IPXACTmodels/Component/Port.h>
+#include <IPXACTmodels/Component/Component.h>
+#include <IPXACTmodels/Component/BusInterface.h>
+#include <IPXACTmodels/Component/PortMap.h>
+
+class tst_VerilogWriterFactory : public QObject
+{
+    Q_OBJECT
+
+public:
+    tst_VerilogWriterFactory();
+
+private slots:
+    void initTestCase();
+    void cleanupTestCase();
+    void init();
+    void cleanup();
+
+    // Hierarchy test cases:
+    void testTopLevelComponent();
+
+    void testFileHeaderIsPrinted();
+
+    void testHierarchicalConnections();
+    void testInstanceSlicedHierarchicalConnections();
+    void testTopSlicedHierarchicalConnections();
+
+    void testMasterToSlaveInterconnection();
+    void testPhysicalSlicedMasterToSlaveInterconnection();
+    void testLogicalSlicedMasterToSlaveInterconnection();
+    void testPortSlicedMasterToSlaveInterconnection();
+    void testMasterToMultipleSlavesInterconnections();
+
+    void testAdhocConnectionBetweenComponentInstances();    
+    void testAdhocTieOffInComponentInstance();
+    void testHierarchicalAdhocConnection();
+    void testHierarchicalAdHocTieOffValues();
+    
+    void testDescriptionAndVLNVIsPrintedAboveInstance();
+    void testDescriptionAndVLNVIsPrintedAboveInstance_data();
+
+    void testTopLevelParametersAreWritten();
+    void testInstanceParametersAreWritten();
+
+    // Flat test cases:
+    void testFlatComponent();
+
+	void testGenerationWithImplementation();
+	void testGenerationWithImplementationWithTag();
+	void testGenerationWithImplementationWithPostModule();
+
+private:
+
+    QSharedPointer<MetaInterconnection> addInterconnectToDesign(QSharedPointer<MetaInterface> first,
+        QSharedPointer<MetaInterface> second,
+        QString name);
+
+    QSharedPointer<MetaWire> addWireToDesign(QString name, QString leftBound, QString rightBound, QSharedPointer<MetaInterconnection> mInterconnect = QSharedPointer<MetaInterconnection>::QSharedPointer());
+
+    QSharedPointer<MetaPort> addPort(QString const& portName, int portSize, DirectionTypes::Direction direction, 
+        QSharedPointer<MetaInstance> component, QSharedPointer<MetaInterface> mInterface =  QSharedPointer<MetaInterface>::QSharedPointer());
+
+    void addParameter(QString const& name, QString const& value);
+
+    void runGenerator(bool useDesign);
+
+    void createPortAssignment(QSharedPointer<MetaPort> mPort, QSharedPointer<MetaWire> wire,
+        QString const& logicalLeft, QString const& logicalRight, QString const& physicalLeft, QString const& physicalRight);
+
[…]
