### 2017-01-26 3fe1459
[ADDITIVE] Iterated the generator plugin structure.
