### 2017-01-27 d52c3d6
[ADDITIVE] Proceeded in refactoring Verilog generator plugin structure.
