+++ Plugins/VerilogGenerator/VerilogGeneratorPlugin.cpp
@@ -127,108 +127,54 @@ QIcon VerilogGeneratorPlugin::getIcon() const
 //-----------------------------------------------------------------------------
 // Function: VerilogGeneratorPlugin::checkGeneratorSupport()
 //-----------------------------------------------------------------------------
-bool VerilogGeneratorPlugin::checkGeneratorSupport(QSharedPointer<Document const> libComp,
-    QSharedPointer<Document const> libDesConf,
-    QSharedPointer<Document const> libDes) const
+bool VerilogGeneratorPlugin::checkGeneratorSupport(QSharedPointer<Component const> component,
+    QSharedPointer<Design const> design,
+    QSharedPointer<DesignConfiguration const> designConfiguration) const
 {
-    QSharedPointer<const Component> targetComponent = libComp.dynamicCast<const Component>();
-    QSharedPointer<const Design> design = libDes.dynamicCast<const Design>();
-    QSharedPointer<const DesignConfiguration> designConfig = libDesConf.dynamicCast<const DesignConfiguration>();
-
     // If design or design configuration exists, their implementation overrides the top component.
     if (design)
     {
         return design->getImplementation() == KactusAttribute::HW;
     }
 
-    if (designConfig)
+    if (designConfiguration)
     {
-        return designConfig->getImplementation() == KactusAttribute::HW;
+        return designConfiguration->getImplementation() == KactusAttribute::HW;
     }
     
     // Else the availability is determined based on the top component.
-    return targetComponent && targetComponent->getImplementation() == KactusAttribute::HW;    
+    return component && component->getImplementation() == KactusAttribute::HW;    
 }
 
 //-----------------------------------------------------------------------------
 // Function: VerilogGeneratorPlugin::runGenerator()
 //-----------------------------------------------------------------------------
 void VerilogGeneratorPlugin::runGenerator(IPluginUtility* utility, 
-    QSharedPointer<Document> libComp,
-    QSharedPointer<Document> libDesConf,
-    QSharedPointer<Document> libDes)
+    QSharedPointer<Component> component,
+    QSharedPointer<Design> design,
+    QSharedPointer<DesignConfiguration> designConfiguration)
 {
     // First state we are running. Tell the version.
     utility->printInfo(tr("Running %1 %2.").arg(getName(), getVersion()));
 
+    // Must have a component under any condition.
+    if (!component)
+    {
+        utility->printError(tr("Null  component given as a parameter."));
+        return;
+    }
+
     // For later use.
     utility_ = utility;
+    topComponent_ = component;
 
-    // Must have a top component under any condition.
-    if (!libComp)
-    {
-        utility->printError(tr("Null top component given as a parameter."));
-        return;
-    }
-
-    // It must be castable to the proper type.
-    topComponent_ = libComp.dynamicCast<Component>();
-
-    if (!topComponent_)
-    {
-        utility->printError(tr("Could not cast the top component to the proper type! VLNV: %1").arg(libComp->getVlnv().toString()));
-        return;
-    }
-
-    // Try to cast design and design configuration as well.
-    QSharedPointer<Design> design = libDes.dynamicCast<Design>();
-    QSharedPointer<DesignConfiguration> designConfig = libDesConf.dynamicCast<DesignConfiguration>();
-
-    // Return errors if it does not work.
-    if (libDes && !design)
-    {
-        utility->printError(tr("A design was specified, but could not cast it to the proper type! VLNV: %1").arg(libDes->getVlnv().toString()));
-        return;
-    }
-
-    if (libDesConf && !designConfig)
-    {
[…]
