### 2017-01-27 e363b69
[PERFECTIVE] Refined component wizard file set section and layout. [PERFECTIVE] Refined layout for API definition and component editor. [CLEANUP] Removed obsolete files.
