+++ wizards/ComponentWizard/ComponentWizardConclusionPage.cpp
@@ -22,6 +22,7 @@
 #include <IPXACTmodels/Component/Component.h>
 #include <IPXACTmodels/Component/FileSet.h>
 
+#include <QFormLayout>
 #include <QGroupBox>
 #include <QHBoxLayout>
 #include <QLabel>
@@ -48,14 +49,14 @@ parametersLabel_(new QLabel(this)),
 portsLabel_(new QLabel(this)),
 viewsLabel_(new QLabel(this)),
 descriptionLabel_(new QLabel(this)),
-previewBox_(lh),
+previewBox_(lh, this),
 diffView_(new ComponentDiffWidget(expressionFormatter, this))
 {
     setTitle(tr("Summary"));
     setSubTitle(tr("You have successfully completed the wizard. Verify the choices by clicking Finish."));
     setFinalPage(true);
     
-    previewBox_.setFixedWidth(300);
+    //previewBox_.setFixedWidth(300);
 
     directoryLabel_->setWordWrap(true);
     descriptionLabel_->setWordWrap(true);
@@ -187,83 +188,44 @@ ComponentWizardConclusionPage::DiffSummary ComponentWizardConclusionPage::creteS
 //-----------------------------------------------------------------------------
 void ComponentWizardConclusionPage::setupLayout()
 {
+    QFormLayout* leftLayout = new QFormLayout();
+
+    leftLayout->addRow(tr("<b>VLNV:</b>"), vlnvLabel_);
+
+    if (originalComponent_->getImplementation() == KactusAttribute::HW)
+    {
+        leftLayout->addRow(tr("<b>Product Hierarchy:</b>"), hierarchyLabel_);
+        leftLayout->addRow(tr("<b>Firmness:</b>"), firmnessLabel_);
+    }
+
+    leftLayout->addRow(tr("<b>File location:</b>"), directoryLabel_);
+    leftLayout->addRow(tr("<b>Author:</b>"), authorLabel_);
+    leftLayout->addRow(tr("<b>Description:</b>"), descriptionLabel_);
+    leftLayout->addRow(tr("<b>File sets:</b>"), filesetsLabel_);
+
+    if (originalComponent_->getImplementation() == KactusAttribute::HW)
+    {
+        leftLayout->addRow(tr("<b>Parameters:</b>"), parametersLabel_);
+        leftLayout->addRow(tr("<b>Ports:</b>"), portsLabel_);
+        leftLayout->addRow(tr("<b>Views:</b>"), viewsLabel_);
+    }
+
     QGridLayout* layout = new QGridLayout(this);
-    int rows = 0;
 
-    SummaryLabel* detailsLabel = new SummaryLabel(QObject::tr("Summary"), this);
-    layout->addWidget(detailsLabel, rows, 0, 1, 3, Qt::AlignCenter);    
-    rows++;
+    SummaryLabel* summaryLabel = new SummaryLabel(tr("Summary"), this);
+    layout->addWidget(summaryLabel, 0, 0, 1, 2, Qt::AlignCenter);    
 
-    QLabel* vendorTitleLabel = new QLabel("<b>VLNV:</b>",  this);
-    layout->addWidget(vendorTitleLabel, rows, 0, 1, 1, Qt::AlignTop);    
-    layout->addWidget(vlnvLabel_, rows, 1, 1, 1);
-    rows++;
+    layout->addLayout(leftLayout, 1, 0, 1, 1);
+    layout->addWidget(&previewBox_, 1, 1, 1, 1);
 
-    if (originalComponent_->getImplementation() == KactusAttribute::HW)
-    {
-        QLabel* hierarchyTitleLabel = new QLabel("<b>Product Hierarchy:</b>",  this);
[…]
