### 2017-01-27 11d235b
[ADDITIVE] Generator dialog now has the preview.
