+++ Plugins/AlteraBSPGenerator/alterabspgenerator.h
@@ -84,38 +84,34 @@ public:
      */
     virtual QIcon getIcon() const;
 
-    /*! \brief Checks whether the generator supports generation for the given library component.
+    /*!
+     *  Checks whether the generator may run for the given component or design. 
      *
-     * Method: 		checkGeneratorSupport
-     * Full name:	MemoryMapHeaderGenerator::checkGeneratorSupport
-     * Access:		virtual public 
+     *      @param [in] component	        The component for which to check support. If design is not null, component
+	 *                                      will refer to design or designConfiguration.
+     *      @param [in] design	            The design, if the generator is ran for a design.
+     *      @param [in] designConfiguration The design configuration for design, if it is not null.
      *
-     * \param libComp The library component for which to check support.
-     * \param libDesConf The design configuration object if the generator is run on a hierarchical component.
-     * \param libDes The design object if the generator is run on a hierarchical component.
-     *
-     * \return True, if the generator supports the given component. Otherwise false.
-    */
-	 virtual bool checkGeneratorSupport(QSharedPointer<Document const> libComp,
-		 QSharedPointer<Document const> libDesConf = QSharedPointer<Document const>(),
-		 QSharedPointer<Document const> libDes = QSharedPointer<Document const>()) const;
+     *      @return True, if the generator may run the given component. Otherwise false.
+     */
+    virtual bool checkGeneratorSupport(QSharedPointer<Component const> component,
+        QSharedPointer<Design const> design,
+        QSharedPointer<DesignConfiguration const> designConfiguration);
 
-    /*! \brief Runs the generator.
+    /*!
+     *  Runs the generation, creating new files and/or modifying the IP-XACT metadata. The function has
+	 *  also access to the parent window widget, so that it can show dialogs for the user to aid the generation.
      *
-     * Method: 		runGenerator
-     * Full name:	MemoryMapHeaderGenerator::runGenerator
-     * Access:		virtual public 
-     *
-     * \param utility The plugin utility interface.
-     * \param libComp The component for which the generator is run.
-     * \param libDesConf The design configuration object if the generator is run on a hierarchical component.
-     * \param libDes The design object if the generator is run on a hierarchical component.
-     *
-    */
-	 virtual void runGenerator(IPluginUtility* utility, 
-		 QSharedPointer<Document> libComp, 
-		 QSharedPointer<Document> libDesConf = QSharedPointer<Document>(), 
-		 QSharedPointer<Document> libDes = QSharedPointer<Document>());
+     *      @param [in] utility			    The plugin utility interface.
+     *      @param [in] component	        The component for which to check support. If design is not null, component
+     *                                      will refer to design or designConfiguration.
+     *      @param [in] design	            The design, if the generator is ran for a design.
+     *      @param [in] designConfiguration The design configuration for design, if it is not null.
+     */
+    virtual void runGenerator(IPluginUtility* utility, 
+        QSharedPointer<Component> component,
+        QSharedPointer<Design> design,
+        QSharedPointer<DesignConfiguration> designConfiguration);
 
 	 //! \brief Returns the external program requirements of the plugin.
 	 virtual QList<IPlugin::ExternalProgramRequirement> getProgramRequirements();
