### 2017-01-27 60d7dbb
[CORRECTIVE] Now the generator plugins receive correct document types directly, so that they do not need to dynamic cast them.
