+++ Plugins/MemoryMapHeaderGenerator/memorymapheadergenerator.h
@@ -114,29 +114,33 @@ public:
     virtual QIcon getIcon() const;
 
 	 /*!
-	  *  Checks whether the generator supports generation for the given library component.
-	  *
-	  *      @param [in] libComp        The library component for which to check support.
-	  *      @param [in] libDesConf     The design configuration of a hierarchical component.
-	  *      @param [in] libDes         The design of a hierarchical component.
-      *
-      *     @return True, if the generator supports the given component, false otherwise.
-	  */
-	 virtual bool checkGeneratorSupport(QSharedPointer<Document const> libComp,
-		 QSharedPointer<Document const> libDesConf = QSharedPointer<Document const>(),
-		 QSharedPointer<Document const> libDes = QSharedPointer<Document const>()) const;
+     *  Checks whether the generator may run for the given component or design. 
+     *
+     *      @param [in] component	        The component for which to check support. If design is not null, component
+	 *                                      will refer to design or designConfiguration.
+     *      @param [in] design	            The design, if the generator is ran for a design.
+     *      @param [in] designConfiguration The design configuration for design, if it is not null.
+     *
+     *      @return True, if the generator may run the given component. Otherwise false.
+     */
+    virtual bool checkGeneratorSupport(QSharedPointer<Component const> component,
+        QSharedPointer<Design const> design,
+        QSharedPointer<DesignConfiguration const> designConfiguration) const;
 
 	 /*!
-	  *  Runs the generator.
-	  *
-	  *      @param [in] utility        The plugin utility interface.
-	  *      @param [in] libComp        The component for which the generator is run.
-	  *      @param [in] libDesConf     The design configuration of a hierarchical component.
-	  *      @param [in] libDes         The design of a hierarchical component.
-	  */
-	 virtual void runGenerator(IPluginUtility* utility, QSharedPointer<Document> libComp,
-         QSharedPointer<Document> libDesConf = QSharedPointer<Document>(),
-         QSharedPointer<Document> libDes = QSharedPointer<Document>());
+     *  Runs the generation, creating new files and/or modifying the IP-XACT metadata. The function has
+	 *  also access to the parent window widget, so that it can show dialogs for the user to aid the generation.
+     *
+     *      @param [in] utility			    The plugin utility interface.
+     *      @param [in] component	        The component for which to check support. If design is not null, component
+     *                                      will refer to design or designConfiguration.
+     *      @param [in] design	            The design, if the generator is ran for a design.
+     *      @param [in] designConfiguration The design configuration for design, if it is not null.
+     */
+    virtual void runGenerator(IPluginUtility* utility, 
+        QSharedPointer<Component> component,
+        QSharedPointer<Design> design,
+        QSharedPointer<DesignConfiguration> designConfiguration);
 
 	 /*!
 	  *  Get the external program requirements of the plugin.
