+++ Plugins/AlteraBSPGenerator/alterabspgenerator.cpp
@@ -70,55 +70,51 @@ bool AlteraBSPGenerator::checkGeneratorSupport(QSharedPointer<Component const> c
     QSharedPointer<Design const> design,
     QSharedPointer<DesignConfiguration const> designConfiguration) const
 {
-	
 	// BSP package can only be run on component editor 
-	if (libDesConf || libDes)
+	if (design || designConfiguration)
     {
 		return false;
 	}
 
     // Must have a component.
-	QSharedPointer<const Component> comp = libComp.dynamicCast<const Component>();
-
-    if (comp == 0)
+    if (!component)
     {
         return false;
     }
 	
 	// HW component must be CPU and contain at least one view.
-	if (comp->getImplementation() == KactusAttribute::HW)
+	if (component->getImplementation() == KactusAttribute::HW)
     {
-		return comp->isCpu() && comp->getViews()->size() > 0;
+		return component->isCpu() && component->getViews()->size() > 0;
     }
 
 	// By default, always return false.
 	return false;
 }
 
-void AlteraBSPGenerator::runGenerator(IPluginUtility* utility,
-	QSharedPointer<Document> libComp,
-    QSharedPointer<Document> /*libDesConf*/,
-    QSharedPointer<Document> /*libDes*/) {
-
+void AlteraBSPGenerator::runGenerator(IPluginUtility* utility, 
+    QSharedPointer<Component> component,
+    QSharedPointer<Design> design,
+    QSharedPointer<DesignConfiguration> designConfiguration)
+{
 	utility_ = utility;
 	Q_ASSERT(utility_);
 
-	Q_ASSERT(libComp);
+	Q_ASSERT(component);
 
-	QSharedPointer<Component> comp = libComp.dynamicCast<Component>();
-	Q_ASSERT(comp);
 	
-	AlteraBSPGeneratorDialog dialog(utility_->getLibraryInterface(), comp, utility_->getParentWidget());
+	AlteraBSPGeneratorDialog dialog(utility_->getLibraryInterface(), component, utility_->getParentWidget());
 
 	int result = dialog.exec();
 
 	// if user did not want to add the generated files to the component metadata.
-	if (result == QDialog::Rejected) {
+	if (result == QDialog::Rejected)
+    {
 		return;
 	}
 
 	// the path to the component's xml file for creating the relative paths.
-	QString xmlPath = utility_->getLibraryInterface()->getPath(comp->getVlnv());
+	QString xmlPath = utility_->getLibraryInterface()->getPath(component->getVlnv());
 
 	// contains the file type settings for adding files
 	QSettings settings;
@@ -135,7 +131,7 @@ void AlteraBSPGenerator::runGenerator(IPluginUtility* utility,
 		QString fileSetName = QString("%1_bsp").arg(opt.swViewName_);
 
 		// find the file set and if none exists then create one
-		QSharedPointer<FileSet> fileSet = comp->getFileSet(fileSetName);
+		QSharedPointer<FileSet> fileSet = component->getFileSet(fileSetName);
 		fileSet->setGroups("generatedFiles");
 		fileSet->setDescription(tr("Contains the BSP files generated by Altera tools for SW view %1").arg(opt.swViewName_));
 
@@ -149,11 +145,11 @@ void AlteraBSPGenerator::runGenerator(IPluginUtility* utility,
 		}
 
 
-        QSharedPointer<View> swView = comp->getModel()->findView(opt.swViewName_);
+        QSharedPointer<View> swView = component->getModel()->findView(opt.swViewName_);
         Q_ASSERT(swView);
 
 
-        QSharedPointer<ComponentInstantiation> insta = comp->getModel()->findComponentInstantiation(swView->getComponentInstantiationRef());
[…]
