### 2017-01-30 b970b1c
[PERFECTIVE] Changed frozen columns to use vertical headers instead of specialized class.
