### 2017-01-30 85c02c1
[ADDITIVE] Now the Verilog generator plugin writes files again. Deleted Verilog plugin test suite, as it would require a total rework and retargeting to generation structure rather than Verilog generation.
