+++ tests/Plugins/VerilogGenerator/VerilogGeneratorPlugin/tst_VerilogGeneratorPlugin.cpp
@@ -1,367 +1,0 @@
-//-----------------------------------------------------------------------------
-// File: tst_VerilogGenerator.cpp
-//-----------------------------------------------------------------------------
-// Project: Kactus 2
-// Author: Esko Pekkarinen
-// Date: 28.7.2014
-//
-// Description:
-// Unit test for class VerilogGenerator.
-//-----------------------------------------------------------------------------
-
-#include <QtTest>
-
-#include <Plugins/VerilogGenerator/veriloggeneratorplugin_global.h>
-#include <Plugins/VerilogGenerator/VerilogGeneratorPlugin.h>
-
-#include <Plugins/PluginSystem/PluginUtilityAdapter.h>
-#include <Plugins/PluginSystem/GeneratorPlugin/GeneratorConfiguration.h>
-
-#include <IPXACTmodels/Design/Design.h>
-
-#include <IPXACTmodels/designConfiguration/DesignConfiguration.h>
-
-#include <IPXACTmodels/Component/Component.h>
-
-#include <IPXACTmodels/Component/FileSet.h>
-
-#include <tests/MockObjects/LibraryMock.h>
-#include "IPXACTmodels/Component/View.h"
-#include "Plugins/common/NameGenerationPolicy.h"
-
-class tst_VerilogGenerator : public VerilogGeneratorPlugin
-{
-    Q_OBJECT
-
-public:
-    tst_VerilogGenerator();
-
-private slots:
-    void initTestCase();
-    void cleanupTestCase();
-    void init();
-    void cleanup();
-
-    // Test cases:
-    void testFilesetIsCreatedWhenRunForComponent();
-    void testFilesetIsCreatedWhenRunForDesign();
-	void testFilesetIsCreatedWhenRunForDesign_data();
-
-    void testFlatViewsArePossibleForTopComponent();
-    void testRefenecedDesignViewIsPossible();
-    void testRefenecedDesignConfigurationViewIsPossible();
-
-protected:
-
-    virtual bool couldConfigure(QSharedPointer<QList<QSharedPointer<View> > > const possibleViews,
-        QSharedPointer<QList<QSharedPointer<ComponentInstantiation> > > possibleInstantiations,
-        QSharedPointer<QList<QSharedPointer<FileSet> > > possibleFileSets,
-        HDLComponentParser* componentParser,
-        HDLDesignParser* designParser);
-
-    virtual QSharedPointer<GeneratorConfiguration> getConfiguration();
-
-private:
-
-   QSharedPointer<Component> createTestComponent();
-
-   QSharedPointer<Design> createTestDesign() const;
-
-   QSharedPointer<DesignConfiguration> createTestDesignConfig() const;
-
-   void verifyRTLView(QSharedPointer<View> rtlView, QSharedPointer<ComponentInstantiation> cimp);
-
-   void verifyHierarchicalView(QSharedPointer<View> hierView, QString const& viewName, QSharedPointer<ComponentInstantiation> cimp);    
-
-
-    //! The test mock for library interface.
-    LibraryMock library_;
-
-    PluginUtilityAdapter utilityMock_;
-
-    VerilogGeneratorPlugin plugin_;
-
-    QWidget* activeWindow_;
-
-    QSharedPointer<GeneratorConfiguration> configuration_;
-
-    HDLDesignParser* designParser_;
-    HDLComponentParser* componentParser_;
-};
-
-//-----------------------------------------------------------------------------
-// Function: tst_VerilogGenerator::tst_VerilogGenerator()
-//-----------------------------------------------------------------------------
-tst_VerilogGenerator::tst_VerilogGenerator(): library_(this), utilityMock_(&library_, 0, "", this), plugin_(),
-	activeWindow_(0)
-{
-}
-
[…]
