### 2017-01-30 5a20321
[CORRECTIVE] Now the hierarchical parsing plus Verilog generation has two sets of connections and port assignments: Up for instances and down for top. Moreover, the optional memory generation is executed again.
