+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -388,22 +388,22 @@ void tst_HDLParser::testHierarchicalConnections()
     QCOMPARE(mInterface->ports_.size(), 1);
 
     QSharedPointer<MetaPort> mPort = mInterface->ports_["clk"];
-    QSharedPointer<MetaPortAssignment> mpa = mPort->assignments_.value("CLK");
+    QSharedPointer<MetaPortAssignment> mpa = mPort->upAssignments_.value("CLK");
     QCOMPARE(mpa->wire_->hierPorts_.first()->port_->name(), QString("top_clk"));
 
     mInterface = mInstance->interfaces_["data"];
     QCOMPARE(mInterface->ports_.size(), 3);
 
     mPort = mInterface->ports_["data_in"];
-    mpa = mPort->assignments_.value("DATA");
+    mpa = mPort->upAssignments_.value("DATA");
     QCOMPARE(mpa->wire_->hierPorts_.first()->port_->name(), QString("data_to_instance"));
 
     mPort = mInterface->ports_["enable"];
-    mpa = mPort->assignments_.value("ENABLE");
+    mpa = mPort->upAssignments_.value("ENABLE");
     QCOMPARE(mpa->wire_->hierPorts_.first()->port_->name(), QString("enable_to_instance"));
 
     mPort = mInterface->ports_["full"];
-    mpa = mPort->assignments_.value("FULL");
+    mpa = mPort->upAssignments_.value("FULL");
     QCOMPARE(mpa->wire_->hierPorts_.first()->port_->name(), QString("full_from_instance"));
 }
 
@@ -488,7 +488,7 @@ void tst_HDLParser::testHierarchicalConnectionsWithExpressions()
     QCOMPARE(mInterface->ports_.size(), 1);
 
     QSharedPointer<MetaPort> mPort = mInterface->ports_["instance_clk"];
-    QSharedPointer<MetaPortAssignment> mpa = mPort->assignments_.value("CLK");
+    QSharedPointer<MetaPortAssignment> mpa = mPort->upAssignments_.value("CLK");
     QCOMPARE(mpa->wire_->hierPorts_.first()->port_->name(), QString("top_clk"));
     QCOMPARE(mpa->logicalBounds_.first, QString("2"));
     QCOMPARE(mpa->logicalBounds_.second, QString("0"));
@@ -679,13 +679,13 @@ void tst_HDLParser::testMasterToSlaveInterconnection()
     QSharedPointer<MetaInterface> mInterface = mInstance->interfaces_["data_bus"];
 
     QSharedPointer<MetaPort> mPort = mInterface->ports_["data_out"];
-    QSharedPointer<MetaPortAssignment> mpa = mPort->assignments_.value("DATA");
+    QSharedPointer<MetaPortAssignment> mpa = mPort->upAssignments_.value("DATA");
     QCOMPARE(mpa->logicalBounds_.first, QString("7"));
     QCOMPARE(mpa->logicalBounds_.second, QString("0"));
     QCOMPARE(mInterface->ports_.size(), 2);
 
     mPort = mInterface->ports_["enable_out"];
-    mpa = mPort->assignments_.value("ENABLE");
+    mpa = mPort->upAssignments_.value("ENABLE");
     QCOMPARE(mpa->logicalBounds_.first, QString("0"));
     QCOMPARE(mpa->logicalBounds_.second, QString("0"));
 
@@ -696,12 +696,12 @@ void tst_HDLParser::testMasterToSlaveInterconnection()
     QCOMPARE(mInterface->ports_.size(), 2);
 
     mPort = mInterface->ports_["data_in"];
-    mpa = mPort->assignments_.value("DATA");
[…]
