+++ Plugins/VerilogGenerator/PortVerilogWriter/PortVerilogWriter.cpp
@@ -53,11 +53,11 @@ bool PortVerilogWriter::nothingToWrite() const
 //-----------------------------------------------------------------------------
 QString PortVerilogWriter::createDeclaration() const
 {
-    QString portDeclaration("<direction> <type> <bounds> <name>");
+    QString portDeclaration("<direction> <type> <vectorBounds> <name>");
 
     portDeclaration.replace("<direction>", direction().leftJustified(6));
     portDeclaration.replace("<type>", port_->port_->getTypeName().leftJustified(7));
-    portDeclaration.replace("<bounds>", arrayAndVectorBounds().leftJustified(20));
+    portDeclaration.replace("<vectorBounds>", vectorBounds().leftJustified(20));
     portDeclaration.replace("<name>", port_->port_->name());
 
     return portDeclaration;
@@ -87,19 +87,10 @@ QString PortVerilogWriter::direction() const
 }
 
 //-----------------------------------------------------------------------------
-// Function: PortVerilogWriter::arrayAndVectorBounds()
+// Function: PortVerilogWriter::vectorBounds()
 //-----------------------------------------------------------------------------
-QString PortVerilogWriter::arrayAndVectorBounds() const
+QString PortVerilogWriter::vectorBounds() const
 {
-    QString arrayDefinition = "[" + port_->arrayBounds_.first + ":" + port_->arrayBounds_.second + "]";
-
-    arrayDefinition.remove(" ");
-
-    if (arrayDefinition == "[:]")
-    {
-        arrayDefinition.clear();
-    }
-
     QString vectorDefinition = "[" + port_->vectorBounds_.first + ":" + port_->vectorBounds_.second + "]";
 
     vectorDefinition.remove(" ");
@@ -109,5 +100,23 @@ QString PortVerilogWriter::arrayAndVectorBounds() const
         vectorDefinition.clear();
     }
 	
-    return arrayDefinition + vectorDefinition;
+    return vectorDefinition;
 }
+
+//-----------------------------------------------------------------------------
+// Function: PortVerilogWriter::arrayBounds()
+//-----------------------------------------------------------------------------
+QString PortVerilogWriter::arrayBounds() const
+{
+    QString arrayDefinition = "[" + port_->arrayBounds_.first + ":" + port_->arrayBounds_.second + "]";
+
+    arrayDefinition.remove(" ");
+
+    if (arrayDefinition == "[0:0]" || arrayDefinition == "[:]")
+    {
+        arrayDefinition.clear();
+    }
+
+    return arrayDefinition;
+}
+
