### 2017-01-30 160818d
[CORRECTIVE] Fixed an issue where interconnection was not assigned to interfaces.
