### 2017-01-31 3907a14
[CLEANUP] Omitted unused features, writing interfaces and registers from Verilog generation. Moreover, remap state definitions are now always written to flat components.
