### 2017-02-01 dcac2b3
[ADDITIVE] Added initial Catalog class and XML writer (refs #296).
