+++ IPXACTmodels/Catalog/Catalog.cpp
@@ -1,0 +1,230 @@
+//-----------------------------------------------------------------------------
+// File: Catalog.cpp
+//-----------------------------------------------------------------------------
+// Project: Kactus 2
+// Author: Esko Pekkarinen
+// Date: 30.01.2017
+//
+// Description:
+// Implementation for ipxact:catalog element.
+//-----------------------------------------------------------------------------
+
+#include "Catalog.h"
+
+#include "IpxactFile.h"
+
+#include <QString>
+#include <QObject>
+#include <QSharedPointer>
+#include <QStringList>
+
+//-----------------------------------------------------------------------------
+// Function: Catalog::Catalog()
+//-----------------------------------------------------------------------------
+Catalog::Catalog(VLNV const & vlnv): 
+Document(vlnv),
+    catalogs_(new QList<QSharedPointer<IpxactFile> >()),
+    busDefinitions_(new QList<QSharedPointer<IpxactFile> >()),
+    abstractionDefinitions_(new QList<QSharedPointer<IpxactFile> >()),
+    components_(new QList<QSharedPointer<IpxactFile> >()),
+    abstractors_(new QList<QSharedPointer<IpxactFile> >()),
+    designs_(new QList<QSharedPointer<IpxactFile> >()),
+    designConfigurations_(new QList<QSharedPointer<IpxactFile> >()),
+    generatorChains_(new QList<QSharedPointer<IpxactFile> >())
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: Catalog::Catalog()
+//-----------------------------------------------------------------------------
+Catalog::Catalog(Catalog const& other):
+Document(other),
+    catalogs_(new QList<QSharedPointer<IpxactFile> >()),
+    busDefinitions_(new QList<QSharedPointer<IpxactFile> >()),
+    abstractionDefinitions_(new QList<QSharedPointer<IpxactFile> >()),
+    components_(new QList<QSharedPointer<IpxactFile> >()),
+    abstractors_(new QList<QSharedPointer<IpxactFile> >()),
+    designs_(new QList<QSharedPointer<IpxactFile> >()),
+    designConfigurations_(new QList<QSharedPointer<IpxactFile> >()),
+    generatorChains_(new QList<QSharedPointer<IpxactFile> >())
+{
+    copyIpxactFiles(other);
+}
+
+//-----------------------------------------------------------------------------
+// Function: Catalog::operator=()
+//-----------------------------------------------------------------------------
+Catalog & Catalog::operator=(Catalog const& other)
+{
+	if (this != &other)
+    {
+		Document::operator=(other);
+        
+        catalogs_->clear();
+        busDefinitions_->clear();
+        abstractionDefinitions_->clear();
+        components_->clear();
+        abstractors_->clear();
+        designs_->clear();
+        designConfigurations_->clear();
+        generatorChains_->clear();
+
+        copyIpxactFiles(other);
+	}
+
+	return *this;
+}
+
+//-----------------------------------------------------------------------------
+// Function: Catalog::~Catalog()
+//-----------------------------------------------------------------------------
+Catalog::~Catalog()
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: Catalog::clone()
+//-----------------------------------------------------------------------------
+QSharedPointer<Document> Catalog::clone() const
+{
+	return QSharedPointer<Document>(new Catalog(*this));
+}
+
+//-----------------------------------------------------------------------------
+// Function: Catalog::setVlnv()
+//-----------------------------------------------------------------------------
+void Catalog::setVlnv(VLNV const& vlnv)
+{
+    VLNV CatalogVLNV(vlnv);
+    CatalogVLNV.setType(VLNV::CATALOG);
+
+    Document::setVlnv(CatalogVLNV);
+}
+
[…]
