+++ IPXACTmodels/Catalog/CatalogWriter.cpp
@@ -1,0 +1,182 @@
+//-----------------------------------------------------------------------------
+// File: CatalogWriter.cpp
+//-----------------------------------------------------------------------------
+// Project: Kactus 2
+// Author: Esko Pekkarinen
+// Date: 30.01.2017
+//
+// Description:
+// Writer class for IP-XACT Catalog element.
+//-----------------------------------------------------------------------------
+
+#include "CatalogWriter.h"
+
+#include <IPXACTmodels/Catalog/Catalog.h>
+#include <IPXACTmodels/Catalog/IpxactFile.h>
+
+//-----------------------------------------------------------------------------
+// Function: CatalogWriter::CatalogWriter()
+//-----------------------------------------------------------------------------
+CatalogWriter::CatalogWriter(QObject *parent): DocumentWriter(parent)
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogWriter::~CatalogWriter()
+//-----------------------------------------------------------------------------
+CatalogWriter::~CatalogWriter()
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogWriter::writeCatalog()
+//-----------------------------------------------------------------------------
+void CatalogWriter::writeCatalog(QXmlStreamWriter& writer, QSharedPointer<Catalog> catalog) const
+{
+    writer.writeStartDocument();
+    
+    writeTopComments(writer, catalog);
+
+    writeXmlProcessingInstructions(writer, catalog);
+
+    writer.writeStartElement(QStringLiteral("ipxact:catalog"));
+    writeNamespaceDeclarations(writer);
+
+    writeVLNVElements(writer, catalog->getVlnv());
+
+    writeDescription(writer, catalog);
+
+    writeCatalogs(writer, catalog);
+
+    writeBusDefinitions(writer, catalog);
+
+    writeAbstractionDefinitions(writer, catalog);
+
+    writeComponents(writer, catalog);
+
+    writeAbstractors(writer, catalog);
+
+    writeDesigns(writer, catalog);
+
+    writeDesignConfigurations(writer, catalog);
+
+    writeGeneratorChains(writer, catalog);
+
+    writeVendorExtensions(writer, catalog);
+
+    writer.writeEndElement(); // "ipxact:catalog"
+    writer.writeEndDocument();
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogWriter::writeCatalogs()
+//-----------------------------------------------------------------------------
+void CatalogWriter::writeCatalogs(QXmlStreamWriter& writer, QSharedPointer<Catalog> catalog) const
+{
+    writeIpxactFileGroup(writer, catalog->getCatalogs(), QStringLiteral("ipxact:catalogs"));
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogWriter::writeBusDefinitions()
+//-----------------------------------------------------------------------------
+void CatalogWriter::writeBusDefinitions(QXmlStreamWriter& writer, QSharedPointer<Catalog> catalog) const
+{
+    writeIpxactFileGroup(writer, catalog->getBusDefinitions(), QStringLiteral("ipxact:busDefinitions"));
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogWriter::writeAbstractionDefinitions()
+//-----------------------------------------------------------------------------
+void CatalogWriter::writeAbstractionDefinitions(QXmlStreamWriter& writer, QSharedPointer<Catalog> catalog) const
+{
+    writeIpxactFileGroup(writer, catalog->getAbstractionDefinitions(), QStringLiteral("ipxact:abstractionDefinitions"));
+}
+
[…]
