### 2017-02-01 3601f51
[CORRECTIVE] Reconfigured the memory designer graphics item compression (refs #80).
