+++ designEditors/MemoryDesigner/MemoryConnectionHandler.cpp
@@ -108,6 +108,9 @@ void MemoryConnectionHandler::createMemoryConnections(QSharedPointer<Connectivit
     }
 
     compressGraphicsItems(placedSpaceItems, spaceYPlacement, spaceColumn);
+
+    repositionOverlappingSpaceItems(placedSpaceItems);
+
     if (condenseMemoryItems_)
     {
         repositionCompressedMemoryMaps(placedMapItems, memoryMapColumn);
@@ -157,8 +160,9 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
             QVector<MainMemoryGraphicsItem*> spaceItemChain;
             spaceItemChain.append(connectionStartItem);
 
-            QVector<MemoryConnectionItem*> connectionChain;
+            bool isChainedSpaceConnection = false;
 
+            quint64 spaceChainConnectionBaseAddress = 0;
             foreach(QSharedPointer<ConnectivityInterface> pathInterface, connectionPath)
             {
                 if (pathInterface != startInterface && pathInterface != endInterface)
@@ -173,41 +177,39 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
                     }
                     else if (pathInterface->getMode().compare(QStringLiteral("master"), Qt::CaseInsensitive) == 0)
                     {
+                        spaceChainConnectionBaseAddress += baseAddressNumber;
                         if (pathInterface->isConnectedToMemory())
                         {
                             MainMemoryGraphicsItem* connectionMiddleItem = getMainGraphicsItem(
                                 pathInterface, MemoryDesignerConstants::ADDRESSSPACECOLUMN_NAME);
                             if (connectionMiddleItem)
                             {
+                                connectionMiddleItem->changeAddressRange(spaceChainConnectionBaseAddress);
+
                                 if (filterAddressSpaceChains_)
                                 {
                                     connectionMiddleItem->hide();
                                 }
                                 else
                                 {
-                                    MemoryConnectionItem* newSpaceConnection = createSpaceConnection(
-                                        connectionStartItem, connectionMiddleItem, pathInterface, spaceColumn,
-                                        placedSpaceItems, spaceItemChain, spaceYPlacement);
-                                    if (newSpaceConnection)
-                                    {
-                                        connectionChain.append(newSpaceConnection);
-                                    }
+//                                     MemoryConnectionItem* newSpaceConnection = createSpaceConnection(
+//                                         connectionStartItem, spaceChainConnectionBaseAddress, connectionMiddleItem,
+//                                         pathInterface , spaceColumn, placedSpaceItems, spaceItemChain,
+//                                         spaceYPlacement);
+                                    createSpaceConnection(connectionStartItem, spaceChainConnectionBaseAddress,
+                                        connectionMiddleItem, pathInterface , spaceColumn, placedSpaceItems,
+                                        spaceItemChain, spaceYPlacement);
 
                                     spaceItemChain.append(connectionMiddleItem);
 
                                     connectionStartItem = connectionMiddleItem;
+
+                                    isChainedSpaceConnection = true;
                                 }
                             }
[…]
