### 2017-02-01 fac40b7
[ADDITIVE] Added XML read for catalog elements (refs #296).
