+++ IPXACTmodels/Catalog/CatalogReader.cpp
@@ -1,0 +1,189 @@
+//-----------------------------------------------------------------------------
+// File: CatalogReader.cpp
+//-----------------------------------------------------------------------------
+// Project: Kactus 2
+// Author: Esko Pekkarinen
+// Date: 07.08.2015
+//
+// Description:
+// XML reader class for IP-XACT Parameter element.
+//-----------------------------------------------------------------------------
+
+#include "CatalogReader.h"
+
+#include "Catalog.h"
+#include "IpxactFile.h"
+
+
+//-----------------------------------------------------------------------------
+// Function: CatalogReader::CatalogReader()
+//-----------------------------------------------------------------------------
+CatalogReader::CatalogReader(QObject *parent):
+DocumentReader(parent)
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogReader::~CatalogReader()
+//-----------------------------------------------------------------------------
+CatalogReader::~CatalogReader()
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogReader::createCatalogFrom()
+//-----------------------------------------------------------------------------
+QSharedPointer<Catalog> CatalogReader::createCatalogFrom(QDomNode const& document) const
+{
+    QSharedPointer<Catalog> catalog(new Catalog());
+
+    parseTopComments(document, catalog);
+
+    parseXMLProcessingInstructions(document, catalog);
+
+    QDomNode catalogNode = document.firstChildElement();
+    parseVLNVElements(catalogNode, catalog, VLNV::CATALOG);
+
+    parseDescription(catalogNode, catalog);
+  
+    parseCatalogs(catalogNode, catalog);
+
+    parseBusDefinitions(catalogNode, catalog);
+
+    parseAbstractionDefinitions(catalogNode, catalog);
+
+    parseComponents(catalogNode, catalog);
+
+    parseAbstractors(catalogNode, catalog);
+
+    parseDesigns(catalogNode, catalog);
+    
+    parseDesignCongfigurations(catalogNode, catalog);
+
+    parseGeneratorChains(catalogNode, catalog);
+
+    parseKactusAndVendorExtensions(catalogNode, catalog);
+
+    return catalog;
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogReader::parseCatalogs()
+//-----------------------------------------------------------------------------
+void CatalogReader::parseCatalogs(QDomNode const& catalogNode, QSharedPointer<Catalog> catalog) const
+{   
+    QDomElement catalogsNode = catalogNode.firstChildElement(QStringLiteral("ipxact:catalogs"));
+    
+    catalog->getCatalogs()->append(parseIpxactFileList(catalogsNode));
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogReader::parseBusDefinitions()
+//-----------------------------------------------------------------------------
+void CatalogReader::parseBusDefinitions(QDomNode const& catalogNode, QSharedPointer<Catalog> catalog) const
+{   
+    QDomElement busDefinitionsNode = catalogNode.firstChildElement(QStringLiteral("ipxact:busDefinitions"));
+
+    catalog->getBusDefinitions()->append(parseIpxactFileList(busDefinitionsNode));
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogReader::parseAbstractionDefinitions()
+//-----------------------------------------------------------------------------
+void CatalogReader::parseAbstractionDefinitions(QDomNode const& catalogNode, QSharedPointer<Catalog> catalog) const
+{   
[…]
