+++ IPXACTmodels/Catalog/CatalogReader.h
@@ -1,0 +1,138 @@
+//-----------------------------------------------------------------------------
+// File: CatalogReader.h
+//-----------------------------------------------------------------------------
+// Project: Kactus 2
+// Author: Esko Pekkarinen
+// Date: 01.02.2017
+//
+// Description:
+// XML reader class for IP-XACT Catalog element.
+//-----------------------------------------------------------------------------
+
+#ifndef CATALOGREADER_H
+#define CATALOGREADER_H
+
+#include <QObject>
+#include <QDomNode>
+#include <QSharedPointer>
+
+#include <IPXACTmodels/common/DocumentReader.h>
+#include <IPXACTmodels/ipxactmodels_global.h>
+
+class Catalog;
+class IpxactFile;
+
+//-----------------------------------------------------------------------------
+//! XML reader class for IP-XACT Catalog element.
+//-----------------------------------------------------------------------------
+class IPXACTMODELS_EXPORT CatalogReader : public DocumentReader
+{
+    Q_OBJECT
+public:
+
+	//! The constructor.
+	CatalogReader(QObject* parent = 0);
+
+	//! The destructor.
+	~CatalogReader();
+
+    /*!
+     *  Creates a catalog from XML description.
+     *
+     *      @param [in] document   The XML document to create the catalog from.
+     *
+     *      @return The created catalog.
+     */
+    QSharedPointer<Catalog> createCatalogFrom(QDomNode const& document) const;
+ 
+private:
+
+	//! Disable copying.
+	CatalogReader(CatalogReader const& rhs);
+	CatalogReader& operator=(CatalogReader const& rhs);
+
+    /*!
+     *  Reads the catalog file descriptions to the catalog.
+     *
+     *      @param [in]     catalogNode     The XML description of the catalog.
+     *      @param [in/out] catalog         The catalog to insert the file descriptions into.
+     */
+    void parseCatalogs(QDomNode const& catalogNode, QSharedPointer<Catalog> catalog) const;
+    
+    /*!
+     *  Reads the bus definition file descriptions to the catalog.
+     *
+     *      @param [in]     catalogNode     The XML description of the catalog.
+     *      @param [in/out] catalog         The catalog to insert the file descriptions into.
+     */
+    void parseBusDefinitions(QDomNode const& catalogNode, QSharedPointer<Catalog> catalog) const;
+    
+    /*!
+     *  Reads the abstraction definition file descriptions to the catalog.
+     *
+     *      @param [in]     catalogNode     The XML description of the catalog.
+     *      @param [in/out] catalog         The catalog to insert the file descriptions into.
+     */
+    void parseAbstractionDefinitions(QDomNode const& catalogNode, QSharedPointer<Catalog> catalog) const;
+    
+    /*!
+     *  Reads the component file descriptions to the catalog.
+     *
+     *      @param [in]     catalogNode     The XML description of the catalog.
+     *      @param [in/out] catalog         The catalog to insert the file descriptions into.
+     */
+    void parseComponents(QDomNode const& catalogNode, QSharedPointer<Catalog> catalog) const;
+    
+    /*!
+     *  Reads the abstractor file descriptions to the catalog.
+     *
+     *      @param [in]     catalogNode     The XML description of the catalog.
+     *      @param [in/out] catalog         The catalog to insert the file descriptions into.
+     */
+    void parseAbstractors(QDomNode const& catalogNode, QSharedPointer<Catalog> catalog) const;
+   
+    /*!
+     *  Reads the design file descriptions to the catalog.
+     *
+     *      @param [in]     catalogNode     The XML description of the catalog.
[…]
