### 2017-02-01 1f8efa6
[ADDITIVE] Added initial validation for Catalog element (refs #296).
