+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.151.0
+#define VERSION_FULL           3.3.152.0
 
 #define VERSION_BASEYEAR       0
-#define VERSION_DATE           "2017-01-30"
-#define VERSION_TIME           "13:00:26"
+#define VERSION_DATE           "2017-02-01"
+#define VERSION_TIME           "14:13:32"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        151
+#define VERSION_BUILDNO        152
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,151,0
-#define VERSION_PRODUCT        3,3,151,0
-#define VERSION_FILESTR        "3,3,151,0"
-#define VERSION_PRODUCTSTR     "3,3,151,0"
+#define VERSION_FILE           3,3,152,0
+#define VERSION_PRODUCT        3,3,152,0
+#define VERSION_FILESTR        "3,3,152,0"
+#define VERSION_PRODUCTSTR     "3,3,152,0"
 
 #endif
