### 2017-02-01 967dc1d
[PERFECTIVE] Now the implementation is highlighted for Verilog generation preview.
