+++ tests/Plugins/VerilogGenerator/ImplementationSelection/tst_ImplementationSelection.cpp
@@ -94,11 +94,6 @@ void tst_ImplementationSelection::cleanup()
 //-----------------------------------------------------------------------------
 void tst_ImplementationSelection::testImplementationSelection()
 {
-	QFile existingFile("./generatorOutput.v");
-
-	existingFile.open(QIODevice::WriteOnly);
-	QTextStream outputStream(&existingFile);
-
 	QString content = QString(
 		"module TestComponent #(\n"
 		"    parameter                              dataWidth        = 8,\n"
@@ -115,17 +110,13 @@ void tst_ImplementationSelection::testImplementationSelection()
 		"endmodule\n"
 		); 
 
-	outputStream << content;
-
-	existingFile.close();
-
 	
 
 	QString implementation;
     QString postModule;
     QString error;
 
-	VerilogSyntax::selectImplementation("./generatorOutput.v", implementation, postModule, error);
+	VerilogSyntax::selectImplementation(content, implementation, postModule, error);
 
 	implementation = implementation.trimmed();
 
@@ -137,11 +128,6 @@ void tst_ImplementationSelection::testImplementationSelection()
 //-----------------------------------------------------------------------------
 void tst_ImplementationSelection::testImplementationSelectionWithEvilComments()
 {
-    QFile existingFile("./generatorOutput.v");
-
-    existingFile.open(QIODevice::WriteOnly);
-    QTextStream outputStream(&existingFile);
-
     QString content = QString(
         "module TestComponent #(\n"
         "    parameter                              dataWidth        = 8, // module ending );\n"
@@ -158,17 +144,11 @@ void tst_ImplementationSelection::testImplementationSelectionWithEvilComments()
         "endmodule\n"
         ); 
 
-    outputStream << content;
-
-    existingFile.close();
-
-    
-
     QString implementation;
     QString postModule;
     QString error;
 
-    VerilogSyntax::selectImplementation("./generatorOutput.v", implementation, postModule, error);
+    VerilogSyntax::selectImplementation(content, implementation, postModule, error);
 
     implementation = implementation.trimmed();
 
@@ -180,11 +160,6 @@ void tst_ImplementationSelection::testImplementationSelectionWithEvilComments()
 //-----------------------------------------------------------------------------
 void tst_ImplementationSelection::testImplementationSelectionWithTag()
 {
-	QFile existingFile("./generatorOutput.v");
-
-	existingFile.open(QIODevice::WriteOnly);
-	QTextStream outputStream(&existingFile);
-
 	QString content = QString(
 		"module TestComponent #(\n"
 		"    parameter                              dataWidth        = 8,\n"
@@ -202,17 +177,11 @@ void tst_ImplementationSelection::testImplementationSelectionWithTag()
 		"endmodule\n"
 		); 
 
-	outputStream << content;
-
-	existingFile.close();
-
-	
-
 	QString implementation;
     QString postModule;
     QString error;
 
-	VerilogSyntax::selectImplementation("./generatorOutput.v", implementation, postModule, error);
+    VerilogSyntax::selectImplementation(content, implementation, postModule, error);
 
 	implementation = implementation.trimmed();
 
@@ -224,11 +193,6 @@ void tst_ImplementationSelection::testImplementationSelectionWithTag()
 //-----------------------------------------------------------------------------
 void tst_ImplementationSelection::testImplementationSelectionWithoutParameters()
 {
-	QFile existingFile("./generatorOutput.v");
-
-	existingFile.open(QIODevice::WriteOnly);
-	QTextStream outputStream(&existingFile);
-
 	QString content = QString(
 		"module TestComponent(\n"
[…]
