### 2017-02-02 2a887e6
[PERFECTIVE] Supplemented the manual on colors.
