### 2017-02-02 69690dd
[ADDITIVE] Added new page for Catalogs (refs #296). [CLEANUP] Simplified layout for other new pages.
