+++ common/dialogs/propertyPageDialog/PropertyPageDialog.cpp
@@ -16,33 +16,32 @@
 #include <QHBoxLayout>
 #include <QVBoxLayout>
 #include <QGroupBox>
+#include <QDialogButtonBox>
 
 //-----------------------------------------------------------------------------
 // Function: PropertyPageDialog()
 //-----------------------------------------------------------------------------
 PropertyPageDialog::PropertyPageDialog(QSize const& listIconSize, int iconColumnCount,
-                                       ViewMode viewMode, ApplyMode applyMode, QWidget* parent)
-    : QDialog(parent),
-      iconColumnCount_(iconColumnCount),
-      contentsList_(0),
-      pages_(0),
-      btnOk_(0),
-      viewMode_(viewMode),
-      applyMode_(applyMode)
+    ViewMode viewMode, ApplyMode applyMode, QWidget* parent):
+QDialog(parent),
+    iconColumnCount_(iconColumnCount),
+    contentsList_(new QListWidget(this)),
+    pages_(new QStackedWidget(this)),
+    btnOk_(new QPushButton(tr("OK") , this)),
+    viewMode_(viewMode),
+    applyMode_(applyMode)
 {
-    // Create the contents list.
-    contentsList_ = new QListWidget(this);
     contentsList_->setIconSize(listIconSize);
 
     if (viewMode == VIEW_ICONS)
     {
         contentsList_->setViewMode(QListView::IconMode);
         contentsList_->setMovement(QListView::Static);
+        
+        contentsList_->setUniformItemSizes(true);
+        contentsList_->setSelectionRectVisible(true);
 
-        if (iconColumnCount == 1)
-        {
-            contentsList_->setWrapping(false);
-        }
+        contentsList_->setWrapping(iconColumnCount != 1);
 
         contentsList_->setSpacing(ICON_SPACING);
         contentsList_->setHorizontalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
@@ -53,39 +52,10 @@ PropertyPageDialog::PropertyPageDialog(QSize const& listIconSize, int iconColumn
         contentsList_->setViewMode(QListView::ListMode);
     }
     
-    // Create the stacked pages widget and a layout with a separator (group box).
-    pages_ = new QStackedWidget(this);
-
-    QGroupBox* separator = new QGroupBox(this);
-    separator->setFlat(true);
-
-    QVBoxLayout* pageLayout = new QVBoxLayout();
-    pageLayout->addWidget(pages_);
-    pageLayout->addWidget(separator);
-
-    // Create the layout for the contents list and pages.
-    QHBoxLayout* horizontalLayout = new QHBoxLayout();
-    horizontalLayout->addWidget(contentsList_);
-    horizontalLayout->addLayout(pageLayout, 1);
-
-    // Create the buttons and their layout.
-    btnOk_ = new QPushButton(tr("OK") , this);
-    QPushButton* btnCancel = new QPushButton(tr("Cancel"), this);
-
-    QHBoxLayout* buttonLayout = new QHBoxLayout();
-    buttonLayout->addStretch(1);
-    buttonLayout->addWidget(btnOk_);
-    buttonLayout->addWidget(btnCancel);
-
-    // Create the main layout.
-    QVBoxLayout* mainLayout = new QVBoxLayout(this);
-    mainLayout->addLayout(horizontalLayout, 1);
-    mainLayout->addLayout(buttonLayout);
-
-    connect(btnOk_, SIGNAL(clicked()), this, SLOT(accept()));
-    connect(btnCancel, SIGNAL(clicked()), this, SLOT(reject()));
+    setupLayout();
+   
[…]
