### 2017-02-02 e3edc1d
[CLEANUP] Adjusted the nomeclature of some generation classes.
