+++ Plugins/PluginSystem/GeneratorPlugin/GeneratorConfigurationDialog.cpp
@@ -1,193 +1,0 @@
-﻿//-----------------------------------------------------------------------------
-// File: GeneratorConfigurationDialog.cpp
-//-----------------------------------------------------------------------------
-// Project: Kactus2
-// Author: Janne Virtanen
-// Date: 26.01.2017
-//
-// Description:
-// Dialog for setting file generation options.
-//-----------------------------------------------------------------------------
-
-#include "GeneratorConfigurationDialog.h"
-#include "GeneratorConfiguration.h"
-
-#include <Plugins/VerilogImport/VerilogSyntax.h>
-
-#include <QDialogButtonBox>
-#include <QFileDialog>
-#include <QVBoxLayout>
-#include <QCheckBox>
-#include <QGroupBox>
-#include <QTextCursor>
-
-//-----------------------------------------------------------------------------
-// Function: GeneratorConfigurationDialog::GeneratorConfigurationDialog()
-//-----------------------------------------------------------------------------
-GeneratorConfigurationDialog::GeneratorConfigurationDialog(QSharedPointer<GeneratorConfiguration> configuration,
-	QWidget *parent) : 
-	QDialog(parent,Qt::WindowTitleHint), 
-    configuration_(configuration),
-    viewSelection_(new ViewSelectionWidget(configuration->getViewSelection())),
-    fileOutput_(new FileOutputWidget(configuration->getFileOuput())),
-    generalWarningLabel_(new QLabel)
-{
-    setWindowTitle(tr("Configure file generation for %1.").arg(configuration->getViewSelection()->getTargetLanguage()));
-
-    // Create font for previewing.
-    QFont font("Courier");
-    font.setStyleHint(QFont::Monospace);
-    font.setFixedPitch(true);
-    font.setPointSize(9);
-
-    // Create the previewer.
-    previewer_ = new QPlainTextEdit;
-    previewer_->setFont(font);
-    previewer_->setTabStopWidth(4 * previewer_->fontMetrics().width(' '));
-    previewer_->setReadOnly(true);
-    previewer_->setCursorWidth(0);
-    previewer_->setLineWrapMode(QPlainTextEdit::NoWrap);
-    previewer_->setMinimumWidth(850);
-
-    // Add everything it their proper position in the final layout.
-	QVBoxLayout* leftLayout = new QVBoxLayout();
-    leftLayout->addWidget(viewSelection_);
-    leftLayout->addWidget(fileOutput_);
-    leftLayout->addWidget(generalWarningLabel_);
-
-    // Layout for things coming to the bottom part of the dialog.
-
-    // Add Ok and cancel give the dialog results.
-    QDialogButtonBox* dialogButtons = new QDialogButtonBox(QDialogButtonBox::Ok | QDialogButtonBox::Cancel, 
-        Qt::Horizontal);
-
-    QHBoxLayout* topLayout = new QHBoxLayout;
-    QGroupBox* leftBox = new QGroupBox("Settings");
-    leftBox->setLayout(leftLayout);
-    QGroupBox* rightBox = new QGroupBox("Preview");
-    QVBoxLayout* rightLayout = new QVBoxLayout();
-    rightLayout->addWidget(previewer_);
-    rightBox->setLayout(rightLayout);
-
-    topLayout->addWidget(leftBox);
-    topLayout->addWidget(rightBox);
-
-    QGroupBox* botBox = new QGroupBox("Log");
-    QVBoxLayout* botLayout = new QVBoxLayout();
-    botBox->setLayout(botLayout);
-    botLayout->addWidget(new QPlainTextEdit);
-
-    QVBoxLayout* topmostLayout = new QVBoxLayout(this);
-    topmostLayout->addLayout(topLayout);
[…]
