### 2017-02-02 ea98ff4
[PERFECTIVE] Now the Verilog generation and HDL parsing features a message console.
