+++ Plugins/PluginSystem/GeneratorPlugin/FileOutput.cpp
@@ -1,5 +1,5 @@
 //-----------------------------------------------------------------------------
-// File: FileOutput.h
+// File: FileOutput.cpp
 //-----------------------------------------------------------------------------
 // Project: Kactus2
 // Author: Janne Virtanen
