### 2017-02-02 aa46a71
[CLEANUP] Supplemented the comments of the Verilog generator plugin.
