### 2017-02-03 184c5a4
[CLEANUP] Encapsulated methods of meta instance.
