### 2017-02-03 632fa67
[ADDITIVE] Added the display of overlapping sub items to memory designer (refs #80).
