+++ designEditors/MemoryDesigner/SubMemoryLayout.cpp
@@ -18,7 +18,6 @@
 #include <designEditors/MemoryDesigner/MemoryDesignerGraphicsItem.h>
 #include <designEditors/MemoryDesigner/MemoryDesignerConstants.h>
 #include <designEditors/MemoryDesigner/MemoryConnectionItem.h>
-#include <designEditors/MemoryDesigner/MainMemoryGraphicsItem.h>
 
 #include <QBrush>
 #include <QPen>
@@ -345,34 +344,66 @@ void SubMemoryLayout::setFaultySubItemColor(MemoryDesignerChildGraphicsItem* sub
 //-----------------------------------------------------------------------------
 // Function: SubMemoryLayout::getSubItemHeightAddition()
 //-----------------------------------------------------------------------------
-int SubMemoryLayout::getSubItemHeightAddition()
+quint64 SubMemoryLayout::getSubItemHeightAddition() const
 {
-    int heightAddition = 0;
+    quint64 heightAddition = 0;
 
-    MemoryDesignerGraphicsItem* mainItem = dynamic_cast<MemoryDesignerGraphicsItem*>(this);
-    if (mainItem && !subItemsAreFiltered())
+    if (mainGraphicsItem_ && !subItemsAreFiltered())
     {
-        quint64 mainItemLastAddress = mainItem->getLastAddress();
+        quint64 mainItemLastAddress = mainGraphicsItem_->getLastAddress();
+        quint64 mainItemLow = mainGraphicsItem_->sceneBoundingRect().bottom();
 
-            QMapIterator<quint64, MemoryDesignerChildGraphicsItem*> subItemIterator(getSubMemoryItems());
+        QMapIterator<quint64, MemoryDesignerChildGraphicsItem*> subItemIterator(getSubMemoryItems());
         while (subItemIterator.hasNext())
         {
             subItemIterator.next();
 
             MemoryDesignerChildGraphicsItem* subItem = subItemIterator.value();
             quint64 subItemLastAddress = subItem->getLastAddress();
-            if (subItemLastAddress > mainItemLastAddress)
+            SubMemoryLayout* subLayout = dynamic_cast<SubMemoryLayout*>(subItem);
+            if (subLayout && !subLayout->subItemsAreFiltered())
             {
-                heightAddition = subItem->sceneBoundingRect().bottom();
+                heightAddition = subLayout->getSubItemHeightAddition();
             }
-        }
-        if (heightAddition > 0)
-        {
-            heightAddition = heightAddition - mainItem->sceneBoundingRect().bottom();
+            else if (subItemLastAddress > mainItemLastAddress)
+            {
+                quint64 subItemLow = subItem->sceneBoundingRect().bottom();
+                quint64 subItemAddition = subItemLow - mainItemLow;
+                if (subItemAddition > heightAddition)
+                {
+                    heightAddition = subItemAddition;
+                }
+            }
         }
     }
 
     return heightAddition;
+}
+
+//-----------------------------------------------------------------------------
+// Function: SubMemoryLayout::getHeightWithSubItems()
+//-----------------------------------------------------------------------------
+quint64 SubMemoryLayout::getHeightWithSubItems() const
+{
+    quint64 height = mainGraphicsItem_->boundingRect().height() + getSubItemHeightAddition();
+
+    return height;
+}
+
+//-----------------------------------------------------------------------------
+// Function: SubMemoryLayout::getSceneRectangleWithSubItems()
+//-----------------------------------------------------------------------------
+QRectF SubMemoryLayout::getSceneRectangleWithSubItems() const
+{
[…]
