+++ designEditors/MemoryDesigner/MemoryConnectionHandler.cpp
@@ -192,10 +192,6 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
                                 }
                                 else
                                 {
-//                                     MemoryConnectionItem* newSpaceConnection = createSpaceConnection(
-//                                         connectionStartItem, spaceChainConnectionBaseAddress, connectionMiddleItem,
-//                                         pathInterface , spaceColumn, placedSpaceItems, spaceItemChain,
-//                                         spaceYPlacement);
                                     createSpaceConnection(connectionStartItem, spaceChainConnectionBaseAddress,
                                         connectionMiddleItem, pathInterface , spaceColumn, placedSpaceItems,
                                         spaceItemChain, spaceYPlacement);
@@ -277,23 +273,29 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
             {
                 placeSpaceItemToOtherColumn(connectionStartItem, spaceColumn, connectionEndItem, yTransfer);
 
-                spaceYPlacement = spaceYPlacement - (connectionStartItem->boundingRect().height() +
+                spaceYPlacement = spaceYPlacement - (connectionStartItem->getHeightWithSubItems() +
                     MemoryDesignerConstants::SPACEITEMINTERVAL);
             }
 
             connectionEndItem->changeAddressRange(remappedAddress);
 
-            qreal spaceItemEndPointBefore = connectionStartItem->getSceneEndPoint();
+            quint64 spaceItemEndPointBefore = connectionStartItem->getSceneEndPoint();
 
             MemoryConnectionItem* newConnectionItem =
                 new MemoryConnectionItem(connectionStartItem, remappedAddress, remappedEndAddress,
                 connectionEndItem, spaceColumn->scene(), condenseMemoryItems_, yTransfer);
             connectionsToMemoryMaps_.append(newConnectionItem);
 
-            qreal spaceItemEndPointAfter = connectionStartItem->getSceneEndPoint();
+            quint64 spaceItemEndPointAfter = connectionStartItem->getSceneEndPoint();
 
             spaceYPlacement += spaceItemEndPointAfter - spaceItemEndPointBefore;
-            spaceYPlacement += connectionStartItem->getSubItemHeightAddition();
+
+            qreal endItemChangeToPlacement = connectionEndItem->getSceneRectangleWithSubItems().bottom() +
+                MemoryDesignerConstants::SPACEITEMINTERVAL;
+            if (endItemChangeToPlacement > spaceYPlacement)
+            {
+                spaceYPlacement = endItemChangeToPlacement;
+            }
         }
     }
 }
@@ -373,7 +375,7 @@ void MemoryConnectionHandler::createSpaceConnection(MainMemoryGraphicsItem* conn
         if (recalculateYPosition)
         {
             quint64 newYPosition =
-                connectionMiddleItem->sceneBoundingRect().height() + MemoryDesignerConstants::SPACEITEMINTERVAL;
+                connectionMiddleItem->getHeightWithSubItems() + MemoryDesignerConstants::SPACEITEMINTERVAL;
             if (newYPosition > spaceYPlacement)
             {
                 spaceYPlacement = newYPosition;
[…]
