### 2017-02-03 0054dfd
[CLEANUP] HDLComponentParser is replaced by having meta instance inherit meta component, which also contains formatting functions.
