+++ Plugins/common/HDLParser/MetaComponent.cpp
@@ -1,17 +1,15 @@
 //-----------------------------------------------------------------------------
-// File: HDLComponentParser.cpp
+// File: MetaComponent.cpp
 //-----------------------------------------------------------------------------
 // Project: Kactus2
 // Author: Janne Virtanen
-// Date: 11.08.2016
+// Date: 03.02.2017
 //
 // Description:
 // Class used to parse relevant information from IP-XACT component for HDL generation.
 //-----------------------------------------------------------------------------
 
-#include "HDLComponentParser.h"
-
-#include <library/LibraryManager/libraryinterface.h>
+#include "MetaComponent.h"
 
 #include <IPXACTmodels/AbstractionDefinition/AbstractionDefinition.h>
 
@@ -23,12 +21,6 @@
 #include <IPXACTmodels/Component/View.h>
 #include <IPXACTmodels/Component/ComponentInstantiation.h>
 
-#include <IPXACTmodels/Component/MemoryMap.h>
-#include <IPXACTmodels/Component/MemoryRemap.h>
-#include <IPXACTmodels/Component/MemoryBlockBase.h>
-#include <IPXACTmodels/Component/AddressBlock.h>
-#include <IPXACTmodels/Component/Register.h>
-#include <IPXACTmodels/Component/Field.h>
 #include <IPXACTmodels/Component/RemapState.h>
 #include <IPXACTmodels/Component/RemapPort.h>
 
@@ -37,91 +29,52 @@
 #include <editors/ComponentEditor/common/ListParameterFinder.h>
 
 //-----------------------------------------------------------------------------
-// Function: HDLComponentParser::HDLComponentParser
+// Function: MetaComponent::MetaComponent
 //-----------------------------------------------------------------------------
-HDLComponentParser::HDLComponentParser(LibraryInterface* library,
-    MessagePasser* messages,
+MetaComponent::MetaComponent(MessagePasser* messages,
     QSharedPointer<Component> component,
     QSharedPointer<View> activeView) :
-    MetaInstance(library,
-    messages,
-    component,
-    activeView,
-    QSharedPointer<ComponentInstance>::QSharedPointer(),
-    QSharedPointer<ListParameterFinder>::QSharedPointer(),
-    QSharedPointer<QList<QSharedPointer<ConfigurableElementValue> > >::QSharedPointer())
+    messages_(messages),
+    component_(component),
+    activeView_(activeView),
+    parameters_(new QList<QSharedPointer<Parameter> >()),
+    ports_(new QMap<QString,QSharedPointer<MetaPort> >()),
+    remapStates_(new QList<QSharedPointer<FormattedRemapState> >())
 {
-    // Initialize the parameter finder and formatter.
-    QSharedPointer<ComponentParameterFinder> parameterFinder(new ComponentParameterFinder(component_));
-    parameterFinder->setActiveView(activeView_);
-    formatter_ = QSharedPointer<ExpressionFormatter>(new ExpressionFormatter(parameterFinder));
+    // Try to find a component instantiation for the view.
+    if (activeView_)
+    {
+        activeInstantiation_ = component_->getModel()->
+            findComponentInstantiation(activeView_->getComponentInstantiationRef());
 
-    // Cull again the parameters.
-    parameters_->clear();
+        if (activeInstantiation_)
+        {
+            // If there is a named component instantiation, its module name shall be used.
+            moduleName_ = activeInstantiation_->getModuleName();
+        }
+    }
+
+    if (moduleName_.isEmpty())
+    {
+        // No module name. -> Take the name from the VLNV of the component.
+        moduleName_ = component_->getVlnv().getName();
+    }
+
+    // Must cull the parameters before can use them!
     cullParameters();
[…]
