+++ Plugins/PluginSystem/GeneratorPlugin/IWriterFactory.h
@@ -13,7 +13,6 @@
 #define IWRITERFACTORY_H
 
 #include <Plugins/common/HDLParser/MetaDesign.h>
-#include <Plugins/common/HDLParser/HDLComponentParser.h>
 
 //-----------------------------------------------------------------------------
 // Verilog file generator.
@@ -35,7 +34,7 @@ public:
      *      @return The objects that bundles the writers. Will be null, if could not be created.
      */
     virtual QSharedPointer<GenerationFile> prepareComponent(QString const& outputPath,
-        QSharedPointer<HDLComponentParser> component) = 0;
+        QSharedPointer<MetaComponent> component) = 0;
 
     /*!
      *  Creates writers for the given parsed meta design.
