### 2017-02-03 e30e8dd
[CLEANUP] Encapsulated fields of meta design.
