+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -258,8 +258,8 @@ void tst_HDLParser::testTopLevelComponent()
     QCOMPARE(designs.size(), 1);
     QSharedPointer<MetaDesign> design = designs.first();
 
-    QVERIFY(design->topInstance_);
-    QSharedPointer<MetaInstance> mInstance = design->topInstance_;
+    QVERIFY(design->getTopInstance());
+    QSharedPointer<MetaInstance> mInstance = design->getTopInstance();
 
     QCOMPARE(mInstance->getParameters()->size(), 2);
     QCOMPARE(mInstance->getParameters()->value(0)->name(), QString("dataWidth"));
@@ -300,8 +300,8 @@ void tst_HDLParser::testTopLevelComponentExpressions()
     QCOMPARE(designs.size(), 1);
     QSharedPointer<MetaDesign> design = designs.first();
 
-    QVERIFY(design->topInstance_);
-    QSharedPointer<MetaInstance> mInstance = design->topInstance_;
+    QVERIFY(design->getTopInstance());
+    QSharedPointer<MetaInstance> mInstance = design->getTopInstance();
 
     QCOMPARE(mInstance->getParameters()->size(), 2);
     QCOMPARE(mInstance->getParameters()->value(0)->name(), QString("module"));
@@ -381,9 +381,9 @@ void tst_HDLParser::testHierarchicalConnections()
     QCOMPARE(designs.size(), 1);
     QSharedPointer<MetaDesign> design = designs.first();
 
-    QCOMPARE(design->instances_.size(), 1);
+    QCOMPARE(design->getInstances()->size(), 1);
 
-    QSharedPointer<MetaInstance> mInstance = design->instances_.value("instance1");
+    QSharedPointer<MetaInstance> mInstance = design->getInstances()->value("instance1");
     QCOMPARE(mInstance->getInterfaces()->size(), 2);
 
     QSharedPointer<MetaInterface> mInterface = mInstance->getInterfaces()->value("clk");
@@ -477,9 +477,9 @@ void tst_HDLParser::testHierarchicalConnectionsWithExpressions()
     QCOMPARE(designs.size(), 1);
     QSharedPointer<MetaDesign> design = designs.first();
 
-    QCOMPARE(design->instances_.size(), 1);
+    QCOMPARE(design->getInstances()->size(), 1);
 
-    QSharedPointer<MetaInstance> mInstance = design->instances_.value("instance1");
+    QSharedPointer<MetaInstance> mInstance = design->getInstances()->value("instance1");
     QCOMPARE(mInstance->getInterfaces()->size(), 1);
     QCOMPARE(mInstance->getParameters()->size(), 1);
 
@@ -673,9 +673,9 @@ void tst_HDLParser::testMasterToSlaveInterconnection()
     QCOMPARE(designs.size(), 1);
     QSharedPointer<MetaDesign> design = designs.first();
 
-    QCOMPARE(design->instances_.size(), 2);
+    QCOMPARE(design->getInstances()->size(), 2);
 
-    QSharedPointer<MetaInstance> mInstance = design->instances_.value("sender");
+    QSharedPointer<MetaInstance> mInstance = design->getInstances()->value("sender");
     QCOMPARE(mInstance->getInterfaces()->size(), 1);
 
     QSharedPointer<MetaInterface> mInterface = mInstance->getInterfaces()->value("data_bus");
@@ -691,7 +691,7 @@ void tst_HDLParser::testMasterToSlaveInterconnection()
     QCOMPARE(mpa->logicalBounds_.first, QString("0"));
     QCOMPARE(mpa->logicalBounds_.second, QString("0"));
 
[…]

[…]
