### 2017-02-03 9b689c3
[CLEANUP] Missing comments in HDL parser.
