### 2017-02-06 b7fcd4f
[CLEANUP] Further adjustments to comments of HDL parsing.
