+++ Plugins/common/HDLParser/MetaDesign.cpp
@@ -42,8 +42,8 @@ library_(library),
 messages_(messages),
 design_(design),
 designConf_(designConf),
-topFinder_(new ListParameterFinder),
 topInstance_(topInstance),
+topFinder_(new ListParameterFinder),
 instances_(new QMap<QString,QSharedPointer<MetaInstance> >),
 interconnections_(new QList<QSharedPointer<MetaInterconnection> >),
 adHocWires_(new QList<QSharedPointer<MetaWire> >)
