### 2017-02-06 dd5c172
[ADDITIVE] Added editor for catalogs (refs #296).
