+++ editors/CatalogEditor/CatalogFileView.cpp
@@ -1,0 +1,352 @@
+//-----------------------------------------------------------------------------
+// File: CatalogFileView.h
+//-----------------------------------------------------------------------------
+// Project: Kactus 2
+// Author: Esko Pekkarinen
+// Date: 03.02.2017
+//
+// Description:
+// The widget to display the ipxact files of a catalog.
+//-----------------------------------------------------------------------------
+
+#include "CatalogFileView.h"
+
+#include "CatalogFileColumns.h"
+
+#include <QApplication>
+#include <QClipboard>
+#include <QDesktopServices>
+#include <QDrag>
+#include <QFileInfo>
+#include <QHeaderView>
+#include <QMenu>
+#include <QMessageBox>
+#include <QMimeData>
+#include <QVector>
+
+//-----------------------------------------------------------------------------
+// Function: CatalogFileView::CatalogFileView()
+//-----------------------------------------------------------------------------
+CatalogFileView::CatalogFileView(QWidget* parent): QTreeView(parent),
+    addAction_(new QAction(tr("Add row"), this)),
+    removeAction_(new QAction(tr("Remove row"), this)),
+    cutAction_(new QAction(tr("Cut"), this)),
+    copyAction_(new QAction(tr("Copy"), this)),
+    pasteAction_(new QAction(tr("Paste"), this)),
+    clearAction_(new QAction(tr("Clear"), this)),
+    expandAllIAction_(new QAction(tr("Expand all"), this)),
+    collapseAllIAction_(new QAction(tr("Collapse all"), this))
+{
+	setSelectionBehavior(QAbstractItemView::SelectItems);
+	setSelectionMode(QAbstractItemView::ContiguousSelection);
+
+    setDropIndicatorShown(true);   
+    viewport()->setAcceptDrops(true); 
+    setDragDropMode(QAbstractItemView::DropOnly);
+
+	setupActions();
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogFileView::~CatalogFileView()
+//-----------------------------------------------------------------------------
+CatalogFileView::~CatalogFileView()
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogFileView::contextMenuEvent()
+//-----------------------------------------------------------------------------
+void CatalogFileView::contextMenuEvent(QContextMenuEvent* event)
+{
+    QModelIndex index = indexAt(event->pos());
+
+    QMenu menu(this);
+    menu.addAction(addAction_);
+
+    bool validIndex = index.isValid() && index.parent().isValid();
+
+    // if at least one valid item is selected.
+    if (validIndex)
+    {
+        menu.addAction(removeAction_);
+        menu.addSeparator();
+        menu.addAction(cutAction_);
+        menu.addAction(copyAction_);
+        menu.addAction(pasteAction_);
+        menu.addAction(clearAction_);
+    }
+
+    menu.addSeparator();
+    menu.addAction(expandAllIAction_);
+    menu.addAction(collapseAllIAction_);
+
+    menu.exec(event->globalPos());
+
+    event->accept();
+}
+
+//-----------------------------------------------------------------------------
+// Function: CatalogFileView::onAddAction()
+//-----------------------------------------------------------------------------
+void CatalogFileView::onAddAction()
+{
+    QModelIndexList indexes = selectedIndexes();
+    QModelIndex posToAdd;
+
+    if (!indexes.isEmpty())
+    {
+        posToAdd = indexes.first();
+    }
+
+    emit addItem(posToAdd);
+
+    clearSelection();
+}
+
+//-----------------------------------------------------------------------------
[…]
