### 2017-02-06 5f1646a
[ADDITIVE] Added feature to open document from catalog (refs #296).
