+++ editors/CatalogEditor/CatalogEditor.cpp
@@ -75,6 +75,19 @@ TabDocument(parent, DOC_PROTECTION_SUPPORT),
         proxy, SLOT(onRemoveItem(QModelIndex const&)), Qt::UniqueConnection);
     connect(proxy, SIGNAL(removeItem(QModelIndex const&)),
         fileModel_, SLOT(onRemoveItem(QModelIndex const&)), Qt::UniqueConnection);
+
+    connect(fileView_, SIGNAL(openItem(QModelIndex const&)),
+        proxy, SLOT(onOpenItem(QModelIndex const&)), Qt::UniqueConnection);
+    connect(proxy, SIGNAL(openItem(QModelIndex const&)),
+        fileModel_, SLOT(onOpenItem(QModelIndex const&)), Qt::UniqueConnection);
+
+    connect(fileModel_, SIGNAL(openCatalog(VLNV const&)),
+        this, SIGNAL(openCatalog(VLNV const&)), Qt::UniqueConnection);
+    connect(fileModel_, SIGNAL(openComponent(VLNV const&)),
+        this, SIGNAL(openComponent(VLNV const&)), Qt::UniqueConnection);
+    connect(fileModel_, SIGNAL(openBus(VLNV const&, VLNV const&)),
+        this, SIGNAL(openBus(VLNV const&, VLNV const&)), Qt::UniqueConnection);
+
 }
 
 //-----------------------------------------------------------------------------
