### 2017-02-06 c385d4f
[CORRECTIVE] Fixed a glitch in parsing of physical port bounds.
