+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -57,6 +57,8 @@ private slots:
     void testHierarchicalConnections();
     void testHierarchicalConnectionsWithExpressions();
 
+    void testPortMapping();
+
     void testMasterToSlaveInterconnection();
     void testEmptyBounds();
     void testMasterToSlaveInterconnectionWithExpressions();
@@ -109,7 +111,7 @@ private:
     void mapPortToInterface(QString const& portName, int left, int right, 
         QString const& logicalName, QString const& interfaceName, QSharedPointer<Component> component);
     
-    void addInterfaceToComponent(QString const& interfaceName, QSharedPointer<Component> component);
+    QSharedPointer<BusInterface> addInterfaceToComponent(QString const& interfaceName, QSharedPointer<Component> component, QSharedPointer<ConfigurableVLNVReference> absRef);
 
     QSharedPointer<View> addSenderComponentToLibrary(VLNV senderVLNV, General::InterfaceMode mode);
 
@@ -360,8 +362,8 @@ void tst_HDLParser::testHierarchicalConnections()
     addPort("enable_to_instance", 1, DirectionTypes::IN, topComponent_);
     addPort("full_from_instance", 1, DirectionTypes::OUT, topComponent_);
 
-    addInterfaceToComponent("clk_if", topComponent_);
-    addInterfaceToComponent("data_bus", topComponent_);
+    addInterfaceToComponent("clk_if", topComponent_, clkAbstractionVLNV_);
+    addInterfaceToComponent("data_bus", topComponent_, dataAbstractionVLNV_);
     
     mapPortToInterface("top_clk", "CLK", "clk_if", topComponent_);
     mapPortToInterface("data_to_instance", "DATA", "data_bus", topComponent_);
@@ -435,7 +437,7 @@ void tst_HDLParser::testHierarchicalConnectionsWithExpressions()
     clkPort->setRightBound("0");
     topComponent_->getPorts()->append(clkPort);
 
-    addInterfaceToComponent("clk_if", topComponent_);
+    addInterfaceToComponent("clk_if", topComponent_, clkAbstractionVLNV_);;
 
     mapPortToInterface("top_clk", "CLK", "clk_if", topComponent_);
 
@@ -457,7 +459,7 @@ void tst_HDLParser::testHierarchicalConnectionsWithExpressions()
     instanceClkPort->setRightBound("4-2*2");
     instanceComponent->getPorts()->append(instanceClkPort);
 
-    addInterfaceToComponent("instanceInterface", instanceComponent);
+    addInterfaceToComponent("instanceInterface", instanceComponent, clkAbstractionVLNV_);
     mapPortToInterface("instance_clk", "CLK", "instanceInterface", instanceComponent);
 
 	library_.addComponent(instanceComponent);
@@ -499,12 +501,21 @@ void tst_HDLParser::testHierarchicalConnectionsWithExpressions()
 //-----------------------------------------------------------------------------
 // Function: tst_HDLParser::addInterfaceToComponent()
 //-----------------------------------------------------------------------------
-void tst_HDLParser::addInterfaceToComponent(QString const& interfaceName,
-    QSharedPointer<Component> component)
+QSharedPointer<BusInterface> tst_HDLParser::addInterfaceToComponent(QString const& interfaceName,
+    QSharedPointer<Component> component,
+    QSharedPointer<ConfigurableVLNVReference> absRef)
 {
     QSharedPointer<BusInterface> busInterface(new BusInterface());
     busInterface->setName(interfaceName);
[…]
