### 2017-02-07 0edaf78
[ADDITIVE] Added suggestion of existing IP-XACT items when editing catalog files (refs #296).
