+++ editors/CatalogEditor/CatalogFileModel.cpp
@@ -21,13 +21,18 @@
 #include <IPXACTmodels/Catalog/Catalog.h>
 #include <IPXACTmodels/Catalog/IpxactFile.h>
 
+#include <QFileInfo>
 #include <QIcon>
 #include <QSize>
 #include <QMimeData>
 
+//-----------------------------------------------------------------------------
+// Private namespace.
+//-----------------------------------------------------------------------------
 namespace
 {
-    const QString CATEGORIES[CatalogFileColumns::CATEGORY_COUNT] = {
+    const QString CATEGORIES[CatalogFileColumns::CATEGORY_COUNT] =
+    {
         QLatin1String("Catalogs"),
         QLatin1String("Bus definitions"),
         QLatin1String("Abstraction definitions"),
@@ -39,7 +44,8 @@ namespace
         QLatin1String("Unknown")
     };
 
-    const QString ICONS[CatalogFileColumns::CATEGORY_COUNT] = {
+    const QString ICONS[CatalogFileColumns::CATEGORY_COUNT] =
+    {
         QLatin1String(":icons/common/graphics/catalog.png"),
         QLatin1String(":icons/common/graphics/new-bus.png"),
         QLatin1String(":icons/common/graphics/abs-def.png"),        
@@ -59,7 +65,6 @@ CatalogFileModel::CatalogFileModel(LibraryInterface* library, QObject *parent) :
 QAbstractItemModel(parent),
     catalog_(),
     library_(library),
-    topLevelTypes_(),
     topLevelRows_(),
     locked_(true)
 {
@@ -80,14 +85,6 @@ CatalogFileModel::~CatalogFileModel()
 void CatalogFileModel::setProtection(bool locked)
 {
     locked_ = locked;
-
-    QVector<int> roles(Qt::ForegroundRole);
-
-    for (int i = 0; i < CatalogFileColumns::CATEGORY_COUNT; i++)
-    {
-        QModelIndex parentIndex = index(i, 0, QModelIndex());
-        emit dataChanged(index(0, 0, parentIndex), index(rowCount(parentIndex), CatalogFileColumns::PATH, parentIndex), roles);
-    }    
 } 
 
 //-----------------------------------------------------------------------------
@@ -98,39 +95,20 @@ void CatalogFileModel::refresh(QSharedPointer<Catalog> catalog)
     catalog_ = catalog;
     
     beginResetModel();
-    topLevelTypes_.clear();
     topLevelRows_.clear();
-
-    topLevelTypes_.append(VLNV::CATALOG);
     topLevelRows_.append(catalog->getCatalogs());
-
-    topLevelTypes_.append(VLNV::BUSDEFINITION);
     topLevelRows_.append(catalog->getBusDefinitions());
-
-    topLevelTypes_.append(VLNV::ABSTRACTIONDEFINITION);
     topLevelRows_.append(catalog->getAbstractionDefinitions());
-
-    topLevelTypes_.append(VLNV::COMPONENT);
     topLevelRows_.append(catalog->getComponents());
-
-    topLevelTypes_.append(VLNV::ABSTRACTOR);
     topLevelRows_.append(catalog->getAbstractors());
-
-    topLevelTypes_.append(VLNV::DESIGN);
     topLevelRows_.append(catalog->getDesigns());
-
-    topLevelTypes_.append(VLNV::DESIGNCONFIGURATION);
     topLevelRows_.append(catalog->getDesignConfigurations());
-
-    topLevelTypes_.append(VLNV::GENERATORCHAIN);
     topLevelRows_.append(catalog->getGeneratorChains());
-
-    topLevelTypes_.append(VLNV::INVALID);
[…]
