### 2017-02-07 ab4e590
[ADDITIVE] Added help page for catalog (refs #296). [PERFECTIVE] Allowed URIs in catalog file paths.
