+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.305.0
+#define VERSION_FULL           3.3.318.0
 
 #define VERSION_BASEYEAR       0
 #define VERSION_DATE           "2017-02-07"
-#define VERSION_TIME           "08:43:16"
+#define VERSION_TIME           "10:06:07"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        305
+#define VERSION_BUILDNO        318
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,305,0
-#define VERSION_PRODUCT        3,3,305,0
-#define VERSION_FILESTR        "3,3,305,0"
-#define VERSION_PRODUCTSTR     "3,3,305,0"
+#define VERSION_FILE           3,3,318,0
+#define VERSION_PRODUCT        3,3,318,0
+#define VERSION_FILESTR        "3,3,318,0"
+#define VERSION_PRODUCTSTR     "3,3,318,0"
 
 #endif
