### 2017-02-07 7cc5338
[ADDITIVE] Added feature to automatically add dependent files when drag-dropping VLNV into catalog (refs #296). [PERFECTIVE] Disabled duplicate VLNVs in catalog. [PERFECTIVE] Improved suggestions for VLNVs when editing catalog files.
