### 2017-02-07 8d91500
[CLEANUP] Removed obsolete tree column from catalog file editor (refs #296).
