### 2017-02-08 f81ed32
[CLEANUP] Culled unneeded includes from Verilog generator. Did some other minor cleanups.
