### 2017-02-08 a0ccdcc
[PERFECTIVE] Added catalogs to hierarchy view of the library (refs #296). [ADDITIVE] Added library filter for catalogs. [PERFECTIVE] Separated bus/API/COM library filters to bus and API/COM filters.
