+++ Administrative/releaseFiles/upgrade.cfg
@@ -554,3 +554,9 @@ ADD Workspaces/Design/ComponentEditorFilters/HW/IP/Instantiations Boolean true
 VERSION 3.0.30
 ADD Workspaces/Default/AdHocEditorVisibility Boolean true
 ADD Workspaces/Design/AdHocEditorVisibility Boolean true
+
+VERSION 3.3.340
+ADD Workspaces/Default/LibraryFilters/Type/ShowCatalogs Boolean true
+ADD Workspaces/Design/LibraryFilters/Type/ShowCatalogs Boolean true
+ADD Workspaces/Default/LibraryFilters/Type/ShowApis Boolean true
+ADD Workspaces/Design/LibraryFilters/Type/ShowApis Boolean true
