### 2017-02-08 6025138
[CORRECTIVE] Fixed potential glitch that would have prevented a tie-off value from appearing in generated Verilog.
