+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -94,7 +94,7 @@ private slots:
 
 private:
 
-    void addPort(QString const& portName, int portSize, DirectionTypes::Direction direction, 
+    QSharedPointer<Port> addPort(QString const& portName, int portSize, DirectionTypes::Direction direction, 
         QSharedPointer<Component> component);
 
     void addModuleParameter(QString const& name, QString const& value, QString const& valueID);
@@ -320,12 +320,14 @@ void tst_HDLParser::testTopLevelComponentExpressions()
 //-----------------------------------------------------------------------------
 // Function: tst_HDLParser::addPort()
 //-----------------------------------------------------------------------------
-void tst_HDLParser::addPort(QString const& portName, int portSize, 
+QSharedPointer<Port> tst_HDLParser::addPort(QString const& portName, int portSize, 
     DirectionTypes::Direction direction, QSharedPointer<Component> component)
 {
     QSharedPointer<Port> port = QSharedPointer<Port>(new Port(portName, direction));
     port->setPortSize(portSize);
     component->getPorts()->append(port);
+
+    return port;
 }
 
 //-----------------------------------------------------------------------------
@@ -1636,15 +1638,21 @@ void tst_HDLParser::testAdhocTieOffInComponentInstance()
 
     QString instanceName = "tieOffer";
 
+    QString portDefaultName = "portDefault";
+    QString defaultValueName = "defaultValue";
     QString zeroName = "zeroTieOff";
     QString oneName = "oneTieOff";
     QString naName = "n/aTieOff";
     QString numberedName = "numberedTieOff";
     QString outName = "tieOffOut";
     QString inOutName = "tieOffInOut";
-    QString defaultName = "defaultTieOff";
     QString openName = "openTieOff";
     QString expressionName = "expressionTieOff";
+
+    QSharedPointer<Port> portDefault = addPort(portDefaultName, 4, DirectionTypes::OUT, tieOffComponent);
+    portDefault->setDefaultValue("20");
+    QSharedPointer<Port> defaultPort = addPort(defaultValueName, 4, DirectionTypes::IN, tieOffComponent);
+    defaultPort->setDefaultValue("35");
 
     addPort(zeroName, 2, DirectionTypes::IN, tieOffComponent);
     addPort(oneName, 4, DirectionTypes::IN, tieOffComponent);
@@ -1652,7 +1660,6 @@ void tst_HDLParser::testAdhocTieOffInComponentInstance()
     addPort(numberedName, 10, DirectionTypes::IN, tieOffComponent);
     addPort(outName, 2, DirectionTypes::OUT, tieOffComponent);
     addPort(inOutName, 10, DirectionTypes::INOUT, tieOffComponent);
-    addPort(defaultName, 1, DirectionTypes::IN, tieOffComponent);
     addPort(openName, 1, DirectionTypes::IN, tieOffComponent);
     addPort(expressionName, 1, DirectionTypes::IN, tieOffComponent);
 
@@ -1660,22 +1667,19 @@ void tst_HDLParser::testAdhocTieOffInComponentInstance()
 
     addInstanceToDesign(instanceName, tieOffVLNV, activeView);
 
-    QSharedPointer<Port> defaultPort = tieOffComponent->getPort("defaultTieOff");
-    defaultPort->setDefaultValue("20");
-
     QSharedPointer<Parameter> expressionParameter (new Parameter());
     expressionParameter->setName("expName");
     expressionParameter->setValueId("expID");
     expressionParameter->setValue("5");
[…]
