### 2017-02-08 53ca5ea
[CORRECTIVE] Fixed crashing when unknown item type was added to catalog (refs #296).
