### 2017-02-09 1f829f5
[CORRECTIVE] Fixed issue with selecting items in catalog file list (refs #296).
