### 2017-02-09 0a6e272
[PERFECTIVE] Fixed visual appearance of catalog editor files (refs #296).
