### 2017-02-09 6a8fee5
[CORRECTIVE] Fixed issue with configurable element values when using top level parameters and changing top level view.
