+++ designEditors/common/ComponentInstanceEditor/componentinstanceeditor.cpp
@@ -132,8 +132,11 @@ void ComponentInstanceEditor::setComponentInstance(ComponentItem* component, QSh
 
     containingDesign_ = design;
 	component_ = component;
+
 	instanceFinder_->setComponent(component->componentModel());
-	instanceFinder_->setActiveView(activeViewName);
+	instanceFinder_->setActiveView(designConfiguration->getActiveView(component->name()));
+
+    topFinder_->setActiveView(topComponent_->getModel()->findView(activeViewName));
 
 	// set the vlnv of the component to be displayed
 	vlnvDisplayer_->setVLNV(component->componentModel()->getVlnv(), true);
