### 2017-02-09 9c57a7e
[ADDITIVE] Added filter option for hiding configurable element values whose resolve-property is set to immediate.
