+++ designEditors/common/ComponentInstanceEditor/ConfigurableElementsFilter.cpp
@@ -11,12 +11,14 @@
 
 #include "ConfigurableElementsFilter.h"
 
+#include "ConfigurableElementsColumns.h"
+
 //-----------------------------------------------------------------------------
 // Function: ConfigurableElementsFilter::ConfigurableElementsFilter()
 //-----------------------------------------------------------------------------
 ConfigurableElementsFilter::ConfigurableElementsFilter(QObject* parent):
 QSortFilterProxyModel(parent),
-indexOfRemovedElement_(QModelIndex())
+    showImmediateValues_(false)
 {
 
 }
@@ -30,29 +32,36 @@ ConfigurableElementsFilter::~ConfigurableElementsFilter()
 }
 
 //-----------------------------------------------------------------------------
+// Function: ConfigurableElementsView::setShowImmediateValues()
+//-----------------------------------------------------------------------------
+void ConfigurableElementsFilter::setShowImmediateValues(bool show)
+{
+    showImmediateValues_ = show;
+    invalidate();
+}
+
+//-----------------------------------------------------------------------------
 // Function: ConfigurableElementsView::onRemoveItem()
 //-----------------------------------------------------------------------------
 void ConfigurableElementsFilter::onRemoveItem(QModelIndex const& filterIndex)
 {
-    indexOfRemovedElement_ = mapToSource(filterIndex);
-
-    emit removeItem(indexOfRemovedElement_);
+    emit removeItem(mapToSource(filterIndex));
 }
 
 //-----------------------------------------------------------------------------
 // Function: ConfigurableElementsFilter::filterAcceptsRow()
 //-----------------------------------------------------------------------------
-bool ConfigurableElementsFilter::filterAcceptsRow(int source_row, const QModelIndex &source_parent) const
+bool ConfigurableElementsFilter::filterAcceptsRow(int source_row, QModelIndex const& source_parent) const
 {
     if (!source_parent.isValid())
     {
-        QModelIndex itemIndex = sourceModel()->index(source_row, 0, source_parent);
-
-        if (sourceModel()->rowCount(itemIndex) == 0)
-        {
-            return false;
-        }
+        QModelIndex categoryIndex = sourceModel()->index(source_row, 0, QModelIndex());
+        return sourceModel()->rowCount(categoryIndex) != 0;
     }
 
-    return true;
+    else
+    {
+        QModelIndex itemIndex = sourceModel()->index(source_row, ConfigurableElementsColumns::VALUE, source_parent);
+        return (showImmediateValues_ || itemIndex.flags().testFlag(Qt::ItemIsEditable));
+    }
 }
