+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.386.0
+#define VERSION_FULL           3.3.393.0
 
 #define VERSION_BASEYEAR       0
 #define VERSION_DATE           "2017-02-09"
-#define VERSION_TIME           "11:15:48"
+#define VERSION_TIME           "12:19:01"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        386
+#define VERSION_BUILDNO        393
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,386,0
-#define VERSION_PRODUCT        3,3,386,0
-#define VERSION_FILESTR        "3,3,386,0"
-#define VERSION_PRODUCTSTR     "3,3,386,0"
+#define VERSION_FILE           3,3,393,0
+#define VERSION_PRODUCT        3,3,393,0
+#define VERSION_FILESTR        "3,3,393,0"
+#define VERSION_PRODUCTSTR     "3,3,393,0"
 
 #endif
