### 2017-02-09 2b2a2ee
[CORRECTIVE] Fixed glitches in HDL parsing that may occur if design configuration is missing or is not complete. Moverover, renamed GenerationFile to GenerationOutput and FileOutput to OutputControl.
