+++ Plugins/PluginSystem/GeneratorPlugin/IWriterFactory.h
@@ -33,7 +33,7 @@ public:
      *
      *      @return The objects that bundles the writers. Will be null, if could not be created.
      */
-    virtual QSharedPointer<GenerationFile> prepareComponent(QString const& outputPath,
+    virtual QSharedPointer<GenerationOutput> prepareComponent(QString const& outputPath,
         QSharedPointer<MetaComponent> component) = 0;
 
     /*!
@@ -43,7 +43,7 @@ public:
      *
      *      @return The objects that bundles the writers. Will be null, if could not be created.
      */
-    virtual QSharedPointer<GenerationFile> prepareDesign(QSharedPointer<MetaDesign> design) = 0;
+    virtual QSharedPointer<GenerationOutput> prepareDesign(QSharedPointer<MetaDesign> design) = 0;
     
     /*!
      *  Returns the language of the factory.
