### 2017-02-09 41402dc
[CORRECTIVE] Minor interface change for HDLGenerationDialog.
