### 2017-02-13 6f5ac2a
[ADDITIVE] Added editors for isPresent in memory maps and address blocks.
