### 2017-02-14 0832a9d
[CORRECTIVE] Fixed memory map visualization bug where items were not drawn correctly when presence changes.
