### 2017-02-14 0454e64
[ADDITIVE] Added memory map presence to be evaluated when analyzing memory connectivity.
