+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.441.0
+#define VERSION_FULL           3.3.442.0
 
 #define VERSION_BASEYEAR       0
 #define VERSION_DATE           "2017-02-14"
-#define VERSION_TIME           "08:52:54"
+#define VERSION_TIME           "08:56:00"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        441
+#define VERSION_BUILDNO        442
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,441,0
-#define VERSION_PRODUCT        3,3,441,0
-#define VERSION_FILESTR        "3,3,441,0"
-#define VERSION_PRODUCTSTR     "3,3,441,0"
+#define VERSION_FILE           3,3,442,0
+#define VERSION_PRODUCT        3,3,442,0
+#define VERSION_FILESTR        "3,3,442,0"
+#define VERSION_PRODUCTSTR     "3,3,442,0"
 
 #endif
