### 2017-02-14 de2aef2
[ADDITIVE] Added editor for isPresent in Address spaces and segments. [ADDITIVE] Added changes to address space visualization when segment isPresent is edited.
