+++ designEditors/MemoryDesigner/ConnectivityGraphFactory.cpp
@@ -178,35 +178,41 @@ void ConnectivityGraphFactory::addAddressSpaceMemories(QSharedPointer<Connectivi
 
     foreach (QSharedPointer<AddressSpace> space, *component->getAddressSpaces())
     {
-        QString spaceIdentifier = instanceIdentifier + "." + space->name();
-        QSharedPointer<MemoryItem> spaceItem(new MemoryItem(space->name(), "addressSpace"));
-        spaceItem->setIdentifier(spaceIdentifier);
-        spaceItem->setAUB(space->getAddressUnitBits());
-        spaceItem->setAddress("0");
-
-        if (!space->getRange().isEmpty())
+        if (space->getIsPresent().isEmpty() ||
+            expressionParser_->parseExpression(space->getIsPresent()).toInt() == 1)
         {
-            spaceItem->setRange(expressionParser_->parseExpression(space->getRange()));
-        }
+            QString spaceIdentifier = instanceIdentifier + "." + space->name();
+            QSharedPointer<MemoryItem> spaceItem(new MemoryItem(space->name(), "addressSpace"));
+            spaceItem->setIdentifier(spaceIdentifier);
+            spaceItem->setDisplayName(space->displayName());
+            spaceItem->setAUB(space->getAddressUnitBits());
+            spaceItem->setAddress("0");
 
-        if (!space->getWidth().isEmpty())
-        {
-            spaceItem->setWidth(expressionParser_->parseExpression(space->getWidth()));
-        }
-
-        newInstance->addMemory(spaceItem);
-
-        foreach(QSharedPointer<Segment> segment, *space->getSegments())
-        {
-            if (segment->getIsPresent().isEmpty() ||
-                expressionParser_->parseExpression(segment->getIsPresent()).toInt() == 1)
+            if (!space->getRange().isEmpty())
             {
-                QSharedPointer<MemoryItem> segmentItem(new MemoryItem(segment->name(), "segment"));
-                segmentItem->setIdentifier(spaceIdentifier + '.' + segment->name());
-                segmentItem->setRange(expressionParser_->parseExpression(segment->getRange()));
-                segmentItem->setOffset(expressionParser_->parseExpression(segment->getAddressOffset()));
+                spaceItem->setRange(expressionParser_->parseExpression(space->getRange()));
+            }
 
-                spaceItem->addChild(segmentItem);
+            if (!space->getWidth().isEmpty())
+            {
+                spaceItem->setWidth(expressionParser_->parseExpression(space->getWidth()));
+            }
+
+            newInstance->addMemory(spaceItem);
+
+            foreach(QSharedPointer<Segment> segment, *space->getSegments())
+            {
+                if (segment->getIsPresent().isEmpty() ||
+                    expressionParser_->parseExpression(segment->getIsPresent()).toInt() == 1)
+                {
+                    QSharedPointer<MemoryItem> segmentItem(new MemoryItem(segment->name(), "segment"));
+                    segmentItem->setIdentifier(spaceIdentifier + '.' + segment->name());
+                    segmentItem->setDisplayName(segment->displayName());
+                    segmentItem->setRange(expressionParser_->parseExpression(segment->getRange()));
+                    segmentItem->setOffset(expressionParser_->parseExpression(segment->getAddressOffset()));
+
+                    spaceItem->addChild(segmentItem);
+                }
             }
         }
     }
@@ -248,6 +254,7 @@ QSharedPointer<MemoryItem> ConnectivityGraphFactory::createMemoryMapData(QShared
 
[…]
