+++ editors/ComponentEditor/addressSpaces/segmenteditor.cpp
@@ -111,14 +111,6 @@ SegmentEditor::~SegmentEditor()
 }
 
 //-----------------------------------------------------------------------------
-// Function: SegmentEditor::isValid()
-//-----------------------------------------------------------------------------
-bool SegmentEditor::isValid() const
-{
-	return model_.isValid();
-}
-
-//-----------------------------------------------------------------------------
 // Function: SegmentEditor::refresh()
 //-----------------------------------------------------------------------------
 void SegmentEditor::refresh()
