### 2017-02-15 8366190
[ADDITIVE] Now the Veriog inout interconnections are generated without assignments.
