### 2017-02-16 779930f
[ADDITIVE] Added syntax highlighting for Verilog generation preview (refs #330). [CLEANUP] Refactored CSource highlighting to be applicable for other languages as well.
