### 2017-02-16 e21487b
[CLEANUP] Moved expression editor visual settings to class itself from various locations.
