### 2017-02-16 ab2cb3c
[PERFECTIVE] Allowed expressions in values for parameters with a choice in effect.
