+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.474.0
+#define VERSION_FULL           3.3.479.0
 
 #define VERSION_BASEYEAR       0
 #define VERSION_DATE           "2017-02-16"
-#define VERSION_TIME           "08:30:35"
+#define VERSION_TIME           "09:00:01"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        474
+#define VERSION_BUILDNO        479
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,474,0
-#define VERSION_PRODUCT        3,3,474,0
-#define VERSION_FILESTR        "3,3,474,0"
-#define VERSION_PRODUCTSTR     "3,3,474,0"
+#define VERSION_FILE           3,3,479,0
+#define VERSION_PRODUCT        3,3,479,0
+#define VERSION_FILESTR        "3,3,479,0"
+#define VERSION_PRODUCTSTR     "3,3,479,0"
 
 #endif
