### 2017-02-16 1b751a9
[ADDITIVE] Added display name property to memory items in connectivity analysis.
