### 2017-02-16 a27ea09
[CORRECTIVE] Fixed a glitch in Verilog generation that caused one bit slicing to be omitted.
