### 2017-02-16 4751456
[CORRECTIVE] Fixed another glitch in Verilog generation that misplaced one-bit assignments.
