### 2017-02-16 4393f54
[CORRECTIVE] Fixed more glitches on generating one bit connections. Moreover, now the hierarchical inout connections pick slicing from top port. (refs #331).
