### 2017-02-20 65107c7
[ADDITIVE] Added overlapping and out of bounds register fields to memory designer (refs #80).
