+++ designEditors/MemoryDesigner/FieldGraphicsItem.cpp
@@ -16,29 +16,50 @@
 #include <designEditors/MemoryDesigner/MemoryItem.h>
 #include <designEditors/MemoryDesigner/MemoryDesignerChildGraphicsItem.h>
 #include <designEditors/MemoryDesigner/MemoryDesignerConstants.h>
+#include <designEditors/MemoryDesigner/FieldOverlapItem.h>
 
 #include <QBrush>
 #include <QFont>
 #include <QFontMetrics>
+#include <QGraphicsScene>
 
 //-----------------------------------------------------------------------------
 // Function: FieldGraphicsItem::FieldGraphicsItem()
 //-----------------------------------------------------------------------------
 FieldGraphicsItem::FieldGraphicsItem(QString const& fieldName, quint64 fieldOffset, quint64 fieldLastBit,
     qreal fieldWidth, quint64 fieldHeight, bool isEmptyField, QFont labelFont, QString const& containingInstance,
-    MemoryDesignerGraphicsItem* parentItem):
+    bool isOutsideRegister, MemoryDesignerGraphicsItem* parentItem):
 MemoryDesignerChildGraphicsItem(fieldName, QStringLiteral("Field"), fieldOffset, fieldHeight, fieldWidth,
     containingInstance, parentItem),
 combinedRangeLabel_(new QGraphicsTextItem("", this)),
-fieldName_(fieldName)
+fieldName_(fieldName),
+overlapFields_(),
+overlapIcon_(new QGraphicsPixmapItem(QPixmap(":icons/common/graphics/triangle_arrow_down.png"), this)),
+overlapAreaRectangle_(0)
 {
+    qreal overlapIconPositionX = boundingRect().right() - overlapIcon_->boundingRect().width() - (GridSize / 2);
+    qreal overlapIconPositionY = boundingRect().top() + (GridSize / 2);
+    overlapIcon_->setPos(overlapIconPositionX, overlapIconPositionY);
+    overlapIcon_->setVisible(false);
+
     getNameLabel()->setFont(labelFont);
     combinedRangeLabel_->setFont(labelFont);
 
     setupLabels(fieldOffset, fieldLastBit);
     setupToolTip(QStringLiteral("Field"));
     setLabelPositions();
-    setColors(KactusColors::FIELD_COLOR, isEmptyField);
+
+    QColor fieldColor = KactusColors::FIELD_COLOR;
+    if (isOutsideRegister)
+    {
+        QString toolTipAddition = QStringLiteral(
+            "<br><br><b><font color=\"red\">This field is not contained within register</font></b>");
+        addToToolTip(toolTipAddition);
+
+        fieldColor = KactusColors::MISSING_COMPONENT;
+    }
+         
+    setColors(fieldColor, isEmptyField);
 }
 
 //-----------------------------------------------------------------------------
@@ -149,6 +170,38 @@ void FieldGraphicsItem::changeWidth(qreal widthChange)
 }
 
 //-----------------------------------------------------------------------------
+// Function: FieldGraphicsItem::resizeAndRepositionOverlappingItems()
+//-----------------------------------------------------------------------------
+void FieldGraphicsItem::resizeAndRepositionOverlappingItems()
+{
+    if (!overlapFields_.isEmpty())
+    {
+        qreal overlapRectangleLeft = sceneBoundingRect().left();
+        qreal overlapRectangleRight = sceneBoundingRect().right();
+        foreach (FieldGraphicsItem::CombinedOverlappingFieldItem overlapCombination, overlapFields_)
+        {
+            FieldOverlapItem* overlapItem = overlapCombination.overlapItem_;
+            FieldGraphicsItem* overlappedField = overlapCombination.overlappedField_;
+            QRectF overlapFieldRectangle = overlappedField->sceneBoundingRect();
[…]
