### 2017-02-20 a1e546c
[ADDITIVE] Non-empty display names are displayed instead of names (refs #80).
