### 2017-02-21 c650ff5
[PERFECTIVE] Cleaned memory designer files (refs #80).
