### 2017-02-22 4a856b6
[PERFECTIVE] Added 'wire' as one of the suggested items for port type (refs #333).
