### 2017-02-22 f6e929a
[ADDITIVE] Added possibility to open a containing component in memory designer (refs #80).
