+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.487.0
+#define VERSION_FULL           3.3.493.0
 
 #define VERSION_BASEYEAR       0
-#define VERSION_DATE           "2017-02-21"
-#define VERSION_TIME           "14:16:23"
+#define VERSION_DATE           "2017-02-22"
+#define VERSION_TIME           "09:39:28"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        487
+#define VERSION_BUILDNO        493
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,487,0
-#define VERSION_PRODUCT        3,3,487,0
-#define VERSION_FILESTR        "3,3,487,0"
-#define VERSION_PRODUCTSTR     "3,3,487,0"
+#define VERSION_FILE           3,3,493,0
+#define VERSION_PRODUCT        3,3,493,0
+#define VERSION_FILESTR        "3,3,493,0"
+#define VERSION_PRODUCTSTR     "3,3,493,0"
 
 #endif
