### 2017-02-22 60ab9a6
[CLEANUP] Cleaned the code used to write Verilog for inout connections. (refs #331).
