### 2017-02-23 49e205b
[PERFECTIVE] Added visual icon and tooltip of possible values to configurable elements with a choice set.
