+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.491.0
+#define VERSION_FULL           3.3.504.0
 
 #define VERSION_BASEYEAR       0
-#define VERSION_DATE           "2017-02-22"
-#define VERSION_TIME           "09:03:53"
+#define VERSION_DATE           "2017-02-23"
+#define VERSION_TIME           "12:09:42"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        491
+#define VERSION_BUILDNO        504
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,491,0
-#define VERSION_PRODUCT        3,3,491,0
-#define VERSION_FILESTR        "3,3,491,0"
-#define VERSION_PRODUCTSTR     "3,3,491,0"
+#define VERSION_FILE           3,3,504,0
+#define VERSION_PRODUCT        3,3,504,0
+#define VERSION_FILESTR        "3,3,504,0"
+#define VERSION_PRODUCTSTR     "3,3,504,0"
 
 #endif
