### 2017-02-27 #151 @labarba
@jflorence Hi there—would you give us an update here of what your assessment is so far? I see that you've been ticking off items on the review checklist, yet some remain. Can you summarize the status of this review? Thanks!
