### 2017-03-01 #151 @jflorence
@labarba @TermosPullo Hello,
Here is my assessment so far:
I would grade the software as MINOR REVISIONS needed.
Mostly, those revisions are in the documentation.
The detail of those can be found as issues 336, 337, 338, 339, 340 on the SW's issues tracker https://kactus2.cs.tut.fi/issues/

Overall, the software meets the criteria of the JOSS.

To address the specific concern you had about whether the SW fits the scope ("Your software should have a research application -Should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better"):

Kactus2 is a design environment for electronic IP cores, system-on-chips and boards, as well as associated software.
Kactus2 indeed has a research as well as industrial application, as it allows, for instance, better communication of an electronic system design. It can not only accelerate the design itself, but also improve transfer of knowledge between research groups.

Now, about the specific points (reviewer questions at the top):

Conflict of interest:
+I confirm that there are no conflicts of interest for me to review this work.

General Checks:
+The software has a repository.
+The license is GNU GPL2. It is also available as dual-licensing for commercial use.
-The version I reviewed was 3.3.59 and not 3.2.173. Is this a problem?
+Authorship: As per the git log, TermosPullo (JanneVirtanen) has made significant contributions to the software.

Functionality
+Installation: Installation went smoothly
+Functionality: Functionality is as described
+Performance: No performance claims made

Documentation:
-A statement of need: The authors clearly state what problems the software is designed to solve and who the target audience is. However, this is done on the website funbase.cs.tut.fi and not in the software itself.
+Installation instruction: The dependencies are clearly stated
-Example usage: Examples cannot be seen straightforward. However, by spending some time looking around, we can find some videos on the website funbase.cs.tut.fi
+Functionality documentation: API methods are documented (functionality, parameters, return value, etc.)
+Automated tests: the project has unit tests
-Community guidelines: When exiting the software, the user has the opportunity to complete a feedback form and submit it to the SW's developers.
This is nice (I would suggest keeping that window on screen for a bit longer). Also, in the SW itself, the "About" menu item contains an email address to contact the SW team.
However, I do not find how to contribute to the software. If it is indicated somewhere, it needs to be more obvious (eg in the README.md)
Also, the in-software help is poor.

Software paper:
+Authors: The file paper.md states the authors and their affiliation
+A statement of need: A summary is included, but does not describe the target audience
+References: paper.md references two documents, both of which have a doi indicated.
