### 2017-03-02 c851b98
[CORRECTIVE] Updated the paper to be more descriptive about target audience and what you can do with Kactus2. Moreover, updated the description in README.md to correspond better the current focus of the project, as well as include the community guide lines. (refs #338, #339, #340)
