### 2017-03-03 bb48745
[CORRECTIVE] Now the read IP-XACT documents will remember their XML namespaces and write them to their files when saving. (refs #317, #282)
