### 2017-03-03 9cd8f8e
[CORRECTIVE] Now ad-hoc parts are assigned to wire as [selection_width-1:0] while assignment to port uses the part bounds [left:right] (refs #335)
