+++ Plugins/VerilogGenerator/VerilogGeneratorPlugin.cpp
@@ -52,7 +52,7 @@ QString VerilogGeneratorPlugin::getName() const
 //-----------------------------------------------------------------------------
 QString VerilogGeneratorPlugin::getVersion() const
 {
-    return "2.0c";
+    return "2.0d";
 }
 
 //-----------------------------------------------------------------------------
