### 2017-03-06 abe0ce7
[CORRECTIVE] Fixed multiple memory connections to one memory map in memory designer (refs #332). [PERFECTIVE] Changed the look of address space bridge connections.
