### 2017-03-06 #151 @labarba
Thank you for your review, @jflorence !!

@TermosPullo Back to you. Minor revisions, as described above. Let us know here when you think the revision is ready.
