### 2017-03-07 93f5a73
[CORRECTIVE] Removed address space reference and base address from mirrored-master bus interface. [PERFECTIVE] Changed coloring of bit steering set for mirrored-master, system and mirrored-system bus interfaces.
