### 2017-03-07 bda92c1
[PERFECTIVE] Enabled keyboard navigation for library tree view. [CORRECTIVE] Removed an incorrect error message on opening a component.
