### 2017-03-07 015c4fe
[CLEANUP] Removed the experimental definition file support from abstraction definition.
