### 2017-03-07 44c0bec
[CLEANUP] Component reader no longer converts software views to IP-XACT views.
