### 2017-03-07 f5c518e
[CORRECTIVE] Fixed a crash that would occur when an interface without selected mode is in a design diagram. [CORRECTIVE] Now red color is reserved for invalid and modeless interfaces. System and mirrored system are purple and light purple, respectively.
