### 2017-03-08 e1f6290
[CLEANUP] Concentrated most of the color definitions to common/KactusColors.h.
