+++ editors/ComponentEditor/common/ParameterFinder.h
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: ParameterFinder.h
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Mikko Teuho
 // Date: 15.01.2015
 //
