+++ editors/ComponentEditor/treeStructure/componenteditorcpusitem.cpp
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: componenteditorcpusitem.cpp
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author:Antti Kamppi
 // Date: 09.05.2012
 //
