+++ editors/ComponentEditor/visualization/memoryvisualizationitem.cpp
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: memoryvisualizationitem.cpp
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Antti Kamppi
 // Date: 15.10.2012
 //
