+++ editors/ComponentEditor/instantiations/ComponentInstantiationDisplayer.cpp
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: ComponentInstantiationDisplayer.cpp
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Esko Pekkarinen
 // Date: 26.01.2016
 //
