+++ designEditors/SystemDesign/SWPortItem.cpp
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: SWPortItem.cpp
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Joni-Matti M‰‰tt‰
 // Date: 2.5.2012
 //
