+++ editors/ComponentEditor/general/generaleditor.cpp
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: generaleditor.cpp
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Joni-Matti M‰‰tt‰
 // Date: 7.2.2011
 //
