+++ designEditors/common/ConfigurationEditor/activeviewmodel.h
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: activeviewmodel.h
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Antti Kamppi
 // Date: 24.08.2011
 //
