+++ editors/ComponentEditor/treeStructure/componenteditorviewsitem.cpp
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: componenteditorviewsitem.cpp
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Antti Kamppi
 // Date: 09.05.2012
 //
