+++ editors/ComponentEditor/software/SWPropertiesEditor.h
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: SWPropertiesEditor.h
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Joni-Matti M‰‰tt‰
 // Date: 21.5.2012
 //
