+++ designEditors/common/DesignDiagramResolver.h
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: DesignDiagramResolver.h
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Mikko Teuho
 // Date: 04.05.2016
 //
