+++ editors/ComponentEditor/memoryMaps/enumeratedvaluedelegate.cpp
@@ -1,7 +1,7 @@
 //-----------------------------------------------------------------------------
 // File: enumeratedvaluedelegate.cpp
 //-----------------------------------------------------------------------------
-// Project: Kactus 2
+// Project: Kactus2
 // Author: Antti Kamppi
 // Date: 28.08.2012
 //
