### 2017-03-09 e319e02
[PERFECTIVE] Added keyboard shortcuts to ribbon button tooltips.
