### 2017-03-09 dc1f39a
[CORRECTIVE] Fixed a crash that would occur when design does not have a configuration. [CLEANUP] Further concentrated the color definitions.
