+++ Plugins/PluginSystem/GeneratorPlugin/HDLGenerationDialog.cpp
@@ -193,7 +193,7 @@ void HDLGenerationDialog::onSelectedFileChanged(QSharedPointer<GenerationOutput>
         return;
     }
 
-    QColor const& highlightColor = QColor::fromRgb(183,225,252);
+    QColor const& highlightColor = KactusColors::WARNING;
 
     QTextCursor cursor = previewer_->textCursor();
     cursor.setPosition(implementationStart);
