+++ designEditors/common/ComponentInstanceEditor/componentinstanceeditor.cpp
@@ -136,7 +136,11 @@ void ComponentInstanceEditor::setComponentInstance(ComponentItem* component, QSh
 	component_ = component;
 
 	instanceFinder_->setComponent(component->componentModel());
-	instanceFinder_->setActiveView(designConfiguration->getActiveView(component->name()));
+
+    if (designConfiguration)
+    {
+	    instanceFinder_->setActiveView(designConfiguration->getActiveView(component->name()));
+    }
 
     
     topFinder_->setActiveView(topComponent_->getModel()->findView(activeViewName));    
