### 2017-03-09 9cd325b
[ADDITIVE] Added options to expand/collapse VLNV library tree items recursively.
