+++ library/LibraryManager/librarytreeview.cpp
@@ -57,8 +57,10 @@ QTreeView(parent),
     openSystemAction_(new QAction(tr("Open System Design"), this)),
     openXmlAction_(new QAction(tr("Open XML File"), this)),
     openContainingFolderAction_(new QAction(tr("Open Containing Folder"), this)),
-    expandAllAction_(new QAction(tr("Expand all"), this)),
-    collapseAllAction_(new QAction(tr("Collapse all"), this))
+    expandChilds_(new QAction(tr("Expand Branches"), this)),
+    collapseChilds_(new QAction(tr("Collapse Branches"), this)),
+    expandAllAction_(new QAction(tr("Expand All"), this)),
+    collapseAllAction_(new QAction(tr("Collapse All"), this))
 {
 	Q_ASSERT_X(filter, "LibraryTreeView constructor", "Null filter pointer given");
 
@@ -217,91 +219,15 @@ void LibraryTreeView::contextMenuEvent(QContextMenuEvent* event)
 	menu.addAction(deleteAction_);
 
     menu.addSeparator();
+    if (model()->rowCount(current) > 0)
+    {
+        menu.addAction(expandChilds_);
+        menu.addAction(collapseChilds_);
+    }
     menu.addAction(expandAllAction_);
     menu.addAction(collapseAllAction_);
 
 	menu.exec(event->globalPos());
-}
-
-//-----------------------------------------------------------------------------
-// Function: LibraryTreeView::setupActions()
-//-----------------------------------------------------------------------------
-void LibraryTreeView::setupActions()
-{
-	openDesignAction_->setStatusTip(tr("Open a HW design"));
-	openDesignAction_->setToolTip(tr("Open a HW design"));
-	connect(openDesignAction_, SIGNAL(triggered()),	this, SLOT(onOpenDesign()), Qt::UniqueConnection);
-
-    openSWDesignAction_->setStatusTip(tr("Open a SW design"));
-    openSWDesignAction_->setToolTip(tr("Open a SW design"));
-    connect(openSWDesignAction_, SIGNAL(triggered()), this, SLOT(onOpenSWDesign()), Qt::UniqueConnection);
-
-    openMemoryDesignAction_->setStatusTip(tr("Open a memory design"));
-    openMemoryDesignAction_->setToolTip(tr("Open a memory design"));
-    connect(openMemoryDesignAction_, SIGNAL(triggered()), this, SLOT(onOpenMemoryDesign()), Qt::UniqueConnection);
-
-	openComponentAction_->setStatusTip(tr("Open component editor"));
-	openComponentAction_->setToolTip(tr("Open component editor"));
-	connect(openComponentAction_, SIGNAL(triggered()), this, SLOT(onOpenComponent()), Qt::UniqueConnection);
-
-	createNewDesignAction_->setStatusTip(tr("Create new HW design"));
-	createNewDesignAction_->setToolTip(tr("Create new HW design"));
-	connect(createNewDesignAction_, SIGNAL(triggered()), this, SLOT(onCreateDesign()), Qt::UniqueConnection);
-
-    createNewSWDesignAction_->setStatusTip(tr("Create new SW design"));
-    createNewSWDesignAction_->setToolTip(tr("Create new SW design"));
-    connect(createNewSWDesignAction_, SIGNAL(triggered()), this, SLOT(onCreateSWDesign()), Qt::UniqueConnection);
-
-    createNewSystemDesignAction_->setStatusTip(tr("Create new system design"));
-    createNewSystemDesignAction_->setToolTip(tr("Create new system design"));
-    connect(createNewSystemDesignAction_, SIGNAL(triggered()),
-        this, SLOT(onCreateSystemDesign()), Qt::UniqueConnection);
-
[…]
