### 2017-03-09 e75202a
[ADDITIVE] Now the manual has a section about library navigation.
