### 2017-03-09 66b18ba
[CORRECTIVE] Fixed some crashes that would occur when a port map does not have a physical port.
