### 2017-03-10 a6d21d9
[ADDITIVE] Component item editor can be opened for memory designer items (refs #80).
