+++ designEditors/MemoryDesigner/MemoryDesignDocument.h
@@ -226,9 +226,10 @@ signals:
     /*!
      *  Open the component document for the selected VLNV.
      *
-     *      @param [in] vlnv    VLNV of the containing component.
+     *      @param [in] vlnv                VLNV of the containing component.
+     *      @param [in] identifierChain     List of strings to identify the component item.
      */
-    void openComponentDocument(VLNV const& vlnv);
+    void openComponentDocument(VLNV const& vlnv, QVector<QString> identifierChain);
 
 protected:
     
