### 2017-03-10 8f99e97
[CORRECTIVE] Fixed error in parsing Verilog ports on import (refs #341).
