+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.3.763.0
+#define VERSION_FULL           3.3.766.0
 
 #define VERSION_BASEYEAR       0
-#define VERSION_DATE           "2017-03-09"
-#define VERSION_TIME           "11:41:07"
+#define VERSION_DATE           "2017-03-10"
+#define VERSION_TIME           "09:51:56"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          3
-#define VERSION_BUILDNO        763
+#define VERSION_BUILDNO        766
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,3,763,0
-#define VERSION_PRODUCT        3,3,763,0
-#define VERSION_FILESTR        "3,3,763,0"
-#define VERSION_PRODUCTSTR     "3,3,763,0"
+#define VERSION_FILE           3,3,766,0
+#define VERSION_PRODUCT        3,3,766,0
+#define VERSION_FILESTR        "3,3,766,0"
+#define VERSION_PRODUCTSTR     "3,3,766,0"
 
 #endif
