### 2017-03-10 c687a68
[CLEANUP] Minor cleanup in the documentation. (refs #337)
