### 2017-03-10 8dbc98e
[PERFECTIVE] Enabled external links to be opened from help using the default browser.
