### 2017-03-10 #151 @TermosPullo
@jflorence Hello again. We have now done the revisions. Your feed back was really useful for us! Moreover, reviewing a newer version than earlier is no problem to us. Please let us know if more is needed.

@labarba Could it be feasible to change the repository address of our submission? The current address is no longer functional and it appears I cannot change it myself.
The submission: http://joss.theoj.org/papers/73e33d6850d24f0d6aad0d5f38937f83
The functional address: https://github.com/kactus2/kactus2dev
