### 2017-03-10 1249afe
[CLEANUP] Minor fixes to README.md
