### 2017-03-12 #151 @labarba
@jflorence The author reports completing the requested revisions. Are you satisfied with them to recommend acceptance? I still see five unchecked boxes in the review checklist.
Thanks!
