### 2017-03-13 3614e40
[CORRECTIVE] Added missing upgrade configurations for component editor visibilities. [CORRECTIVE] Fixed setting Kactus2 version in configuration file. [CLEANUP] Improved layout for settings pages.
