### 2017-03-13 78362ab
[CORRECTIVE] Fixed lines that were not valid for Linux build.
