### 2017-03-13 93f2e70
[CORRECTIVE] Corrected text colors where text could not be read. [PERFECTIVE] Added explanatory text to component wizard views-page.
