### 2017-03-14 5584f34
[CORRECTIVE] Corrected text colors where text could not be read.
