### 2017-03-16 63ec4a7
[PERFECTIVE] Adjusted component editor general page layout. [CORRECTIVE] Fixed error in port map port direction match checking. [CLEANUP] Removed obsolete class. [ADDITIVE] Added option for defining license for a component.
