+++ editors/ComponentEditor/general/generaleditor.cpp
@@ -19,22 +19,24 @@
 
 #include <IPXACTmodels/Component/Component.h>
 
-#include <QHBoxLayout>
-#include <QVBoxLayout>
 #include <QGroupBox>
-#include <QStringList>
+#include <QHBoxLayout>
 #include <QLineEdit>
+#include <QSplitter>
+#include <QStringList>
+#include <QVBoxLayout>
 
 //-----------------------------------------------------------------------------
-// Function: GeneralEditor()
+// Function: GeneralEditor::GeneralEditor()
 //-----------------------------------------------------------------------------
 GeneralEditor::GeneralEditor(LibraryInterface* libHandler, QSharedPointer<Component> component, QWidget *parent):
 ItemEditor(component, libHandler, parent),
     vlnvDisplayer_(new VLNVDisplayer(this, component->getVlnv())),
     attributeEditor_(new KactusAttributeEditor(this)),
-    authorEditor_(new QLineEdit()),
-    descEditor_(new DescEditor()),
-    headerEditor_(new DescEditor()),
+    authorEditor_(new QLineEdit(this)),
+    licenseEditor_(new QLineEdit(this)),
+    descriptionEditor_(new QPlainTextEdit(this)),
+    headerEditor_(new QPlainTextEdit(this)),
     previewBox_(new ComponentPreviewBox(libHandler, this))
 {
     Q_ASSERT(libHandler != 0);
@@ -44,47 +46,27 @@ ItemEditor(component, libHandler, parent),
     vlnvDisplayer_->setTitle(tr("Component VLNV and location"));
 	vlnvDisplayer_->setPath(xmlPath);
 
-    headerEditor_->setTitle(tr("XML header"));
-
     previewBox_->setInteractive(true);
     previewBox_->setComponent(component);
-    previewBox_->setFixedWidth(280);
+    //previewBox_->setFixedWidth(280);
     previewBox_->setHorizontalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
 
-    // Create the VLNV displayer and attribute & description editors.
-    QVBoxLayout* layout = new QVBoxLayout();
-    layout->setContentsMargins(0, 0, 0, 0);
+    setupLayout();
 
-    QGroupBox* authorBox = new QGroupBox("Author", this);
-
-    QVBoxLayout* authorLayout = new QVBoxLayout(authorBox);
-    authorLayout->addWidget(authorEditor_);
-
-    layout->addWidget(new SummaryLabel(tr("Component summary"), this), 0, Qt::AlignCenter);
-    layout->addWidget(vlnvDisplayer_);
-    layout->addWidget(attributeEditor_);
-    layout->addWidget(authorBox);    
-    layout->addWidget(descEditor_);
-	layout->addWidget(headerEditor_);
-
-    QHBoxLayout* topLayout = new QHBoxLayout(this);
-    topLayout->addLayout(layout, 1);
-    topLayout->addWidget(previewBox_);
-    topLayout->setContentsMargins(4, 0, 0, 0);
-
-    // Connect the contentChanged() signals.
     connect(previewBox_, SIGNAL(endpointsRearranged()), this, SIGNAL(contentChanged()), Qt::UniqueConnection);
     connect(attributeEditor_, SIGNAL(contentChanged()), this, SLOT(onAttributesChange()), Qt::UniqueConnection);
-    connect(descEditor_, SIGNAL(contentChanged()), this, SLOT(onDescriptionChange()), Qt::UniqueConnection);
-	connect(headerEditor_, SIGNAL(contentChanged()), this, SLOT(onHeaderChange()), Qt::UniqueConnection);
+    connect(descriptionEditor_, SIGNAL(textChanged()), this, SLOT(onDescriptionChange()), Qt::UniqueConnection);
[…]
