### 2017-03-16 fa6a454
[ADDITIVE] Added local memory maps to memory designer (refs #80). [CORRECTIVE] Fixed opening of the component editor in memory designer for register items, when address blocks are hidden (refs #80).
