+++ designEditors/MemoryDesigner/MemoryConnectionHandler.cpp
@@ -20,6 +20,7 @@
 #include <designEditors/MemoryDesigner/MemoryConnectionItem.h>
 #include <designEditors/MemoryDesigner/MemoryCollisionItem.h>
 #include <designEditors/MemoryDesigner/MemoryColumnHandler.h>
+#include <designEditors/MemoryDesigner/MemoryItem.h>
 
 //-----------------------------------------------------------------------------
 // Function: MemoryConnectionHandler::MemoryConnectionHandler()
@@ -136,12 +137,22 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
     QSharedPointer<ConnectivityInterface> startInterface = connectionPath.first();
     QSharedPointer<ConnectivityInterface> endInterface = connectionPath.last();
 
-    if (startInterface && endInterface && startInterface != endInterface)
+    if (startInterface && endInterface)
     {
         MainMemoryGraphicsItem* connectionStartItem =
             getMainGraphicsItem(startInterface, MemoryDesignerConstants::ADDRESSSPACECOLUMN_NAME);
-        MainMemoryGraphicsItem* connectionEndItem =
-            getMainGraphicsItem(endInterface, MemoryDesignerConstants::MEMORYMAPCOLUMNCOMMON_NAME);
+        MainMemoryGraphicsItem* connectionEndItem(0);
+
+        if (startInterface == endInterface)
+        {
+            connectionEndItem = getLocalMemoryMapItem(endInterface);
+        }
+        else
+        {
+            connectionEndItem =
+                getMainGraphicsItem(endInterface, MemoryDesignerConstants::MEMORYMAPCOLUMNCOMMON_NAME);
+        }
+
         if (connectionStartItem && connectionEndItem)
         {
             if (!placedSpaceItems->contains(connectionStartItem))
@@ -151,10 +162,13 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
             }
 
             quint64 baseAddressNumber = 0;
-            QString startInterfaceBaseAddress = startInterface->getBaseAddress();
-            if (startInterfaceBaseAddress.compare(QStringLiteral("x"), Qt::CaseInsensitive) != 0)
+            if (startInterface != endInterface)
             {
-                baseAddressNumber = startInterfaceBaseAddress.toULongLong();
+                QString startInterfaceBaseAddress = startInterface->getBaseAddress();
+                if (startInterfaceBaseAddress.compare(QStringLiteral("x"), Qt::CaseInsensitive) != 0)
+                {
+                    baseAddressNumber = startInterfaceBaseAddress.toULongLong();
+                }
             }
 
             bool hasRemapRange = false;
@@ -238,10 +252,18 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
 
             quint64 remappedAddress =
                 baseAddressNumber + spaceChainConnectionBaseAddress + mirroredSlaveAddressChange;
+
+            unsigned int yTransfer = baseAddressNumber + mirroredSlaveAddressChange;
+
+            if (!hasRemapRange)
+            {
+                remappedAddress += memoryMapBaseAddress;
+                yTransfer += memoryMapBaseAddress;
+            }
+
             quint64 remappedEndAddress =
                 endAddressNumber + spaceChainConnectionBaseAddress + mirroredSlaveAddressChange;
 
-            unsigned int yTransfer = baseAddressNumber + memoryMapBaseAddress + mirroredSlaveAddressChange;
             if (filterAddressSpaceChains_)
             {
                 yTransfer += spaceChainConnectionBaseAddress;
@@ -252,14 +274,21 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
 
[…]

[…]
