### 2017-03-16 90b886d
[CORRECTIVE] Revised the rules when to use abstraction definition default instead of port default. (refs #342)
