+++ tests/Plugins/HDLParser/tst_HDLParser.cpp
@@ -57,7 +57,7 @@ private slots:
     void testHierarchicalConnections();
     void testHierarchicalConnectionsWithExpressions();
 
-    void testPortMapping();
+    //void testPortMapping();
 
     void testMasterToSlaveInterconnection();
     void testEmptyBounds();
@@ -66,12 +66,13 @@ private slots:
     void testInterconnectionToVaryingSizeLogicalMaps();
     void testSlicedInterconnection();
     void testAbsDefDefault();
+    void testPortDefaultValueInComponentInstance();
+    void testHierarchicalPortDefaultValue();
 
     void testAdhocConnectionBetweenComponentInstances();    
     void testAdhocConnectionToVaryingSizePorts();    
     void testAdhocConnectionWithPartSelect();
     void testAdhocTieOffInComponentInstance();
-    void testPortDefaultValueInComponentInstance();
     void testHierarchicalAdhocConnection();
     void testAdHocConnectionBetweenMultipleComponentInstances();
 
@@ -105,10 +106,10 @@ private:
 
     QSharedPointer<ComponentInstance> addInstanceToDesign(QString instanceName, VLNV instanceVlnv, QSharedPointer<View> activeView);
 
-    void mapPortToInterface(QString const& portName, QString const& logicalName, 
+    QSharedPointer<PortMap> mapPortToInterface(QString const& portName, QString const& logicalName, 
         QString const& interfaceName, QSharedPointer<Component> component);
 
-    void mapPortToInterface(QString const& portName, int left, int right, 
+    QSharedPointer<PortMap> mapPortToInterface(QString const& portName, int left, int right, 
         QString const& logicalName, QString const& interfaceName, QSharedPointer<Component> component);
     
     QSharedPointer<BusInterface> addInterfaceToComponent(QString const& interfaceName, QSharedPointer<Component> component, QSharedPointer<ConfigurableVLNVReference> absRef);
@@ -523,7 +524,7 @@ QSharedPointer<BusInterface> tst_HDLParser::addInterfaceToComponent(QString cons
 //-----------------------------------------------------------------------------
 // Function: tst_HDLParser:::mapPortToInterface()
 //-----------------------------------------------------------------------------
-void tst_HDLParser::mapPortToInterface(QString const& portName, QString const& logicalName, 
+QSharedPointer<PortMap> tst_HDLParser::mapPortToInterface(QString const& portName, QString const& logicalName, 
     QString const& interfaceName, QSharedPointer<Component> component)
 {
 	QSharedPointer<Port> port = component->getPort(portName);
@@ -542,12 +543,14 @@ void tst_HDLParser::mapPortToInterface(QString const& portName, QString const& l
     QSharedPointer<QList<QSharedPointer<PortMap> > > portMaps = containingBusIf->getPortMaps();
 
     portMaps->append(portMap);
+
+    return portMap;
 }
 
 //-----------------------------------------------------------------------------
 // Function: tst_HDLParser::mapPortToInterface()
 //-----------------------------------------------------------------------------
-void tst_HDLParser::mapPortToInterface(QString const& portName, int left, int right, QString const& logicalName, 
+QSharedPointer<PortMap> tst_HDLParser::mapPortToInterface(QString const& portName, int left, int right, QString const& logicalName, 
[…]
