### 2017-03-17 ccc291c
[CLEANUP] Relocated all under library/LibraryManager to library.
