+++ IPXACTmodels/Component/validators/InstantiationsValidator.cpp
@@ -13,7 +13,7 @@
 
 #include <editors/ComponentEditor/common/ExpressionParser.h>
 
-#include <library/LibraryManager/libraryinterface.h>
+#include <library/LibraryInterface.h>
 
 #include <IPXACTmodels/Component/Choice.h>
 #include <IPXACTmodels/Component/DesignConfigurationInstantiation.h>
