+++ IPXACTmodels/Catalog/validators/CatalogValidator.cpp
@@ -11,7 +11,7 @@
 
 #include "CatalogValidator.h"
 
-#include <library/LibraryManager/libraryinterface.h>
+#include <library/LibraryInterface.h>
 
 #include <editors/ComponentEditor/common/ExpressionParser.h>
 
