+++ editors/ComponentEditor/instantiations/DesignInstantiationsEditor.cpp
@@ -18,7 +18,7 @@
 #include <common/widgets/summaryLabel/summarylabel.h>
 #include <common/delegates/LineEditDelegate/lineeditdelegate.h>
 
-#include <library/LibraryManager/libraryinterface.h>
+#include <library/LibraryInterface.h>
 
 #include <QVBoxLayout>
 
