+++ IPXACTmodels/Design/validator/InterconnectionValidator.cpp
@@ -22,7 +22,7 @@
 
 #include <editors/ComponentEditor/common/ExpressionParser.h>
 
-#include <library/LibraryManager/libraryinterface.h>
+#include <library/LibraryInterface.h>
 
 #include <QRegularExpression>
 
