### 2017-03-17 2394370
[PERFECTIVE] Changed the display of local memory map connections in memory designer (refs #80).
