### 2017-03-20 be51c53
[CORRECTIVE] Fixed library mock so that it compiles again.
