### 2017-03-21 b6b8075
[CORRECTIVE] Fixed some crashes that would occur when importing a component.
