### 2017-03-21 2ce04d1
[CORRECTIVE] Fixed an issue in verilog port import for a large number of ports.
