+++ Plugins/VerilogImport/VerilogSyntax.h
@@ -23,7 +23,7 @@ namespace VerilogSyntax
 	const QString TAG_OVERRIDE("WARNING: EVERYTHING ON AND ABOVE THIS LINE MAY BE OVERWRITTEN BY KACTUS2!!!");
 
     //! Module begins with module <name> #(<parameters>) (<ports>);.
-    const QRegularExpression MODULE_BEGIN("module\\s+(\\w+)\\s*(#\\s*[(].*[)])?\\s*(?=([(][^)]*[)])?\\s*;)?",
+    const QRegularExpression MODULE_BEGIN("(?:^|\\r?\\n)[ \t]*(?:macro)?module\\s+(\\w+)\\s*(#\\s*[(].*[)])?\\s*(?=([(][^)]*[)])?\\s*;)?",
         QRegularExpression::DotMatchesEverythingOption);
 
 	//! Module header begins with module <name> #(.
