+++ tests/Plugins/VerilogImport/VerilogPortParser/tst_VerilogPortParser.cpp
@@ -560,6 +560,144 @@ void tst_VerilogPortParser::testMultipleVerilog1995PortsAreParsed_data()
         "\n"
         "endmodule"
         << QString("clk, rst, enable").split(", ");    
+    
+    QTest::newRow("Port definition in port list commments") << 
+        "module test(clk, // input mask,\n"
+        "rst,\n"
+        "enable);\n"
+        "    input clk;\n"
+        "    input rst;\n"
+        "    input enable;\n"
+        "\n"
+        "endmodule"
+        << QString("clk, rst, enable").split(", ");   
+
+    QTest::newRow("C") << 
+        "module  multiport (\n"
+        "\n"
+        "// output ports\n"
+        "port1,                         // Comment on port\n"
+        "port2,                         // Comment\n"
+        "port3,                         // Comment\n"
+        "port4,                         // Comment\n"
+        "port5,                         // Comment\n"
+        "port6,                         // Comment\n"
+        "port7,                         // Comment\n"
+        "port8,                         // Comment\n"
+        "port9,                         // Comment\n"
+        "port10,                        // Comment on port\n"
+        "port11,                        // Comment\n"
+        "port12,                        // Comment\n"
+        "port13,                        // Comment\n"
+        "port14,                        // Comment\n"
+        "port15,                        // Comment\n"
+        "port16,                        // Comment\n"
+        "port17,                        // Comment\n"
+        "port18,                        // Comment\n"
+        "port19,                        // Comment\n"
+        "port20,                        // Comment on port\n"
+        "port21,                        // Comment\n"
+        "port22,                        // Comment\n"
+        "port23,                        // Comment\n"
+        "port24,                        // Comment\n"
+        "port25,                        // Comment\n"
+        "port26,                        // Comment\n"
+        "port27,                        // Comment\n"
+        "\n"
+        "// input ports\n"
+        "port28,                        // Comment\n"
+        "port29,                        // Comment\n"
+        "port30,                        // Comment on port\n"
+        "port31,                        // Comment\n"
+        "port32,                        // Comment\n"
+        "port33,                        // Comment\n"
+        "port34,                        // Comment\n"
+        "port35,                        // Comment\n"
+        "port36,                        // Comment\n"
+        "port37,                        // Comment\n"
+        "port38,                        // Comment\n"
+        "port39,                        // Comment\n"
+        "port40,                        // Comment on port\n"
+        "port41,                        // Comment\n"
+        "port42,                        // Comment\n"
+        "port43,                        // Comment\n"
+        "port44,                        // Comment\n"
+        "port45,                        // Comment\n"
+        "port46,                        // Comment\n"
+        "port47,                        // Comment\n"
+        "port48,                        // Comment\n"
+        "port49,                        // Comment\n"
+        "port50                         // Comment on last port\n"
+        ");\n"
+        "\n"
+        "//=============================================================================\n"
+        "// output ports\n"
+        "//=============================================================================\n"
+        "output               port1;    // Comment on port\n"
+        "output               port2;    // Comment\n"
+        "output         [1:0] port3;    // Comment\n"
+        "output               port4;    // Comment\n"
+        "output         [2:0] port5;    // Comment\n"
+        "output         [3:0] port6;    // Comment\n"
+        "output         [4:0] port7;    // Comment\n"
+        "output               port8;    // Comment\n"
+        "output         [5:0] port9;    // Comment\n"
+        "output         [6:0] port10;   // Comment on port\n"
+        "output               port11;   // Comment\n"
+        "output         [7:0] port12;   // Comment\n"
+        "output               port13;   // Comment\n"
+        "output         [8:0] port14;   // Comment\n"
+        "output         [9:0] port15;   // Comment\n"
[…]
