+++ Administrative/releaseFiles/release_notes.txt
@@ -8,19 +8,19 @@
               * Changed or Information
 ===============================================================================
 
-dd/03/2017 3.4.0
+22/03/2017 3.4.0
 ----------------------------------------------------
 
 + Enhanced features for Memory Designer:
-    * Added feature to visualize overlapping memory items
-	* Added feature to open the containing component and the editor for the selected item
-	* Added feature to visualize local memory maps
-	* Changed the visual look of a connection through a bridge
-	* Changed display name to be shown instead of element name, if defined
+    * Added feature to visualize overlapping memory items    
+    * Added feature to open the containing component and the editor for the selected item
+    * Added feature to visualize local memory maps
+    * Changed the visual look of a connection through a bridge
+    * Changed display name to be shown instead of element name, if defined
 + Improvements to Verilog generation:
-	* Added preview for generated files
-	* Added message console for reporting generation status
-	* Improved port assignment and vector boundary generation logic
+    * Added preview for generated files
+    * Added message console for reporting generation status
+    * Improved port assignment and vector boundary generation logic
 + Fixed error in importing Verilog ports
 + Added editor for Catalogs
 + Component editor layout updated for most editors
@@ -40,7 +40,7 @@ dd/03/2017 3.4.0
 - Fixed crashing when a design did not have a configuration
 - Fixed crashing when closing a design in specific cases
 - Fixed modifications to design connections while the document was unlocked
-- Fixed CSV import/export of signal definitions in abstraction definition	
+- Fixed CSV import/export of signal definitions in abstraction definition
 - Fixed missing component editor visibility options
 
 * System mode bus interface coloring changed from red to purple to avoid confusion with invalid interfaces
@@ -48,15 +48,15 @@ dd/03/2017 3.4.0
 * Toolbar "Check intergrity" changed to show listing of all errors within the library
   in separate window.
 * Exit screen is displayed longer
-	
+
 09/12/2016 3.3.0
 ----------------------------------------------------
 
 + Enhanced features for Memory Designer:
     * Register fields added 
     * Compression of items to minimize the required space
-	* Multiple address spaces can now be shown as part of the same connection
-	* Filtering options added
+    * Multiple address spaces can now be shown as part of the same connection
+    * Filtering options added
 + Memory connectivity analysis now manages also instances with identical names
 + File paths within filesets now accept URI expressions
 + Files can be drag-dropped to filesets from the file system
@@ -66,25 +66,25 @@ dd/03/2017 3.4.0
 + Zooming now follows mouse location in design
 + Enhanced usage of drafts in design:
     * Bus interface definitions copying when connecting hierarchical draft interface fixed
-	* Bus interface definitions copying when connecting non-hierarchical draft interface removed
-	* Bus interface definitions and port copying to draft instances changed to take place on
-	  packaging of the containing draft component instance
-	* Copy-paste of bus interface to a draft component instance changed to copy only the name and 
-	  the mode of the interface
+    * Bus interface definitions copying when connecting non-hierarchical draft interface removed
+    * Bus interface definitions and port copying to draft instances changed to take place on
+      packaging of the containing draft component instance
+    * Copy-paste of bus interface to a draft component instance changed to copy only the name and 
+      the mode of the interface
 + Notification on trying to read XML files of previous standard versions added
 + Improvements to Verilog generation:
     * Generation setup dialog simplified
     * Register definition creation set as optional
-	* Module name is correctly used for component instances
+    * Module name is correctly used for component instances
     * Environmental identifier created by the generator changed
[…]
