+++ README.md
@@ -1,9 +1,7 @@
 
-===============================================================================
                                 README : Kactus2
            Copyright (c) 2012-2017 Tampere University of Technology
                             http://funbase.cs.tut.fi
-===============================================================================
 
 Summary
 ----------------------------------------------------
