### 2017-03-23 86638df
[CORRECTIVE] Fixed the failure in rules defining used default value for port assignments. (refs #342)
