### 2017-03-24 a6d5b68
[PERFECTIVE] Set abstract wire default value to accept any string as input.
