### 2017-03-24 dadfa40
[CORRECTIVE] Now IP-XACT models library takes the system group into account when deducing the correct direction of a logical port. Consequently, fixed core components to use system group as a parameter.
[CORRECTIVE] Fixed a glitch where the name of the existing system group was not displayed in bus interface editor.
