+++ editors/ComponentEditor/busInterfaces/portmaps/BusInterfacePortMapTab.cpp
@@ -158,11 +158,11 @@ void BusInterfacePortMapTab::refresh()
 //-----------------------------------------------------------------------------
 // Function: BusInterfacePortMapTab::setAbsType()
 //-----------------------------------------------------------------------------
-void BusInterfacePortMapTab::setAbsType(const VLNV& vlnv, General::InterfaceMode mode)
+void BusInterfacePortMapTab::setAbsType(const VLNV& vlnv, General::InterfaceMode mode, QString systemGroup)
 {
-    portMapsModel_.setAbsType(vlnv, mode);
+    portMapsModel_.setAbsType(vlnv, mode, systemGroup);
     portMapsDelegate_->updateLogicalPortNames(vlnv, mode);
-    autoConnector_.setAbstractionDefinition(vlnv, mode);
+    autoConnector_.setAbstractionDefinition(vlnv, mode, systemGroup);
 
     refresh();
 }
