### 2017-03-27 25ed9d4
[CORRECTIVE] Fixed issues caused by not checking the system group of an interface or logical port.
