### 2017-03-29 7728615
[CLEANUP] Fixed various occurrences where it is more beneficial to use reference to a string rather than a copy of the string.
