+++ editors/ComponentEditor/busInterfaces/portmaps/BusInterfacePortMapTab.h
@@ -81,7 +81,7 @@ public:
      *      @param [in] mode            The interface mode of the bus interface.
      *      @param [in] systemGroup     The used system group in case of system mode.
 	 */
-	virtual void setAbsType(const VLNV& vlnv, General::InterfaceMode mode, QString systemGroup);
+	virtual void setAbsType(const VLNV& vlnv, General::InterfaceMode mode, QString const& systemGroup);
 
     /*!
      *  Sets a subset of component ports to be visible in the physical port list.
