### 2017-03-29 9dc80f2
[CORRECTIVE] Updated memory designer construction (refs #80).
