+++ designEditors/MemoryDesigner/MemoryConnectionHandler.h
@@ -97,6 +97,34 @@ private:
     MemoryConnectionHandler(MemoryConnectionHandler const& rhs);
     MemoryConnectionHandler& operator=(MemoryConnectionHandler const& rhs);
 
+    //! Path variables determined during connection path examination.
+    struct ConnectionPathVariables
+    {
+        //! The base address number of the path interfaces.
+        quint64 baseAddressNumber_;
+
+        //! Address change caused by mirrored slave bus interfaces.
+        quint64 mirroredSlaveAddressChange_;
+
+        //! End address of the memory map modified by mirrored slave interfaces.
+        quint64 memoryMapEndAddress_;
+
+        //! Base address modified by the interface bridges.
+        quint64 spaceChainBaseAddress_;
+
+        //! Flag for paths that have gone through a mirrored slave with a remapped address range.
+        bool hasRemapRange_;
+
+        //! Flag for paths that have gone through a bridge.
+        bool isChainedSpaceConnection_;
+
+        //! Address space items referenced by the bridged master bus interfaces of the path.
+        QVector<MainMemoryGraphicsItem*> spaceChain_;
+
+        //! The last address space item that will be connected to the memory map item.
+        MainMemoryGraphicsItem* spaceItemConnectedToMapItem_;
+    };
+
     /*!
      *  Create a memory connection between an address space and a memory map.
      *
@@ -113,6 +141,117 @@ private:
         QSharedPointer<QVector<MainMemoryGraphicsItem*> > placedMapItems, MemoryColumn* memoryMapColumn,
         int& spaceYPlacement, QSharedPointer<QVector<MainMemoryGraphicsItem*> > placedSpaceItems,
         MemoryColumn* spaceColumn, qreal& previousConnectionLow);
+
+    /*!
+     *  Get the end item for the memory connection.
+     *
+     *      @param [in] startInterface  Start interface of the memory connection path.
+     *      @param [in] endInterface    End interface of the memory connection path.
+     *
+     *      @return The end item of the memory connection.
+     */
+    MainMemoryGraphicsItem* getConnectionEndItem(QSharedPointer<ConnectivityInterface> startInterface,
+        QSharedPointer<ConnectivityInterface> endInterface) const;
+
+    /*!
+     *  Get the starting base address of the memory connection.
+     *
+     *      @param [in] startInterface  Start interface of the memory connection path.
+     *      @param [in] endInterface    End interface of the memory connection path.
+     *
+     *      @return Starting base address of the memory connection.
+     */
+    quint64 getStartingBaseAddress(QSharedPointer<ConnectivityInterface> startInterface,
+        QSharedPointer<ConnectivityInterface> endInterface) const;
+
+    /*!
+     *  Examine the connection path for master or mirrored slave interfaces..
+     *
+     *      @param [in] baseAddressNumber       Base address number of the memory connection.
+     *      @param [in] connectionStartItem     Start item of the memory connection.
+     *      @param [in] connectionEndItem       End item of the memory connection.
+     *      @param [in] startInterface          Start interface of the connection path.
+     *      @param [in] endInterface            End interface of the connection path.
+     *      @param [in] connectionPath          The connection path.
+     *      @param [in] spaceColumn             Memory column for address space items.
+     *      @param [in] placedSpaceItems        List of all the placed address space items.
+     *      @param [in] spaceYPlacement         Y-coordinate of the next address space item.
+     *
+     *      @return List of variable changes caused by the interfaces on the connection path.
+     */
+    ConnectionPathVariables examineConnectionPath(quint64 baseAddressNumber,
[…]
