+++ designEditors/MemoryDesigner/MemoryConnectionHandler.cpp
@@ -105,8 +105,11 @@ void MemoryConnectionHandler::createMemoryConnections(QSharedPointer<Connectivit
     qreal previousConnectionLow = 0;
     foreach (QVector<QSharedPointer<ConnectivityInterface> > singlePath, masterSlavePaths)
     {
-        createConnection(singlePath, placedMapItems, memoryMapColumn, spaceYPlacement, placedSpaceItems,
-            spaceColumn, previousConnectionLow);
+        if (!singlePath.isEmpty())
+        {
+            createConnection(singlePath, placedMapItems, memoryMapColumn, spaceYPlacement, placedSpaceItems,
+                spaceColumn, previousConnectionLow);
+        }
     }
 
     compressGraphicsItems(placedSpaceItems, spaceYPlacement, spaceColumn);
@@ -137,207 +140,73 @@ void MemoryConnectionHandler::createConnection(QVector<QSharedPointer<Connectivi
     QSharedPointer<ConnectivityInterface> startInterface = connectionPath.first();
     QSharedPointer<ConnectivityInterface> endInterface = connectionPath.last();
 
-    if (startInterface && endInterface)
+    MainMemoryGraphicsItem* connectionStartItem =
+        getMainGraphicsItem(startInterface, MemoryDesignerConstants::ADDRESSSPACECOLUMN_NAME);
+
+    MainMemoryGraphicsItem* connectionEndItem = getConnectionEndItem(startInterface, endInterface);
+
+    if (!connectionStartItem || !connectionEndItem)
     {
-        MainMemoryGraphicsItem* connectionStartItem =
-            getMainGraphicsItem(startInterface, MemoryDesignerConstants::ADDRESSSPACECOLUMN_NAME);
-        MainMemoryGraphicsItem* connectionEndItem(0);
-
-        if (startInterface == endInterface)
-        {
-            connectionEndItem = getLocalMemoryMapItem(endInterface);
-        }
-        else
-        {
-            connectionEndItem =
-                getMainGraphicsItem(endInterface, MemoryDesignerConstants::MEMORYMAPCOLUMNCOMMON_NAME);
-        }
-
-        if (connectionStartItem && connectionEndItem)
-        {
-            if (!placedSpaceItems->contains(connectionStartItem))
-            {
-                spaceColumn->moveGraphicsItem(
-                    connectionStartItem, spaceYPlacement, MemoryDesignerConstants::SPACEITEMINTERVAL);
-            }
-
-            quint64 baseAddressNumber = 0;
-            if (startInterface != endInterface)
-            {
-                QString startInterfaceBaseAddress = startInterface->getBaseAddress();
-                if (startInterfaceBaseAddress.compare(QStringLiteral("x"), Qt::CaseInsensitive) != 0)
-                {
-                    baseAddressNumber = startInterfaceBaseAddress.toULongLong();
-                }
-            }
-
-            bool hasRemapRange = false;
-            quint64 mirroredSlaveAddressChange = 0;
-            quint64 memoryMapEndAddress = 0;
-
-            QVector<MainMemoryGraphicsItem*> spaceItemChain;
-            spaceItemChain.append(connectionStartItem);
-
-            bool isChainedSpaceConnection = false;
-
-            quint64 spaceChainConnectionBaseAddress = 0;
-            foreach(QSharedPointer<ConnectivityInterface> pathInterface, connectionPath)
-            {
-                if (pathInterface != startInterface && pathInterface != endInterface)
-                {
-                    if (pathInterface->getMode().compare(
-                        QStringLiteral("mirroredSlave"), Qt::CaseInsensitive) == 0 &&
-                        !pathInterface->getRemapAddress().isEmpty() && !pathInterface->getRemapRange().isEmpty())
-                    {
[…]
