### 2017-03-30 11c99e9
[CORRECTIVE] Fixed HTML generation for register and field resets.
