+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.4.3.0
+#define VERSION_FULL           3.4.16.0
 
 #define VERSION_BASEYEAR       0
-#define VERSION_DATE           "2017-03-23"
-#define VERSION_TIME           "14:04:19"
+#define VERSION_DATE           "2017-03-30"
+#define VERSION_TIME           "08:58:13"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          4
-#define VERSION_BUILDNO        3
+#define VERSION_BUILDNO        16
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,4,3,0
-#define VERSION_PRODUCT        3,4,3,0
-#define VERSION_FILESTR        "3,4,3,0"
-#define VERSION_PRODUCTSTR     "3,4,3,0"
+#define VERSION_FILE           3,4,16,0
+#define VERSION_PRODUCT        3,4,16,0
+#define VERSION_FILESTR        "3,4,16,0"
+#define VERSION_PRODUCTSTR     "3,4,16,0"
 
 #endif
