### 2017-03-30 3e369c2
[CORRECTIVE] Fixed missing paths for return values.
