### 2017-03-31 334110e
[CORRECTIVE] Fixed reading for connection required in bus interface editor.
