### 2017-03-31 520c2f2
[CORRECTIVE] Now undefined access of a bit field is no longer considered invalid, as such bit field would inherit access of its register.
