### 2017-03-31 915ee65
[CORRECTIVE] Updated overlapping extended memory map placement in memory designer (refs #80).
