### 2017-03-31 a79f377
[PERFECTIVE] Now bus definition editor reverses the unidirectional ports when signal options are added. Moreover, master modes will only get a slave mode and vice versa, and system modes will get signal options for each defined group. (refs #344)
