### 2017-04-03 0423c83
[CORRECTIVE] Fix references to LibraryInterface.h
