+++ editors/ComponentEditor/fileSet/file/filenamelineedit.cpp
@@ -6,7 +6,7 @@
  */
 
 #include "filenamelineedit.h"
-#include <library/LibraryManager/libraryinterface.h>
+#include <library/LibraryInterface.h>
 #include <IPXACTmodels/generaldeclarations.h>
 
 #include <QFileDialog>
