### 2017-04-03 9d5631b
[CORRECTIVE] Updated memory designer chained address space graphics item placement (refs #80).
