### 2017-04-04 #151 @jflorence
Sorry for late reply.
Everything is fine now, except for
"Version: Does the release version given match the GitHub release (3.2.173)?"
What does it mean exactly?
The latest version I reviewed was 3.4.0
