### 2017-04-04 #151 @arfon
> The latest version I reviewed was 3.4.0

@jflorence - I've updated the release at the top of this issue to match.
