### 2017-04-05 5fef49f
[CORRECTIVE] Fixed a glitch that caused duplicate port description in generated Verilog module headers.
[CORRECTIVE] Fixed a glitch in copying component parameters.

[CORRECTIVE] Fixed a rare crash that occurred when running a generator plugin.
