### 2017-04-06 a0f3477
[CLEANUP] Refactored name generation policy.
