### 2017-04-12 573bbcc
[PERFECTIVE] Now the input and output ports of port mapping use more explicit symbols.
