### 2017-04-12 a8fc0b0
[PERFECTIVE] Now also the component ports and abstraction definition ports are illustrated with a direction symbol.
