### 2017-04-13 #151 @labarba
@jflorence We would need your final statement that the submission is complete and ready to accept. Thanks for your review!

And check out the nice [feature on JOSS](http://blogs.nature.com/naturejobs/2017/04/04/joss-gives-computational-scientists-their-academic-due/) that appeared in Nature TechBlogs a few days ago.
