### 2017-04-19 b42c537
[PERFECTIVE] Added installer location to readme
