### 2017-04-20 86766d4
[CLEANUP] Refactored the ModelSim generator to utilize the generic generation framework. Moreover, refactored the framework so that it fits to the needs of the ModelSim generator. Also, now the generated do files add the files to the project before compiling. (refs #348)
