+++ Plugins/ModelSimGenerator/ModelSimWriterFactory/ModelSimWriterFactory.h
@@ -1,27 +1,27 @@
 //-----------------------------------------------------------------------------
-// File: VerilogWriterFactory.h
+// File: ModelSimWriterFactory.h
 //-----------------------------------------------------------------------------
 // Project: Kactus2
 // Author: Janne Virtanen
-// Date: 26.01.2017
+// Date: 20.04.2017
 //
 // Description:
-// Creates writes for generating Verilog.
+// Creates writers for generating do files.
 //-----------------------------------------------------------------------------
 
-#ifndef VERILOGWRITERFACTORY_H
-#define VERILOGWRITERFACTORY_H
+#ifndef MODELSIMFACTORY_H
+#define MODELSIMFACTORY_H
 
-#include "VerilogDocument.h"
+#include "ModelSimDocument.h"
 
 #include <Plugins/PluginSystem/GeneratorPlugin/IWriterFactory.h>
 
 struct GenerationSettings;
 
 //-----------------------------------------------------------------------------
-// Verilog file generator.
+// Creates writers for generating do files.
 //-----------------------------------------------------------------------------
-class VerilogWriterFactory : public IWriterFactory
+class ModelSimWriterFactory : public IWriterFactory
 {
 public:
 
@@ -34,12 +34,12 @@ public:
 	 *      @param [in] kactusVersion		    The version of the current Kactus build.
 	 *      @param [in] generatorVersion	    The current version of the generator.
 	 */
-    VerilogWriterFactory(LibraryInterface* library, 
+    ModelSimWriterFactory(LibraryInterface* library, 
         MessagePasser* messages, GenerationSettings* settings,
         QString const& kactusVersion, QString const& generatorVersion);
 
     //! The destructor.
-    ~VerilogWriterFactory();
+    ~ModelSimWriterFactory();
 
     /*!
      *  Creates writers for the given formatted HDL component.
@@ -53,9 +53,9 @@ public:
         QSharedPointer<MetaComponent> component);
 
     /*!
-     *  Creates writers for the given meta designs.
+     *  Creates writers for the given parsed meta design.
      *
-     *      @param [in] designs             The designs that will yield documents.
+     *      @param [in] design             The design which is needs writing.
      *
      *      @return The objects that bundles the writers. Will be null, if could not be created.
      */
@@ -69,48 +69,8 @@ public:
 private:
 
     // Disable copying.
-    VerilogWriterFactory(VerilogWriterFactory const& rhs);
-    VerilogWriterFactory& operator=(VerilogWriterFactory const& rhs);
-
-    /*!
-     *  Checks if given file exists, and tries to parse its existing implementation.
-     *
-     *      @param [in] document        	    The document, which will get writers for writing design.
-     *      @param [in] outputPath        	    The path of the potentially the existing file.
-	 *
-	 *      @return False, if the file exists, but could not be opened or parsed properly. Else true.
-     */
-    bool readImplementation(QSharedPointer<VerilogDocument> document, QString const& outputPath);
-
-    /*!
-     *  Initializes writers for component.
-     *
-     *      @param [in] document        	    The document, which will get writers for writing component.
-     *      @param [in] component		        The component, which data will be written.
-     */
-    void initializeComponentWriters(QSharedPointer<VerilogDocument> document, QSharedPointer<MetaComponent> component);
-    
-    /*!
-     *  Initializes writers for the given design.
-     *
-     *      @param [in] document        	The document, which will get writers for writing design.
-	 *      @param [in] design		        The design, which data will be written.
-     */
-    void initializeDesignWriters(QSharedPointer<VerilogDocument> document, QSharedPointer<MetaDesign> design);
-   
[…]
