+++ Plugins/PluginSystem/GeneratorPlugin/OutputControl.h
@@ -30,7 +30,15 @@ struct GenerationOutput
     /*!
      *  Writes the content.
      */
-	virtual void write(){}
+	virtual void write(QString const& outputDirectory) = 0;
+    
+    /*!
+     *  Finds position for body text highlight in document, if any exists.
+     *
+	 *      @param [out] begin                  The position where the highlight begins, if successful.
+     *      @param [out] end                    The position where the highlight ends, if successful.
+     */
+    virtual void getBodyHighlight(int& begin, int& end) const = 0;
 };
 
 //-----------------------------------------------------------------------------

[…]
