### 2017-04-21 17f6ba6
[ADDITIVE] Now the generation framework once again features option to not include the generated files to file sets. This is automatically turned off if the output directory is changed to point outside of the top component. Moveover, now the generation may append group identifiers to file sets.
