+++ Plugins/PluginSystem/GeneratorPlugin/ViewSelectionWidget.cpp
@@ -20,9 +20,10 @@
 //-----------------------------------------------------------------------------
 // Function: ViewSelectionWidget::ViewSelectionWidget()
 //-----------------------------------------------------------------------------
-ViewSelectionWidget::ViewSelectionWidget(QSharedPointer<ViewSelection> configuration) : 
-	configuration_(configuration),
+ViewSelectionWidget::ViewSelectionWidget(QSharedPointer<ViewSelection> model) : 
+	model_(model),
     viewSelection_(new QComboBox(this)),
+    addToFileset_(new QGroupBox(tr("Add file to fileset"))),
 	instantiationSelection_(new QLabel),
 	instantiationLanguage_(new QLabel),
     fileSetSelection_(new QComboBox(this)),
@@ -31,16 +32,19 @@ ViewSelectionWidget::ViewSelectionWidget(QSharedPointer<ViewSelection> configura
 	// Make view selection its own layout.
 	QFormLayout* viewSelectionLayout = new QFormLayout();
     viewSelectionLayout->addRow(tr("Select view:"), viewSelection_);
-
-	// The names of available views are the view selection items.
-    viewSelection_->addItems(configuration_->viewNames());
-    // The names of available file sets are the file set selection items.
-    fileSetSelection_->addItems(configuration_->fileSetNames());
-
     // Select an instantiation and the file set in conjunction with view.
     viewSelectionLayout->addRow(tr("The component instantiation:"), instantiationSelection_);
     viewSelectionLayout->addRow(tr("Language of the instantiation:"), instantiationLanguage_);
-    viewSelectionLayout->addRow(tr("Select file set:"), fileSetSelection_);
+
+    // The names of available views are the view selection items.
+    viewSelection_->addItems(model_->viewNames());
+    // The names of available file sets are the file set selection items.
+    fileSetSelection_->addItems(model_->fileSetNames());
+
+    // Checkable group box used to include generated file in the IP-XACT component.
+    addToFileset_->setCheckable(true);
+    addToFileset_->setChecked(model->getSaveToFileset() != 0);
+
     // Is editable, in case a new entry is desired.
     fileSetSelection_->setEditable(true);
 
@@ -48,13 +52,21 @@ ViewSelectionWidget::ViewSelectionWidget(QSharedPointer<ViewSelection> configura
     p.setColor(QPalette::Base, KactusColors::MANDATORY_FIELD);
     fileSetSelection_->setPalette(p);
 
+	// It will have its own sub layout.
+	QFormLayout* filesetLayout = new QFormLayout();
+    addToFileset_->setLayout(filesetLayout);
+    filesetLayout->addRow(tr("Select file set:"), fileSetSelection_);
+
 	// Add everything it their proper position in the final layout.
 	QVBoxLayout* topLayout = new QVBoxLayout(this);
     topLayout->addLayout(viewSelectionLayout);
+    topLayout->addWidget(addToFileset_);
     // Widget for warning messages.
     topLayout->addWidget(fileSetWarningLabel_);
 
 	// Connect the relevant events to their handler functions.
+    connect(addToFileset_, SIGNAL(toggled(bool)), 
+        this, SLOT(onFileSetStateChanged(bool)), Qt::UniqueConnection);
     connect(viewSelection_, SIGNAL(currentIndexChanged(QString const&)),
 		this, SLOT(onViewChanged(QString const&)), Qt::UniqueConnection);
 	connect(fileSetSelection_, SIGNAL(currentIndexChanged(QString const&)),
@@ -63,7 +75,7 @@ ViewSelectionWidget::ViewSelectionWidget(QSharedPointer<ViewSelection> configura
 		this, SLOT(onFileSetChanged(QString const&)), Qt::UniqueConnection);
 
 	// Get the selected view.
[…]
