### 2017-04-24 63e73c9
[CLEANUP] Merged software instance class to component instance class.
