+++ library/LibraryUtils.cpp
@@ -118,7 +118,7 @@ int getInstanceIndexByUUID(QList<QSharedPointer<ComponentInstance> > instances, 
 //-----------------------------------------------------------------------------
 // Function: getInstanceIndex()
 //-----------------------------------------------------------------------------
-int getInstanceIndex(QList<QSharedPointer<SWInstance> > instances, QString const& importRef,
+int getInstanceIndex(QList<QSharedPointer<ComponentInstance> > instances, QString const& importRef,
                      QString const& mapping)
 {
     // Search for a match in the list.
@@ -201,7 +201,7 @@ void parseProgrammableElementsV2(LibraryInterface* lh, VLNV designVLNV,
 //-----------------------------------------------------------------------------
 void addNewInstancesV2(LibraryInterface* lh, QList<QSharedPointer<ComponentInstance> > elements,
                        QSharedPointer<QList<QSharedPointer<ComponentInstance> > > hwInstances,
-                       QList<QSharedPointer<SWInstance> > swInstances,
+                       QList<QSharedPointer<ComponentInstance> > swInstances,
                        QList<QSharedPointer<ApiInterconnection> > apiDependencies)
 {
     foreach (QSharedPointer<ComponentInstance> element, elements)
@@ -233,7 +233,7 @@ void addNewInstancesV2(LibraryInterface* lh, QList<QSharedPointer<ComponentInsta
                 continue;
             }
 
-            foreach (QSharedPointer<SWInstance> instance, swDesign->getSWInstances())
+            foreach (QSharedPointer<ComponentInstance> instance, swDesign->getSWInstances(lh))
             {
                 instance->setPosition(QPointF());
                 instance->setImported(true);
@@ -261,12 +261,12 @@ void generateSystemDesignV2(LibraryInterface* lh, VLNV const& designVLNV, Design
     // Add them as instances to the system design.
     QSharedPointer<QList<QSharedPointer<ComponentInstance> > > 
         hwInstances(new QList<QSharedPointer<ComponentInstance> >());
-    QList<QSharedPointer<SWInstance> > swInstances;
+    QList<QSharedPointer<ComponentInstance> > swInstances;
     QList<QSharedPointer<ApiInterconnection> > apiDependencies;
     addNewInstancesV2(lh, elements, hwInstances, swInstances, apiDependencies);
 
     sysDesign.setComponentInstances(hwInstances);
-    sysDesign.setSWInstances(swInstances);
+    sysDesign.getComponentInstances()->append(swInstances);
     sysDesign.setApiConnections(apiDependencies);
 }
 
@@ -274,7 +274,7 @@ void generateSystemDesignV2(LibraryInterface* lh, VLNV const& designVLNV, Design
 // Function: getMatchingApiDependency()
 //-----------------------------------------------------------------------------
 int getMatchingApiDependency(QList<QSharedPointer<ApiInterconnection> > apiDependencies,
-                             QList<QSharedPointer<SWInstance> > swInstances,
+                             QList<QSharedPointer<ComponentInstance> > swInstances,
                              QSharedPointer<ApiInterconnection> dependency, QString const& mapping) 
 {
     int index = -1;
@@ -345,13 +345,18 @@ void updateSystemDesignV2(LibraryInterface* lh, VLNV const& hwDesignVLNV, Design
     // Reflect changes in the programmable elements to the system design.
     QSharedPointer<QList<QSharedPointer<ComponentInstance> > > 
         hwInstances(new QList<QSharedPointer<ComponentInstance> >());
-    QList<QSharedPointer<SWInstance> > swInstances;
[…]
