+++ Plugins/TLMWGenerator/TLMWGeneratorPlugin.h
@@ -1,113 +1,0 @@
-//-----------------------------------------------------------------------------
-// File: TLMWGeneratorPlugin.h
-//-----------------------------------------------------------------------------
-// Project: Kactus 2
-// Author: Janne Virtanen
-// Date: 13.7.2015
-//
-// Description:
-// TLMW generator plugin.
-//-----------------------------------------------------------------------------
-
-#ifndef TLMWGeneratorPlugin_H
-#define TLMWGeneratorPlugin_H
-
-#include <Plugins/PluginSystem/GeneratorPlugin/IGeneratorPlugin.h>
-#include <QObject>
-
-#include <TLMWParser.h>
-
-//-----------------------------------------------------------------------------
-//! TLMW code generator.
-//-----------------------------------------------------------------------------
-class TLMWGeneratorPlugin : public QObject, public IGeneratorPlugin
-{
-    Q_OBJECT
-    Q_PLUGIN_METADATA(IID "kactus2.plugins.TLMWGenerator" FILE "tlmw.json")
-
-    Q_INTERFACES(IPlugin)
-    Q_INTERFACES(IGeneratorPlugin)
-
-public:
-    TLMWGeneratorPlugin();
-    ~TLMWGeneratorPlugin();
-
-    /*!
-     *  Returns the name of the plugin.
-     */
-    virtual QString getName() const;
-
-    /*!
-     *  Returns the version of the plugin.
-     */
-    virtual QString getVersion() const;
-
-    /*!
-     *  Returns the description of the plugin.
-     */
-    virtual QString getDescription() const;
-
-    /*!
-     *  Returns the vendor of the plugin.
-     */
-    virtual QString getVendor() const;
-
-    /*!
-     *  Returns the licence of the plugin.
-     */
-    virtual QString getLicence() const;
-
-    /*!
-     *  Returns the holder of the licence of the plugin.
-     */
-    virtual QString getLicenceHolder() const;
-
-    /*!
-     *  Returns the settings widget.
-     */
-    virtual QWidget* getSettingsWidget();
-
-    /*!
-     *  This is used to access the settings modified by function getSettingsWidget().
-     */
-    virtual PluginSettingsModel* getSettingsModel(){return NULL;}
-
-    /*!
-     *  Returns the icon for the generator.
-     */
-    virtual QIcon getIcon() const;
-
-    /*!
-     *  Checks whether the generator may run for the given component or design. 
-     *
-     *      @param [in] component	        The component for which to check support. If design is not null, component
-	 *                                      will refer to design or designConfiguration.
-     *      @param [in] design	            The design, if the generator is ran for a design.
-     *      @param [in] designConfiguration The design configuration for design, if it is not null.
-     *
-     *      @return True, if the generator may run the given component. Otherwise false.
-     */
-    virtual bool checkGeneratorSupport(QSharedPointer<Component const> component,
-        QSharedPointer<Design const> design,
-        QSharedPointer<DesignConfiguration const> designConfiguration) const;
-
-    /*!
-     *  Runs the generation, creating new files and/or modifying the IP-XACT metadata. The function has
-	 *  also access to the parent window widget, so that it can show dialogs for the user to aid the generation.
-     *
-     *      @param [in] utility			    The plugin utility interface.
-     *      @param [in] component	        The component for which to check support. If design is not null, component
-     *                                      will refer to design or designConfiguration.
-     *      @param [in] design	            The design, if the generator is ran for a design.
-     *      @param [in] designConfiguration The design configuration for design, if it is not null.
-     */
-    virtual void runGenerator(IPluginUtility* utility, 
-        QSharedPointer<Component> component,
[…]
