+++ Plugins/TLMWGenerator/TLMWGeneratorPlugin.cpp
@@ -1,165 +1,0 @@
-//-----------------------------------------------------------------------------
-// File: TLMWGeneratorPlugin.cpp
-//-----------------------------------------------------------------------------
-// Project: Kactus 2
-// Author: Janne Virtanen
-// Date: 13.7.2015
-//
-// Description:
-// TLMW generator plugin.
-//-----------------------------------------------------------------------------
-
-#include "TLMWGeneratorPlugin.h"
-#include "TLMWHeaderGenerator.h"
-
-#include <QCoreApplication>
-#include <QFileInfo>
-#include <QMessageBox>
-
-#include <library/LibraryInterface.h>
-
-//-----------------------------------------------------------------------------
-// Function: TLMWCGeneratorPlugin::TLMWGeneratorPlugin()
-//-----------------------------------------------------------------------------
-TLMWGeneratorPlugin::TLMWGeneratorPlugin()
-{
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::~TLMWGeneratorPlugin()
-//-----------------------------------------------------------------------------
-TLMWGeneratorPlugin::~TLMWGeneratorPlugin()
-{
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::getName()
-//----------------------------------------------------------------------------
-QString TLMWGeneratorPlugin::getName() const
-{
-    return "TLMW Generator";
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::getVersion()
-//-----------------------------------------------------------------------------
-QString TLMWGeneratorPlugin::getVersion() const
-{
-    return "1.0";
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::getDescription()
-//-----------------------------------------------------------------------------
-QString TLMWGeneratorPlugin::getDescription() const
-{
-    return "Generates TLMW header files based on the metadata.";
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::getVendor()
-//-----------------------------------------------------------------------------
-QString TLMWGeneratorPlugin::getVendor() const
-{
-    return tr("TUT");
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWCodeGeneratorPlugin::getLicence()
-//-----------------------------------------------------------------------------
-QString TLMWGeneratorPlugin::getLicence() const
-{
-    return tr("GPL2");
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::getLicenceHolder()
-//-----------------------------------------------------------------------------
-QString TLMWGeneratorPlugin::getLicenceHolder() const
-{
-    return tr("Public");
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::getSettingsWidget()
-//-----------------------------------------------------------------------------
-QWidget* TLMWGeneratorPlugin::getSettingsWidget()
-{
-    return new QWidget();
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::getIcon()
-//-----------------------------------------------------------------------------
-QIcon TLMWGeneratorPlugin::getIcon() const
-{
-    return QIcon(":icons/TLMWGenerator.png");
-}
-
-//-----------------------------------------------------------------------------
-// Function: TLMWGeneratorPlugin::checkGeneratorSupport()
-//-----------------------------------------------------------------------------
[…]

[…]
