### 2017-04-24 4126cad
[CORRECTIVE] Design no longer distinguishes software instances: It is now responsibility of the caller.
