### 2017-04-24 #151 @labarba
@jflorence Allow me to ping again — The version was updated: if this is the last remaining issue, we await for you to check that box and to state here your recommendation for accepting the paper. Thanks!
