### 2017-04-24 82bef71
[CORRECTIVE] Fixed a fault in code that prevented implementation selection test suite from building.
