### 2017-04-25 e7007d2
[PERFECTIVE] Design creation dialog asks for qualifier for identification of views and configurations.
