### 2017-04-25 d927c7d
[ADDITIVE] Added delete item for context menus in HW design and system design. [PERFECTIVE] Changed position of the protection ribbon group.
