+++ version.h
@@ -10,20 +10,20 @@
 #ifndef VERSIONNO__H
 #define VERSIONNO__H
 
-#define VERSION_FULL           3.4.73.0
+#define VERSION_FULL           3.4.97.0
 
 #define VERSION_BASEYEAR       0
-#define VERSION_DATE           "2017-04-12"
-#define VERSION_TIME           "16:58:16"
+#define VERSION_DATE           "2017-04-25"
+#define VERSION_TIME           "16:06:24"
 
 #define VERSION_MAJOR          3
 #define VERSION_MINOR          4
-#define VERSION_BUILDNO        73
+#define VERSION_BUILDNO        97
 #define VERSION_EXTEND         0
 
-#define VERSION_FILE           3,4,73,0
-#define VERSION_PRODUCT        3,4,73,0
-#define VERSION_FILESTR        "3,4,73,0"
-#define VERSION_PRODUCTSTR     "3,4,73,0"
+#define VERSION_FILE           3,4,97,0
+#define VERSION_PRODUCT        3,4,97,0
+#define VERSION_FILESTR        "3,4,97,0"
+#define VERSION_PRODUCTSTR     "3,4,97,0"
 
 #endif
