### 2017-04-26 1184d25
[CORRECTIVE] Updated component port wire type def writing.
