### 2017-04-27 f37954d
[CORRECTIVE] Now pressing enter in generation dialog does not close the dialog, if the output directory editor is active.
