### 2017-04-28 6891d45
[PERFECTIVE] Replaced flat abstraction definition table with a tree view. (refs #347)
