+++ editors/BusDefinitionEditor/AbsDefTreeModel.cpp
@@ -1,0 +1,933 @@
+//-----------------------------------------------------------------------------
+// File: AbsDefTreeModel.cpp
+//-----------------------------------------------------------------------------
+// Project: Kactus2
+// Author: Janne Virtanen
+// Date: 27.04.2017
+//
+// Description:
+// A model for displaying abstraction definitions.
+//-----------------------------------------------------------------------------
+
+#include "AbsDefTreeModel.h"
+#include "LogicalPortColumns.h"
+
+#include <IPXACTmodels/common/DirectionTypes.h>
+#include <IPXACTmodels/common/PresenceTypes.h>
+
+#include <IPXACTmodels/BusDefinition/BusDefinition.h>
+
+#include <QIcon>
+#include <QSize>
+#include <QMimeData>
+
+#include <common/KactusColors.h>
+
+namespace
+{
+    QString const MULTIPLE_SELECTED = "[multiple]";
+}
+
+//-----------------------------------------------------------------------------
+// Function: AbsDefTreeModel::AbsDefTreeModel()
+//-----------------------------------------------------------------------------
+AbsDefTreeModel::AbsDefTreeModel(QObject *parent):
+QAbstractItemModel(parent),
+absDef_(),
+busDefinition_(),
+table_() 
+{
+}
+
+//-----------------------------------------------------------------------------
+// Function: AbsDefTreeModel::~AbsDefTreeModel()
+//-----------------------------------------------------------------------------
+AbsDefTreeModel::~AbsDefTreeModel()
+{
+
+}
+
+//-----------------------------------------------------------------------------
+// Function: AbsDefTreeModel::columnCount()
+//-----------------------------------------------------------------------------
+int AbsDefTreeModel::columnCount(const QModelIndex&) const
+{
+    return LogicalPortColumns::COLUMN_COUNT;
+}
+
+//-----------------------------------------------------------------------------
+// Function: AbsDefTreeModel::rowCount()
+//-----------------------------------------------------------------------------
+int AbsDefTreeModel::rowCount(const QModelIndex &parent) const
+{
+    if (!parent.isValid())
+    {
+        return table_.size();
+    }
+    else if (!parent.parent().isValid())
+    {
+        return table_.at(parent.row())->wirePorts_.size();
+    }
+
+    return 0;
+}
+
+//-----------------------------------------------------------------------------
+// Function: AbsDefTreeModel::flags()
+//-----------------------------------------------------------------------------
+Qt::ItemFlags AbsDefTreeModel::flags(const QModelIndex& index) const 
+{
+    if (!index.isValid())
+    {
+        return Qt::NoItemFlags;
+    }
+
+    if (index.parent().isValid())
+    {
+        QSharedPointer<PortAbstractionEntry> entry = table_.at(index.parent().row());
+
+        QSharedPointer<WireAbstraction> chosen = entry->portAbstraction_->getWire();
+
+        // Do not allow setting system groups for other than system mode.
+        if (index.column() == LogicalPortColumns::SYSTEM_GROUP && 
+            !chosen->getSystemPorts()->isEmpty())
+        {
+            return Qt::ItemIsEnabled | Qt::ItemIsSelectable;
+        }
+    }
+
+    return Qt::ItemIsEditable | Qt::ItemIsEnabled | Qt::ItemIsSelectable;
+}
+
+//-----------------------------------------------------------------------------
+// Function: AbsDefTreeModel::index()
+//-----------------------------------------------------------------------------
+QModelIndex AbsDefTreeModel::index(int row, int column, QModelIndex const& parent) const
+{
+    if (!parent.isValid() && 0 <= row && row < table_.size())
+    {
[…]
