### 2017-05-01 #151 @jflorence
Yes, I recommend accepting the paper.
