### 2017-05-02 03bec05
[PERFECTIVE] The output path of generation dialog no longer determines the default setting of add to fileset. Instead, it is determined by writer factory implementation.
