### 2017-05-02 #151 @arfon
@jflorence many thanks for your review here and to @labarba for editing this submission ✨

@TermosPullo - your submission is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00151 ⚡️ 🚀 💥
